Contrasts Between Adjusted Predictions

This function calculates contrasts (or comparisons) between adjusted predictions for each row of the dataset. The resulting object can processed by the tidy() or summary() functions, which compute Average Contrasts (see ?summary.marginaleffects). The newdata argument can be used to calculate a variety of contrasts, including "Contrasts at the Mean," "Contrasts at User-Specified values" (aka Contrasts at Representative values), "Contrasts in Marginal Means", "Adjusted Risk Ratios", and much more. For more information, see the Details and Examples sections below, and in the vignettes on the marginaleffects website: https://vincentarelbundock.github.io/marginaleffects/

Getting Started

Contrasts Vignette

Supported Models

Case Studies

Bayesian analyses with brms

Mixed effects models

Generalized Additive Models

Multinomial Logit and Discrete Choice Models

Tables and plots

Robust standard errors and more

Transformations and Custom Contrasts: Adjusted Risk Ratio Example

comparisons(
  model,
  newdata = NULL,
  variables = NULL,
  type = "response",
  vcov = TRUE,
  conf_level = 0.95,
  contrast_factor = "reference",
  contrast_numeric = 1,
  transform_pre = "difference",
  transform_post = NULL,
  interaction = NULL,
  eps = 1e-04,
  ...
)

Arguments

model

Model object

newdata

NULL, data frame, string, or datagrid() call. Determines the predictor values for which to compute contrasts.

NULL (default): Unit-level contrasts for each observed value in the original dataset.

data frame: Unit-level contrasts for each row of the newdata data frame.

string:

"mean": Contrasts at the Mean. Contrasts when each predictor is held at its mean or mode.

"median": Contrasts at the Median. Contrasts when each predictor is held at its median or mode.

"marginalmeans": Contrasts at Marginal Means.

datagrid() call to specify a custom grid of regressors. For example:

newdata = datagrid(cyl = c(4, 6)): cyl variable equal to 4 and 6 and other regressors fixed at their means or modes.

See the Examples section and the datagrid documentation.

variables

NULL, character vector, or named list. The subset of variables for which to compute contrasts.

NULL: compute contrasts for all the variables in the model object (can be slow).

Character vector: subset of variables (usually faster).

Named list: subset of variables with the type of contrasts to use, following the conventions in the contrast_factor and contrast_numeric arguments. Examples:

variables = list(gear = "pairwise", hp = 10)

variables = list(gear = "sequential", hp = c(100, 120))

See the Examples section below.

type

string indicates the type (scale) of the predictions used to compute marginal effects or contrasts. This can differ based on the model type, but will typically be a string such as: "response", "link", "probs", or "zero". When an unsupported string is entered, the model-specific list of acceptable values is returned in an error message.

vcov

Type of uncertainty estimates to report (e.g., for robust standard errors). Acceptable values:

FALSE: Do not compute standard errors. This can speed up computation considerably.

TRUE: Unit-level standard errors using the default vcov(model) variance-covariance matrix.

String which indicates the kind of uncertainty estimates to return.

Heteroskedasticity-consistent: "HC", "HC0", "HC1", "HC2", "HC3", "HC4", "HC4m", "HC5". See ?sandwich::vcovHC

Heteroskedasticity and autocorrelation consistent: "HAC"

Other: "NeweyWest", "KernHAC", "OPG". See the sandwich package documentation.

One-sided formula which indicates the name of cluster variables (e.g., ~unit_id). This formula is passed to the cluster argument of the sandwich::vcovCL function.

Square covariance matrix

Function which returns a covariance matrix (e.g., stats::vcov(model))

conf_level

numeric value between 0 and 1. Confidence level to use to build a confidence interval.

contrast_factor

string. Which pairs of factors should be contrasted?

"reference": Each factor level is compared to the factor reference (base) level

"all": All combinations of observed levels

"sequential": Each factor level is compared to the previous factor level

"pairwise": Each factor level is compared to all other levels

contrast_numeric

string or numeric. Which pairs of numeric values should be contrasted?

Numeric of length 1: Contrast for a gap of contrast_numeric, computed at the observed value plus and minus contrast_numeric / 2

Numeric vector of length 2: Contrast between the 2nd element and the 1st element of the contrast_numeric vector.

"iqr": Contrast across the interquartile range of the regressor.

"sd": Contrast across one standard deviation around the regressor mean.

"2sd": Contrast across two standard deviations around the regressor mean.

"minmax": Contrast between the maximum and the minimum values of the regressor.

transform_pre

(experimental) string or function. How should pairs of adjusted predictions be contrasted?

string: shortcuts to common contrast functions.

"difference" (default): function(hi, lo) hi - lo

"differenceavg": function(hi, lo) mean(hi) - mean(lo)

"ratio": function(hi, lo) hi / lo

"lnratio": function(hi, lo) log(hi / lo)

"ratioavg": function(hi, lo) mean(hi) / mean(lo)

"lnratioavg": function(hi, lo) log(mean(hi) / mean(lo))

function: accept two equal-length numeric vectors of adjusted predictions (hi and lo) and returns a vector of contrasts of the same length, or a unique numeric value.

transform_post

(experimental) A function applied to the estimate and confidence interval just before returning the final results. For example, users can exponentiate their final results by setting transform_post=exp or transform contrasts made on the link scale for ease of interpretation.

interaction

TRUE, FALSE, or NULL

FALSE: Contrasts represent the change in adjusted predictions when one predictor changes and all other variables are held constant.

TRUE: Contrasts represent the changes in adjusted predictions when the predictors specified in the variables argument are manipulated simultaneously.

NULL (default): Behaves like TRUE when the variables argument is specified and the model formula includes interactions. Behaves like FALSE otherwise.

eps

A numeric value specifying the “step” size to use when calculating numerical derivatives. See the Details section below. Warning: the marginal effects computed for certain models can be sensitive to the choice of step (e.g., Bayesian mixed effects).

...

Additional arguments are passed to the predict() method supplied by the modeling package.These arguments are particularly useful for mixed-effects or bayesian models (see the online vignettes on the marginaleffects website). Available arguments can vary from model to model, depending on the range of supported arguments by each modeling package. See the "Model-Specific Arguments" section of the ?marginaleffects documentation for a non-exhaustive list of available arguments.

Details

A "contrast" is the difference between two adjusted predictions, calculated for meaningfully different regressor values (e.g., College graduates vs. Others). Uncertainty estimates are computed using the delta method.

Detailed vignettes on contrasts, marginal effects, predictions, and marginal means, as well as a list of supported models can be found on the package website:

https://vincentarelbundock.github.io/marginaleffects/

Model-Specific Arguments

Some model types allow model-specific arguments to modify the nature of marginal effects, predictions, marginal means, and contrasts.

Package	Class	Argument	Documentation

brms	brmsfit	ndraws	brms::posterior_predict

re_formula

lme4	merMod	include_random	insight::get_predicted

re.form	lme4::predict.merMod

allow.new.levels	lme4::predict.merMod

glmmTMB	glmmTMB	re.form	glmmTMB::predict.glmmTMB

allow.new.levels	glmmTMB::predict.glmmTMB

zitype	glmmTMB::predict.glmmTMB

mgcv	bam	exclude	mgcv::predict.bam

robustlmm	rlmerMod	re.form	robustlmm::predict.rlmerMod

allow.new.levels	robustlmm::predict.rlmerMod

Examples

library(marginaleffects)
library(magrittr)

# Linear model
tmp <- mtcars
tmp$am <- as.logical(tmp$am)
mod <- lm(mpg ~ am + factor(cyl), tmp)
comparisons(mod, contrast_factor = "reference") %>% tidy()
#> Warning: The `cyl` variable is treated as a categorical (factor) variable, but the original data is of class numeric. It is safer and faster to convert such variables to factor before fitting the model and calling `marginaleffects` functions. This warning appears once per session.
#>       type term     contrast   estimate std.error statistic      p.value
#> 1 response   am TRUE - FALSE   2.559954  1.297579  1.972869 4.851044e-02
#> 2 response  cyl        6 - 4  -6.156118  1.535723 -4.008612 6.107658e-05
#> 3 response  cyl        8 - 4 -10.067560  1.452082 -6.933187 4.114709e-12
#>       conf.low conf.high
#> 1   0.01674586  5.103162
#> 2  -9.16607927 -3.146156
#> 3 -12.91358877 -7.221530
comparisons(mod, contrast_factor = "sequential") %>% tidy()
#>       type term     contrast  estimate std.error statistic      p.value
#> 1 response   am TRUE - FALSE  2.559954  1.297579  1.972869 4.851044e-02
#> 2 response  cyl        6 - 4 -6.156118  1.535723 -4.008612 6.107658e-05
#> 3 response  cyl        8 - 6 -3.911442  1.470254 -2.660385 7.805144e-03
#>      conf.low conf.high
#> 1  0.01674586  5.103162
#> 2 -9.16607927 -3.146156
#> 3 -6.79308707 -1.029797
comparisons(mod, contrast_factor = "pairwise") %>% tidy()
#>       type term     contrast   estimate std.error statistic      p.value
#> 1 response   am TRUE - FALSE   2.559954  1.297579  1.972869 4.851044e-02
#> 2 response  cyl        6 - 4  -6.156118  1.535723 -4.008612 6.107658e-05
#> 3 response  cyl        8 - 4 -10.067560  1.452082 -6.933187 4.114709e-12
#> 4 response  cyl        8 - 6  -3.911442  1.470254 -2.660385 7.805144e-03
#>       conf.low conf.high
#> 1   0.01674586  5.103162
#> 2  -9.16607927 -3.146156
#> 3 -12.91358877 -7.221530
#> 4  -6.79308707 -1.029797

# GLM with different scale types
mod <- glm(am ~ factor(gear), data = mtcars)
comparisons(mod, type = "response") %>% tidy()
#>       type term contrast  estimate std.error statistic      p.value  conf.low
#> 1 response gear    4 - 3 0.6666667 0.1174440  5.676462 1.375090e-08 0.4364806
#> 2 response gear    5 - 3 1.0000000 0.1565921  6.386020 1.702589e-10 0.6930852
#>   conf.high
#> 1 0.8968528
#> 2 1.3069148
comparisons(mod, type = "link") %>% tidy()
#>   type term contrast  estimate std.error statistic      p.value  conf.low
#> 1 link gear    4 - 3 0.6666667 0.1174440  5.676462 1.375090e-08 0.4364806
#> 2 link gear    5 - 3 1.0000000 0.1565921  6.386020 1.702589e-10 0.6930852
#>   conf.high
#> 1 0.8968528
#> 2 1.3069148

# Contrasts at the mean
comparisons(mod, newdata = "mean")
#>   rowid     type term contrast comparison std.error statistic      p.value
#> 1     1 response gear    4 - 3  0.6666667 0.1174440  5.676462 1.375090e-08
#> 2     1 response gear    5 - 3  1.0000000 0.1565921  6.386020 1.702589e-10
#>    conf.low conf.high gear
#> 1 0.4364806 0.8968528    3
#> 2 0.6930852 1.3069148    3

# Contrasts between marginal means
comparisons(mod, newdata = "marginalmeans")
#>       type term contrast comparison std.error statistic      p.value  conf.low
#> 1 response gear    4 - 3  0.6666667 0.1174440  5.676462 1.375090e-08 0.4364806
#> 2 response gear    5 - 3  1.0000000 0.1565921  6.386020 1.702589e-10 0.6930852
#>   conf.high
#> 1 0.8968528
#> 2 1.3069148

# Contrasts at user-specified values
comparisons(mod, newdata = datagrid(am = 0, cyl = tmp$cyl))
#> Warning: Some of the variable names are missing from the model data: am, cyl
#>   rowid     type term contrast comparison std.error statistic      p.value
#> 1     1 response gear    4 - 3  0.6666667 0.1174440  5.676462 1.375090e-08
#> 2     2 response gear    4 - 3  0.6666667 0.1174440  5.676462 1.375090e-08
#> 3     3 response gear    4 - 3  0.6666667 0.1174440  5.676462 1.375090e-08
#> 4     1 response gear    5 - 3  1.0000000 0.1565921  6.386020 1.702589e-10
#> 5     2 response gear    5 - 3  1.0000000 0.1565921  6.386020 1.702589e-10
#> 6     3 response gear    5 - 3  1.0000000 0.1565921  6.386020 1.702589e-10
#>    conf.low conf.high gear am cyl
#> 1 0.4364806 0.8968528    3  0   6
#> 2 0.4364806 0.8968528    3  0   4
#> 3 0.4364806 0.8968528    3  0   8
#> 4 0.6930852 1.3069148    3  0   6
#> 5 0.6930852 1.3069148    3  0   4
#> 6 0.6930852 1.3069148    3  0   8

# Numeric contrasts
mod <- lm(mpg ~ hp, data = mtcars)
comparisons(mod, contrast_numeric = 1) %>% tidy()
#>       type term    contrast    estimate std.error statistic      p.value
#> 1 response   hp (x + 1) - x -0.06822828 0.0101193 -6.742389 1.558043e-11
#>      conf.low   conf.high
#> 1 -0.08806175 -0.04839481
comparisons(mod, contrast_numeric = 5) %>% tidy()
#>       type term    contrast   estimate  std.error statistic      p.value
#> 1 response   hp (x + 5) - x -0.3411414 0.05059652 -6.742389 1.558043e-11
#>     conf.low conf.high
#> 1 -0.4403087 -0.241974
comparisons(mod, contrast_numeric = c(90, 100)) %>% tidy()
#>       type term contrast   estimate std.error statistic      p.value   conf.low
#> 1 response   hp 100 - 90 -0.6822828  0.101193 -6.742389 1.558043e-11 -0.8806175
#>    conf.high
#> 1 -0.4839481
comparisons(mod, contrast_numeric = "iqr") %>% tidy()
#>       type term contrast  estimate std.error statistic      p.value  conf.low
#> 1 response   hp  Q3 - Q1 -5.697061 0.8449619 -6.742389 1.558043e-11 -7.353156
#>   conf.high
#> 1 -4.040966
comparisons(mod, contrast_numeric = "sd") %>% tidy()
#>       type term                contrast  estimate std.error statistic
#> 1 response   hp (x + sd/2) - (x - sd/2) -4.677926 0.6938085 -6.742389
#>        p.value  conf.low conf.high
#> 1 1.558043e-11 -6.037766 -3.318087
comparisons(mod, contrast_numeric = "minmax") %>% tidy()
#>       type term  contrast estimate std.error statistic      p.value  conf.low
#> 1 response   hp Max - Min -19.3086  2.863763 -6.742389 1.558043e-11 -24.92147
#>   conf.high
#> 1 -13.69573

# Adjusted Risk Ratio (see Case Study vignette on the website)
mod <- glm(vs ~ mpg, data = mtcars, family = binomial)
cmp <- comparisons(mod, transform_pre = "lnratioavg")
summary(cmp, transform_post = exp)
#> Average contrasts 
#>   Term                  Contrast Effect   Pr(>|z|) 2.5 % 97.5 %
#> 1  mpg ln(mean(x + 1) / mean(x))  1.135 2.3808e-10 1.091   1.18
#> 
#> Model type:  glm 
#> Prediction type:  response 

# Adjusted Risk Ratio: Manual specification of the `transform_pre`
cmp <- comparisons(mod, transform_pre = function(hi, lo) log(mean(hi) / mean(lo)))
summary(cmp, transform_post = exp)
#> Average contrasts 
#>   Term   Contrast Effect   Pr(>|z|) 2.5 % 97.5 %
#> 1  mpg (x + 1), x  1.135 2.3808e-10 1.091   1.18
#> 
#> Model type:  glm 
#> Prediction type:  response 
#> Pre-transformation:  function(hi, lo) log(mean(hi)/mean(lo)) 
# Interactions between contrasts
mod <- lm(mpg ~ factor(cyl) * factor(gear) + hp, data = mtcars)
cmp <- comparisons(mod, variables = c("cyl", "gear"))
#> Warning: Model matrix is rank deficient. Some variance-covariance parameters are
#>   missing.
summary(cmp)
#> Average contrasts 
#>     cyl  gear  Effect Std. Error z value Pr(>|z|)   2.5 % 97.5 %
#> 1 4 - 4 4 - 3  4.2194      3.238  1.3033 0.192489 -2.1262 10.565
#> 2 4 - 4 5 - 3  6.9870      3.700  1.8883 0.058987 -0.2652 14.239
#> 3 6 - 4 3 - 3 -1.1472      3.706 -0.3096 0.756898 -8.4109  6.116
#> 4 6 - 4 4 - 3 -0.6306      3.405 -0.1852 0.853068 -7.3035  6.042
#> 5 6 - 4 5 - 3  2.6778      4.617  0.5800 0.561933 -6.3715 11.727
#> 6 8 - 4 3 - 3 -0.8719      3.829 -0.2277 0.819850 -8.3757  6.632
#> 7 8 - 4 4 - 3  3.3475      6.427  0.5208 0.602495 -9.2501 15.945
#> 8 8 - 4 5 - 3  5.5250      5.868  0.9416 0.346419 -5.9760 17.026
#> 
#> Model type:  lm 
#> Prediction type:  response 

# variable-specific contrasts
cmp <- comparisons(mod, variables = list(gear = "sequential", hp = 10))
#> Warning: Model matrix is rank deficient. Some variance-covariance parameters are
#>   missing.
summary(cmp)
#> Average contrasts 
#>    gear           hp Effect Std. Error z value Pr(>|z|)  2.5 % 97.5 %
#> 1 4 - 3 (x + 10) - x  2.835      2.625  1.0801  0.28010 -2.310  7.980
#> 2 5 - 4 (x + 10) - x  2.054      2.611  0.7866  0.43151 -3.063  7.171
#> 
#> Model type:  lm 
#> Prediction type:  response