Generate "counterfactual" datasets for use in marginaleffects's newdata argument

Generate "counterfactual" datasets for use in marginaleffects's newdata argument

counterfactual(..., model = NULL, data = NULL)

Arguments

...
named arguments with vectors of values for the variables to construct (see Examples below.)

model
Model object

data
data.frame (one and only one of the model and data arguments must be true).

Value

A data.frame where each row of the original data is repeated multiple times for each of the values of the variables in the at list. See example below.

If users supply a model, the data used to fit that model is retrieved using the insight::get_data function, and then replicated with different values of the variables in the at list.

Examples

# All rows are repeated twice, with different values of `hp`
cd <- counterfactual(data = mtcars, hp = c(100, 110))
cd[cd$rowid %in% 1:3,]

#>    rowid  mpg cyl disp drat    wt  qsec vs am gear carb  hp
#> 1      1 21.0   6  160 3.90 2.620 16.46  0  1    4    4 100
#> 2      2 21.0   6  160 3.90 2.875 17.02  0  1    4    4 100
#> 3      3 22.8   4  108 3.85 2.320 18.61  1  1    4    1 100
#> 33     1 21.0   6  160 3.90 2.620 16.46  0  1    4    4 110
#> 34     2 21.0   6  160 3.90 2.875 17.02  0  1    4    4 110
#> 35     3 22.8   4  108 3.85 2.320 18.61  1  1    4    1 110

# We get the same result by feeding a model instead of a data.frame
mod <- lm(mpg ~ hp + wt, mtcars)
cd <- counterfactual(model = mod, hp = c(100, 110))
cd[cd$rowid %in% 1:3,]

#>    rowid    wt  hp
#> 1      1 2.620 100
#> 2      2 2.875 100
#> 3      3 2.320 100
#> 33     1 2.620 110
#> 34     2 2.875 110
#> 35     3 2.320 110

# Use in `marginaleffects` to compute "Counterfactual Average Marginal Effects"
marginaleffects(mod, newdata = counterfactual(hp = c(100, 110)))

#>     rowid term        dydx   std.error    wt  hp
#> 1       1   hp -0.03177295 0.009029704 2.620 100
#> 2       1   wt -3.87783074 0.632733547 2.620 100
#> 3       2   hp -0.03177295 0.009029704 2.875 100
#> 4       2   wt -3.87783074 0.632733322 2.875 100
#> 5       3   hp -0.03177295 0.009029704 2.320 100
#> 6       3   wt -3.87783074 0.632733097 2.320 100
#> 7       4   wt -3.87783074 0.632733322 3.215 100
#> 8       4   hp -0.03177295 0.009029704 3.215 100
#> 9       5   hp -0.03177295 0.009029704 3.440 100
#> 10      5   wt -3.87783074 0.632733322 3.440 100
#> 11      6   hp -0.03177295 0.009029704 3.460 100
#> 12      6   wt -3.87783074 0.632733547 3.460 100
#> 13      7   hp -0.03177295 0.009029704 3.570 100
#> 14      7   wt -3.87783074 0.632733322 3.570 100
#> 15      8   wt -3.87783074 0.632733322 3.190 100
#> 16      8   hp -0.03177295 0.009029704 3.190 100
#> 17      9   hp -0.03177295 0.009029704 3.150 100
#> 18      9   wt -3.87783074 0.632733547 3.150 100
#> 19     10   hp -0.03177295 0.009029704 3.440 100
#> 20     10   wt -3.87783074 0.632733322 3.440 100
#> 21     11   hp -0.03177295 0.009029704 3.440 100
#> 22     11   wt -3.87783074 0.632733322 3.440 100
#> 23     12   wt -3.87783074 0.632733547 4.070 100
#> 24     12   hp -0.03177295 0.009029704 4.070 100
#> 25     13   wt -3.87783074 0.632733547 3.730 100
#> 26     13   hp -0.03177295 0.009029704 3.730 100
#> 27     14   hp -0.03177295 0.009029704 3.780 100
#> 28     14   wt -3.87783074 0.632733322 3.780 100
#> 29     15   hp -0.03177295 0.009029704 5.250 100
#> 30     15   wt -3.87783074 0.632733547 5.250 100
#> 31     16   hp -0.03177295 0.009029704 5.424 100
#> 32     16   wt -3.87783074 0.632733322 5.424 100
#> 33     17   wt -3.87783074 0.632733547 5.345 100
#> 34     17   hp -0.03177295 0.009029704 5.345 100
#> 35     18   hp -0.03177295 0.009029704 2.200 100
#> 36     18   wt -3.87783074 0.632733322 2.200 100
#> 37     19   hp -0.03177295 0.009029704 1.615 100
#> 38     19   wt -3.87783074 0.632733772 1.615 100
#> 39     20   hp -0.03177295 0.009029704 1.835 100
#> 40     20   wt -3.87783074 0.632733322 1.835 100
#> 41     21   wt -3.87783074 0.632733097 2.465 100
#> 42     21   hp -0.03177295 0.009029704 2.465 100
#> 43     22   hp -0.03177295 0.009029704 3.520 100
#> 44     22   wt -3.87783074 0.632733322 3.520 100
#> 45     23   hp -0.03177295 0.009029704 3.435 100
#> 46     23   wt -3.87783074 0.632733547 3.435 100
#> 47     24   hp -0.03177295 0.009029704 3.840 100
#> 48     24   wt -3.87783074 0.632733996 3.840 100
#> 49     25   wt -3.87783074 0.632733322 3.845 100
#> 50     25   hp -0.03177295 0.009029704 3.845 100
#> 51     26   wt -3.87783074 0.632733322 1.935 100
#> 52     26   hp -0.03177295 0.009029704 1.935 100
#> 53     27   wt -3.87783074 0.632733547 2.140 100
#> 54     27   hp -0.03177295 0.009029704 2.140 100
#> 55     28   hp -0.03177295 0.009029704 1.513 100
#> 56     28   wt -3.87783074 0.632733322 1.513 100
#> 57     29   hp -0.03177295 0.009029704 3.170 100
#> 58     29   wt -3.87783074 0.632733547 3.170 100
#> 59     30   wt -3.87783074 0.632733547 2.770 100
#> 60     30   hp -0.03177295 0.009029704 2.770 100
#> 61     31   wt -3.87783074 0.632733322 3.570 100
#> 62     31   hp -0.03177295 0.009029704 3.570 100
#> 63     32   hp -0.03177295 0.009029704 2.780 100
#> 64     32   wt -3.87783074 0.632733322 2.780 100
#> 65     33   hp -0.03177295 0.009029710 2.620 110
#> 66     33   wt -3.87783074 0.632733547 2.620 110
#> 67     34   wt -3.87783074 0.632733322 2.875 110
#> 68     34   hp -0.03177295 0.009029710 2.875 110
#> 69     35   wt -3.87783074 0.632733097 2.320 110
#> 70     35   hp -0.03177295 0.009029710 2.320 110
#> 71     36   hp -0.03177295 0.009029710 3.215 110
#> 72     36   wt -3.87783074 0.632733322 3.215 110
#> 73     37   hp -0.03177295 0.009029710 3.440 110
#> 74     37   wt -3.87783074 0.632733322 3.440 110
#> 75     38   wt -3.87783074 0.632733547 3.460 110
#> 76     38   hp -0.03177295 0.009029710 3.460 110
#> 77     39   wt -3.87783074 0.632733322 3.570 110
#> 78     39   hp -0.03177295 0.009029710 3.570 110
#> 79     40   wt -3.87783074 0.632733322 3.190 110
#> 80     40   hp -0.03177295 0.009029710 3.190 110
#> 81     41   hp -0.03177295 0.009029710 3.150 110
#> 82     41   wt -3.87783074 0.632733547 3.150 110
#> 83     42   hp -0.03177295 0.009029710 3.440 110
#> 84     42   wt -3.87783074 0.632733322 3.440 110
#> 85     43   hp -0.03177295 0.009029710 3.440 110
#> 86     43   wt -3.87783074 0.632733322 3.440 110
#> 87     44   wt -3.87783074 0.632733547 4.070 110
#> 88     44   hp -0.03177295 0.009029710 4.070 110
#> 89     45   hp -0.03177295 0.009029710 3.730 110
#> 90     45   wt -3.87783074 0.632733547 3.730 110
#> 91     46   hp -0.03177295 0.009029710 3.780 110
#> 92     46   wt -3.87783074 0.632733322 3.780 110
#> 93     47   hp -0.03177295 0.009029710 5.250 110
#> 94     47   wt -3.87783074 0.632733547 5.250 110
#> 95     48   wt -3.87783074 0.632733322 5.424 110
#> 96     48   hp -0.03177295 0.009029710 5.424 110
#> 97     49   hp -0.03177295 0.009029710 5.345 110
#> 98     49   wt -3.87783074 0.632733547 5.345 110
#> 99     50   hp -0.03177295 0.009029710 2.200 110
#> 100    50   wt -3.87783074 0.632733322 2.200 110
#> 101    51   hp -0.03177295 0.009029710 1.615 110
#> 102    51   wt -3.87783074 0.632733772 1.615 110
#> 103    52   wt -3.87783074 0.632733322 1.835 110
#> 104    52   hp -0.03177295 0.009029710 1.835 110
#> 105    53   wt -3.87783074 0.632733097 2.465 110
#> 106    53   hp -0.03177295 0.009029710 2.465 110
#> 107    54   hp -0.03177295 0.009029710 3.520 110
#> 108    54   wt -3.87783074 0.632733322 3.520 110
#> 109    55   hp -0.03177295 0.009029710 3.435 110
#> 110    55   wt -3.87783074 0.632733547 3.435 110
#> 111    56   hp -0.03177295 0.009029710 3.840 110
#> 112    56   wt -3.87783074 0.632733996 3.840 110
#> 113    57   wt -3.87783074 0.632733322 3.845 110
#> 114    57   hp -0.03177295 0.009029710 3.845 110
#> 115    58   hp -0.03177295 0.009029710 1.935 110
#> 116    58   wt -3.87783074 0.632733322 1.935 110
#> 117    59   hp -0.03177295 0.009029710 2.140 110
#> 118    59   wt -3.87783074 0.632733547 2.140 110
#> 119    60   hp -0.03177295 0.009029710 1.513 110
#> 120    60   wt -3.87783074 0.632733322 1.513 110
#> 121    61   wt -3.87783074 0.632733547 3.170 110
#> 122    61   hp -0.03177295 0.009029710 3.170 110
#> 123    62   hp -0.03177295 0.009029710 2.770 110
#> 124    62   wt -3.87783074 0.632733547 2.770 110
#> 125    63   hp -0.03177295 0.009029710 3.570 110
#> 126    63   wt -3.87783074 0.632733322 3.570 110
#> 127    64   hp -0.03177295 0.009029710 2.780 110
#> 128    64   wt -3.87783074 0.632733322 2.780 110