A "counterfactual" version of the datagrid() function.

For each combination of the variable values specified, this function duplicates the entire data frame supplied to newdata, or the entire dataset used to fit model. This is a convenience shortcut to call the datagrid() function with argument grid_type="counterfactual".

datagridcf(..., model = NULL, newdata = NULL)

Arguments

...

named arguments with vectors of values or functions for user-specified variables.

Functions are applied to the variable in the model dataset or newdata, and must return a vector of the appropriate type.

Character vectors are automatically transformed to factors if necessary. +The output will include all combinations of these variables (see Examples below.)

model

Model object

newdata

data.frame (one and only one of the model and newdata arguments

Examples

# Fit a model with 32 observations from the `mtcars` dataset.
nrow(mtcars)
#> [1] 32

mod <- lm(mpg ~ hp + am, data = mtcars)

# We specify two values for the `am` variable and obtain a counterfactual
# dataset with 64 observations (32 x 2).
dat <- datagridcf(model = mod, am = 0:1)
head(dat)
#>   rowidcf  mpg  hp am
#> 1       1 21.0 110  0
#> 2       2 21.0 110  0
#> 3       3 22.8  93  0
#> 4       4 21.4 110  0
#> 5       5 18.7 175  0
#> 6       6 18.1 105  0
nrow(dat)
#> [1] 64

# We specify 2 values for the `am` variable and 3 values for the `hp` variable
# and obtained a dataset with 192 observations (2x3x32), corresponding to the
# full original data, with each possible combination of `hp` and `am`.
dat <- datagridcf(am = 0:1, hp = c(100, 110, 120), newdata = mtcars)
head(dat)
#>   rowidcf  mpg cyl disp drat    wt  qsec vs gear carb am  hp
#> 1       1 21.0   6  160 3.90 2.620 16.46  0    4    4  0 100
#> 2       2 21.0   6  160 3.90 2.875 17.02  0    4    4  0 100
#> 3       3 22.8   4  108 3.85 2.320 18.61  1    4    1  0 100
#> 4       4 21.4   6  258 3.08 3.215 19.44  1    3    1  0 100
#> 5       5 18.7   8  360 3.15 3.440 17.02  0    3    2  0 100
#> 6       6 18.1   6  225 2.76 3.460 20.22  1    3    1  0 100
dim(dat)
#> [1] 192  12