meffects() is a shortcut to marginaleffects()

Warning: This package is experimental.

meffects(
  model,
  newdata = NULL,
  variables = NULL,
  vcov = TRUE,
  numDeriv_method = "simple",
  prediction_type = "response",
  return_data = TRUE,
  ...
)

Arguments

model
Model object

newdata
A dataset over which to compute marginal effects. NULL uses the original data used to fit the model.

variables
Variables to consider (character vector). NULL calculates marginal effects for all terms in the model object.

vcov
Matrix or boolean

FALSE: does not compute unit-level standard errors.

TRUE: computes unit-level standard errors using the default vcov(model) variance-covariance matrix.

Named square matrix: computes standard errors with a user-supplied variance-covariance matrix. This matrix must be square and have dimensions equal to the number of coefficients in get_coef(model).

numDeriv_method
One of "simple", "Richardson", or "complex", indicating the method to use for the approximation. See numDeriv::grad() for details.

prediction_type
Type of prediction as character. This can differ based on the model type, but will typically be a string such as: "response", "link", or "probs".

return_data
boolean If TRUE, the original data used to fit the model is attached to the output. FALSE will objects which take up less space in memory.

...
Additional arguments are ignored.

Details

A "marginal effect" is the partial derivative of the regression equation with respect to a variable in the model. This package uses automatic differentiation tu compute marginal effects for a vast array of models, including non-linear models with transformations (e.g., polynomials).

Most -- but not all -- of the models below have been checked against alternative software packages to validate results. Visit the marginaleffects website to learn more about the testing procedure and coverage.

Supported models:

stats::lm

stats::glm

stats::loess (no variance)

AER::ivreg

betareg::betareg

fixest::feols

fixest::feglm

lme4::lmer

lme4::glmer

MASS::polr (no variance)

ordinal::clm (no variance)

survey::svyglm

Examples

mod <- glm(am ~ hp * wt, data = mtcars, family = binomial)
mfx <- marginaleffects(mod)
summary(mfx)

#> Average marginal effects 
#>   Term   Effect Std. Error  z value Pr(>|z|)    2.5 %  97.5 %
#> 1   hp  0.00265    0.00356  0.74512  0.45620 -0.00433 0.00963
#> 2   wt -0.43573    0.48985 -0.88952  0.37372 -1.39581 0.52435
#> 
#> Model type:  glm 
#> Prediction type:  response 

tidy(mfx)

#> # A tibble: 2 × 7
#> # Groups:   term [2]
#>   term  estimate std.error statistic p.value conf.low conf.high
#>   <chr>    <dbl>     <dbl>     <dbl>   <dbl>    <dbl>     <dbl>
#> 1 hp     0.00265   0.00356     0.745   0.456 -0.00433   0.00963
#> 2 wt    -0.436     0.490      -0.890   0.374 -1.40      0.524  

head(mfx)

#>   rowid term         dydx   std.error am  hp    wt
#> 1     1   hp  0.006989229 0.005882349  1 110 2.620
#> 2     1   wt -0.827161739 0.819043730  1 110 2.620
#> 3     2   hp  0.016403141 0.013387937  1 110 2.875
#> 4     2   wt -2.253398912 1.128058512  1 110 2.875
#> 5     3   wt -0.265544570 0.321881990  1  93 2.320
#> 6     3   hp  0.002831769 0.003755191  1  93 2.320

if (FALSE) {
plot(mfx)
}

# typical marginal effects
marginaleffects(mod, 
                newdata = typical(hp = c(100, 110)))

#>   rowid term          dydx    std.error    wt  hp
#> 1     1   hp  0.0004318229 0.0008216888 3.325 100
#> 2     1   wt -0.0790665586 0.1554274605 3.325 100
#> 3     2   hp  0.0006842470 0.0011556819 3.325 110
#> 4     2   wt -0.1313277736 0.2268135899 3.325 110

# counterfactual average marginal effects
marginaleffects(mod, 
                newdata = counterfactual(hp = c(100, 110)))

#>     rowid term          dydx    std.error    wt  hp
#> 1       1   hp  1.204245e-02 9.995994e-03 2.620 100
#> 2       1   wt -1.359471e+00 8.718971e-01 2.620 100
#> 3       2   hp  1.415558e-02 1.080248e-02 2.875 100
#> 4       2   wt -1.854887e+00 9.071250e-01 2.875 100
#> 5       3   hp  1.566074e-03 2.204571e-03 2.320 100
#> 6       3   wt -1.519931e-01 2.308674e-01 2.320 100
#> 7       4   wt -1.985550e-01 3.116129e-01 3.215 100
#> 8       4   hp  1.189815e-03 1.779620e-03 3.215 100
#> 9       5   hp  1.453976e-04 3.409128e-04 3.440 100
#> 10      5   wt -2.963387e-02 7.036292e-02 3.440 100
#> 11      6   hp  1.200600e-04 2.909609e-04 3.460 100
#> 12      6   wt -2.496082e-02 6.103757e-02 3.460 100
#> 13      7   hp  4.145963e-05 1.189714e-04 3.570 100
#> 14      7   wt -9.689079e-03 2.748035e-02 3.570 100
#> 15      8   wt -2.436247e-01 3.600688e-01 3.190 100
#> 16      8   hp  1.489280e-03 2.093504e-03 3.190 100
#> 17      9   hp  2.119011e-03 2.680346e-03 3.150 100
#> 18      9   wt -3.360304e-01 4.471099e-01 3.150 100
#> 19     10   hp  1.453976e-04 3.409128e-04 3.440 100
#> 20     10   wt -2.963387e-02 7.036292e-02 3.440 100
#> 21     11   hp  1.453976e-04 3.409128e-04 3.440 100
#> 22     11   wt -2.963387e-02 7.036292e-02 3.440 100
#> 23     12   wt -1.294632e-04 6.017685e-04 4.070 100
#> 24     12   hp  2.415339e-07 1.361964e-06 4.070 100
#> 25     13   wt -2.437708e-03 8.311388e-03 3.730 100
#> 26     13   hp  8.548397e-06 3.053517e-05 3.730 100
#> 27     14   hp  5.170044e-06 1.970122e-05 3.780 100
#> 28     14   wt -1.583280e-03 5.684000e-03 3.780 100
#> 29     15   hp -1.862539e-11 1.590175e-10 5.250 100
#> 30     15   wt -4.863231e-09 4.406971e-08 5.250 100
#> 31     16   hp -5.054491e-12 4.636245e-11 5.424 100
#> 32     16   wt -1.082406e-09 1.052207e-08 5.424 100
#> 33     17   wt -2.141194e-09 2.017329e-08 5.345 100
#> 34     17   hp -9.182237e-12 8.154875e-11 5.345 100
#> 35     18   hp  6.006465e-04 1.053596e-03 2.200 100
#> 36     18   wt -5.519541e-02 1.047910e-01 2.200 100
#> 37     19   hp  4.902651e-06 1.789468e-05 1.615 100
#> 38     19   wt -3.577723e-04 1.377302e-03 1.615 100
#> 39     20   hp  3.021851e-05 8.798615e-05 1.835 100
#> 40     20   wt -2.390264e-03 7.406607e-03 1.835 100
#> 41     21   wt -4.870405e-01 5.208582e-01 2.465 100
#> 42     21   hp  4.677862e-03 4.884891e-03 2.465 100
#> 43     22   hp  6.736539e-05 1.794532e-04 3.520 100
#> 44     22   wt -1.490273e-02 3.961604e-02 3.520 100
#> 45     23   hp  1.525131e-04 3.546083e-04 3.435 100
#> 46     23   wt -3.093196e-02 7.289668e-02 3.435 100
#> 47     24   hp  2.806815e-06 1.154864e-05 3.840 100
#> 48     24   wt -9.432139e-04 3.591236e-03 3.840 100
#> 49     25   wt -9.033649e-04 3.455916e-03 3.845 100
#> 50     25   hp  2.666431e-06 1.104133e-05 3.845 100
#> 51     26   wt -5.664166e-03 1.565144e-02 1.935 100
#> 52     26   hp  6.887674e-05 1.779961e-04 1.935 100
#> 53     27   wt -3.304863e-02 6.910542e-02 2.140 100
#> 54     27   hp  3.692031e-04 7.146808e-04 2.140 100
#> 55     28   hp  2.104980e-06 8.419538e-06 1.513 100
#> 56     28   wt -1.482885e-04 6.231705e-04 1.513 100
#> 57     29   hp  1.778434e-03 2.373731e-03 3.170 100
#> 58     29   wt -2.864049e-01 4.022327e-01 3.170 100
#> 59     30   wt -2.151319e+00 9.807785e-01 2.770 100
#> 60     30   hp  1.750450e-02 1.426497e-02 2.770 100
#> 61     31   wt -9.689079e-03 2.748035e-02 3.570 100
#> 62     31   hp  4.145963e-05 1.189714e-04 3.570 100
#> 63     32   hp  1.745714e-02 1.417906e-02 2.780 100
#> 64     32   wt -2.158255e+00 9.780962e-01 2.780 100
#> 65     33   hp  6.989229e-03 5.882349e-03 2.620 110
#> 66     33   wt -8.271617e-01 8.190437e-01 2.620 110
#> 67     34   wt -2.253399e+00 1.128059e+00 2.875 110
#> 68     34   hp  1.640314e-02 1.338794e-02 2.875 110
#> 69     35   wt -6.682713e-02 1.405033e-01 2.320 110
#> 70     35   hp  6.568482e-04 1.192287e-03 2.320 110
#> 71     36   hp  1.933575e-03 2.449835e-03 3.215 110
#> 72     36   wt -3.382383e-01 4.418278e-01 3.215 110
#> 73     37   hp  2.212605e-04 4.799191e-04 3.440 110
#> 74     37   wt -4.727047e-02 1.025827e-01 3.440 110
#> 75     38   wt -3.951143e-02 8.879671e-02 3.460 110
#> 76     38   hp  1.813040e-04 4.089426e-04 3.460 110
#> 77     39   wt -1.467921e-02 3.925378e-02 3.570 110
#> 78     39   hp  5.992269e-05 1.647079e-04 3.570 110
#> 79     40   wt -4.158796e-01 5.042287e-01 3.190 110
#> 80     40   hp  2.425271e-03 2.862491e-03 3.190 110
#> 81     41   hp  3.446649e-03 3.632695e-03 3.150 110
#> 82     41   wt -5.729388e-01 6.108977e-01 3.150 110
#> 83     42   hp  2.212605e-04 4.799191e-04 3.440 110
#> 84     42   wt -4.727047e-02 1.025827e-01 3.440 110
#> 85     43   hp  2.212605e-04 4.799191e-04 3.440 110
#> 86     43   wt -4.727047e-02 1.025827e-01 3.440 110
#> 87     44   wt -1.594167e-04 7.437378e-04 4.070 110
#> 88     44   hp  2.837340e-07 1.656247e-06 4.070 110
#> 89     45   hp  1.156780e-05 4.085083e-05 3.730 110
#> 90     45   wt -3.457815e-03 1.141872e-02 3.730 110
#> 91     46   hp  6.852489e-06 2.602779e-05 3.780 110
#> 92     46   wt -2.199713e-03 7.700145e-03 3.780 110
#> 93     47   hp -1.338260e-11 1.169300e-10 5.250 110
#> 94     47   wt -3.662809e-09 3.495688e-08 5.250 110
#> 95     48   wt -7.582238e-10 7.789755e-09 5.424 110
#> 96     48   hp -3.377772e-12 3.200157e-11 5.424 110
#> 97     49   hp -6.341546e-12 5.794856e-11 5.345 110
#> 98     49   wt -1.550090e-09 1.541101e-08 5.345 110
#> 99     50   hp  2.364559e-04 5.280246e-04 2.200 110
#> 100    50   wt -2.277759e-02 5.761784e-02 2.200 110
#> 101    51   hp  1.500828e-06 6.574830e-06 1.615 110
#> 102    51   wt -1.148098e-04 5.381970e-04 1.615 110
#> 103    52   wt -8.409269e-04 3.247327e-03 1.835 110
#> 104    52   hp  1.014175e-05 3.605226e-05 1.835 110
#> 105    53   wt -2.385602e-01 3.772460e-01 2.465 110
#> 106    53   hp  2.185727e-03 2.888593e-03 2.465 110
#> 107    54   hp  9.935017e-05 2.504389e-04 3.520 110
#> 108    54   wt -2.303839e-02 5.712794e-02 3.520 110
#> 109    55   hp  2.325302e-04 4.993649e-04 3.435 110
#> 110    55   wt -4.943500e-02 1.063260e-01 3.435 110
#> 111    56   hp  3.628574e-06 1.501668e-05 3.840 110
#> 112    56   wt -1.278164e-03 4.779723e-03 3.840 110
#> 113    57   wt -1.221621e-03 4.592673e-03 3.845 110
#> 114    57   hp  3.439928e-06 1.433750e-05 3.845 110
#> 115    58   hp  2.411108e-05 7.683342e-05 1.935 110
#> 116    58   wt -2.078512e-03 7.256450e-03 1.935 110
#> 117    59   hp  1.413180e-04 3.456070e-04 2.140 110
#> 118    59   wt -1.326046e-02 3.639522e-02 2.140 110
#> 119    60   hp  6.175582e-07 2.967521e-06 1.513 110
#> 120    60   wt -4.560477e-05 2.314894e-04 1.513 110
#> 121    61   wt -4.889993e-01 5.567625e-01 3.170 110
#> 122    61   hp  2.896689e-03 3.229517e-03 3.170 110
#> 123    62   hp  1.491001e-02 1.055618e-02 2.770 110
#> 124    62   wt -1.921154e+00 9.700499e-01 2.770 110
#> 125    63   hp  5.992269e-05 1.647079e-04 3.570 110
#> 126    63   wt -1.467921e-02 3.925378e-02 3.570 110
#> 127    64   hp  1.532558e-02 1.099711e-02 2.780 110
#> 128    64   wt -1.986443e+00 9.750434e-01 2.780 110

# heteroskedasticity robust standard errors
marginaleffects(mod, vcov = sandwich::vcovHC(mod))

#>    rowid term          dydx    std.error am  hp    wt
#> 1      1   hp  6.989229e-03 9.188426e-03  1 110 2.620
#> 2      1   wt -8.271617e-01 1.279010e+00  1 110 2.620
#> 3      2   hp  1.640314e-02 1.340865e-02  1 110 2.875
#> 4      2   wt -2.253399e+00 1.360905e+00  1 110 2.875
#> 5      3   wt -2.655446e-01 5.253557e-01  1  93 2.320
#> 6      3   hp  2.831769e-03 4.916395e-03  1  93 2.320
#> 7      4   wt -3.382383e-01 3.505055e-01  0 110 3.215
#> 8      4   hp  1.933575e-03 1.842493e-03  0 110 3.215
#> 9      5   hp  2.991347e-03 2.781261e-03  0 175 3.440
#> 10     5   wt -8.302249e-01 6.823524e-01  0 175 3.440
#> 11     6   hp  1.475603e-04 2.544284e-04  0 105 3.460
#> 12     6   wt -3.141797e-02 5.517898e-02  0 105 3.460
#> 13     7   wt -2.279347e+00 2.474881e+00  0 245 3.570
#> 14     7   hp  5.738829e-03 8.191035e-03  0 245 3.570
#> 15     8   wt -2.819198e-02 5.701635e-02  0  62 3.190
#> 16     8   hp  2.110144e-04 3.884630e-04  0  62 3.190
#> 17     9   wt -2.546923e-01 2.860002e-01  0  95 3.150
#> 18     9   hp  1.645791e-03 1.726980e-03  0  95 3.150
#> 19    10   hp  3.807685e-04 5.148223e-04  0 123 3.440
#> 20    10   wt -8.621334e-02 1.158666e-01  0 123 3.440
#> 21    11   wt -8.621334e-02 1.158666e-01  0 123 3.440
#> 22    11   hp  3.807685e-04 5.148223e-04  0 123 3.440
#> 23    12   wt -6.505458e-04 3.315794e-03  0 180 4.070
#> 24    12   hp  8.758474e-07 6.572651e-06  0 180 4.070
#> 25    13   wt -3.782430e-02 9.184598e-02  0 180 3.730
#> 26    13   hp  9.571749e-05 3.218446e-04  0 180 3.730
#> 27    14   hp  4.913415e-05 1.903042e-04  0 180 3.780
#> 28    14   wt -2.085110e-02 5.870368e-02  0 180 3.780
#> 29    15   hp -5.789952e-13 8.597919e-12  0 205 5.250
#> 30    15   wt -2.277157e-10 3.866959e-09  0 205 5.250
#> 31    16   wt -1.632782e-11 3.223634e-10  0 215 5.424
#> 32    16   hp -4.904969e-14 8.740137e-13  0 215 5.424
#> 33    17   wt -2.832600e-11 5.807088e-10  0 230 5.345
#> 34    17   hp -7.467144e-14 1.373822e-12  0 230 5.345
#> 35    18   hp  1.108683e-02 1.422135e-02  1  66 2.200
#> 36    18   wt -8.515432e-01 8.227415e-01  1  66 2.200
#> 37    19   hp  1.405293e-03 4.020638e-03  1  52 1.615
#> 38    19   wt -7.879409e-02 1.840977e-01  1  52 1.615
#> 39    20   wt -8.861724e-02 1.767656e-01  1  65 1.835
#> 40    20   hp  1.348054e-03 2.935489e-03  1  65 1.835
#> 41    21   wt -5.952843e-01 9.374444e-01  0  97 2.465
#> 42    21   hp  5.801587e-03 7.810194e-03  0  97 2.465
#> 43    22   wt -1.274388e-01 1.579215e-01  0 150 3.520
#> 44    22   hp  4.641522e-04 6.989948e-04  0 150 3.520
#> 45    23   hp  1.215176e-03 1.226520e-03  0 150 3.435
#> 46    23   wt -3.058856e-01 2.816151e-01  0 150 3.435
#> 47    24   wt -6.583631e-02 2.836171e-01  0 245 3.840
#> 48    24   hp  1.153078e-04 6.933458e-04  0 245 3.840
#> 49    25   wt -8.303109e-03 2.718630e-02  0 175 3.845
#> 50    25   hp  1.799896e-05 8.051093e-05  0 175 3.845
#> 51    26   wt -1.610539e-01 2.839053e-01  1  66 1.935
#> 52    26   hp  2.342925e-03 4.557330e-03  1  66 1.935
#> 53    27   hp  8.707316e-04 1.977080e-03  1  91 2.140
#> 54    27   wt -7.455609e-02 1.882778e-01  1  91 2.140
#> 55    28   hp  4.274703e-07 2.710374e-06  1 113 1.513
#> 56    28   wt -3.200338e-05 2.157386e-04  1 113 1.513
#> 57    29   wt -6.565758e-02 3.314830e-01  1 264 3.170
#> 58    29   hp  2.273419e-04 1.131595e-03  1 264 3.170
#> 59    30   wt -5.330220e-02 2.398867e-01  1 175 2.770
#> 60    30   hp  3.182868e-04 1.378430e-03  1 175 2.770
#> 61    31   hp  4.212583e-03 7.657185e-03  1 335 3.570
#> 62    31   wt -2.103001e+00 4.578801e+00  1 335 3.570
#> 63    32   hp  1.568764e-02 7.585714e-03  1 109 2.780
#> 64    32   wt -2.023983e+00 7.493504e-01  1 109 2.780