Slopes (aka Partial derivatives, Marginal Effects, or Trends)

Partial derivative of the regression equation with respect to a regressor of interest.

slopes(): unit-level (conditional) estimates.

avg_slopes(): average (marginal) estimates.

The newdata argument and the datagrid() function can be used to control where statistics are evaluated in the predictor space: "at observed values", "at the mean", "at representative values", etc.

See the slopes vignette and package website for worked examples and case studies:

https://vincentarelbundock.github.io/marginaleffects/articles/slopes.html

https://vincentarelbundock.github.io/marginaleffects/

Usage

slopes(
  model,
  newdata = NULL,
  variables = NULL,
  type = NULL,
  by = FALSE,
  vcov = TRUE,
  conf_level = 0.95,
  slope = "dydx",
  wts = NULL,
  hypothesis = NULL,
  df = Inf,
  eps = NULL,
  ...
)

avg_slopes(
  model,
  newdata = NULL,
  variables = NULL,
  type = NULL,
  by = TRUE,
  vcov = TRUE,
  conf_level = 0.95,
  slope = "dydx",
  wts = NULL,
  hypothesis = NULL,
  df = Inf,
  eps = NULL,
  ...
)

Arguments

model

Model object

newdata

Grid of predictor values at which we evaluate the slopes.

NULL (default): Unit-level slopes for each observed value in the original dataset.

data frame: Unit-level slopes for each row of the newdata data frame.

datagrid() call to specify a custom grid of regressors. For example:

newdata = datagrid(cyl = c(4, 6)): cyl variable equal to 4 and 6 and other regressors fixed at their means or modes.

See the Examples section and the datagrid() documentation.

string:

"mean": Marginal Effects at the Mean. Slopes when each predictor is held at its mean or mode.

"median": Marginal Effects at the Median. Slopes when each predictor is held at its median or mode.

"marginalmeans": Marginal Effects at Marginal Means. See Details section below.

"tukey": Marginal Effects at Tukey's 5 numbers.

"grid": Marginal Effects on a grid of representative numbers (Tukey's 5 numbers and unique values of categorical predictors).

variables

Focal variables

NULL: compute slopes or comparisons for all the variables in the model object (can be slow).

Character vector: subset of variables (usually faster).

type

string indicates the type (scale) of the predictions used to compute contrasts or slopes. This can differ based on the model type, but will typically be a string such as: "response", "link", "probs", or "zero". When an unsupported string is entered, the model-specific list of acceptable values is returned in an error message. When type is NULL, the default value is used. This default is the first model-related row in the marginaleffects:::type_dictionary dataframe.

by

Aggregate unit-level estimates (aka, marginalize, average over). Valid inputs:

FALSE: return the original unit-level estimates.

TRUE: aggregate estimates for each term.

Character vector of column names in newdata or in the data frame produced by calling the function without the by argument.

Data frame with a by column of group labels, and merging columns shared by newdata or the data frame produced by calling the same function without the by argument.

See examples below.

vcov

Type of uncertainty estimates to report (e.g., for robust standard errors). Acceptable values:

FALSE: Do not compute standard errors. This can speed up computation considerably.

TRUE: Unit-level standard errors using the default vcov(model) variance-covariance matrix.

String which indicates the kind of uncertainty estimates to return.

Heteroskedasticity-consistent: "HC", "HC0", "HC1", "HC2", "HC3", "HC4", "HC4m", "HC5". See ?sandwich::vcovHC

Heteroskedasticity and autocorrelation consistent: "HAC"

Mixed-Models degrees of freedom: "satterthwaite", "kenward-roger"

Other: "NeweyWest", "KernHAC", "OPG". See the sandwich package documentation.

One-sided formula which indicates the name of cluster variables (e.g., ~unit_id). This formula is passed to the cluster argument of the sandwich::vcovCL function.

Square covariance matrix

Function which returns a covariance matrix (e.g., stats::vcov(model))

conf_level

numeric value between 0 and 1. Confidence level to use to build a confidence interval.

slope

string indicates the type of slope or (semi-)elasticity to compute:

"dydx": dY/dX

"eyex": dY/dX * Y / X

"eydx": dY/dX * Y

"dyex": dY/dX / X

wts

string or numeric: weights to use when computing average contrasts or slopes. These weights only affect the averaging in avg_*() or with the by argument, and not the unit-level estimates themselves.

string: column name of the weights variable in newdata. When supplying a column name to wts, it is recommended to supply the original data (including the weights variable) explicitly to newdata.

numeric: vector of length equal to the number of rows in the original data or in newdata (if supplied).

hypothesis

specify a hypothesis test or custom contrast using a numeric value, vector, or matrix, a string, or a string formula.

Numeric:

Single value: the null hypothesis used in the computation of Z and p (before applying transform_post).

Vector: Weights to compute a linear combination of (custom contrast between) estimates. Length equal to the number of rows generated by the same function call, but without the hypothesis argument.

Matrix: Each column is a vector of weights, as describe above, used to compute a distinct linear combination of (contrast between) estimates. The column names of the matrix are used as labels in the output.

String formula to specify linear or non-linear hypothesis tests. If the term column uniquely identifies rows, terms can be used in the formula. Otherwise, use b1, b2, etc. to identify the position of each parameter. Examples:

hp = drat

hp + drat = 12

b1 + b2 + b3 = 0

String:

"pairwise": pairwise differences between estimates in each row.

"reference": differences between the estimates in each row and the estimate in the first row.

"sequential": difference between an estimate and the estimate in the next row.

"revpairwise", "revreference", "revsequential": inverse of the corresponding hypotheses, as described above.

See the Examples section below and the vignette: https://vincentarelbundock.github.io/marginaleffects/articles/hypothesis.html

df

Degrees of freedom used to compute p values and confidence intervals. A single numeric value between 1 and Inf. When df is Inf, the normal distribution is used. When df is finite, the t distribution is used. See insight::get_df for a convenient function to extract degrees of freedom. Ex: slopes(model, df = insight::get_df(model))

eps

NULL or numeric value which determines the step size to use when calculating numerical derivatives: (f(x+eps)-f(x))/eps. When eps is NULL, the step size is 0.0001 multiplied by the difference between the maximum and minimum values of the variable with respect to which we are taking the derivative. Changing eps may be necessary to avoid numerical problems in certain models.

...

Additional arguments are passed to the predict() method supplied by the modeling package.These arguments are particularly useful for mixed-effects or bayesian models (see the online vignettes on the marginaleffects website). Available arguments can vary from model to model, depending on the range of supported arguments by each modeling package. See the "Model-Specific Arguments" section of the ?marginaleffects documentation for a non-exhaustive list of available arguments.

Value

A data.frame with one row per observation (per term/group) and several columns:

rowid: row number of the newdata data frame

type: prediction type, as defined by the type argument

group: (optional) value of the grouped outcome (e.g., categorical outcome models)

term: the variable whose marginal effect is computed

dydx: slope of the outcome with respect to the term, for a given combination of predictor values

std.error: standard errors computed by via the delta method.

Details

A "slope" or "marginal effect" is the partial derivative of the regression equation with respect to a variable in the model. This function uses automatic differentiation to compute slopes for a vast array of models, including non-linear models with transformations (e.g., polynomials). Uncertainty estimates are computed using the delta method.

Numerical derivatives for the slopes function are calculated using a simple epsilon difference approach: \(\partial Y / \partial X = (f(X + \varepsilon) - f(X)) / \varepsilon\), where f is the predict() method associated with the model class, and \(\varepsilon\) is determined by the eps argument.

Warning: Some models are particularly sensitive to eps, so it is good practice to try different values of this argument.

Standard errors for the slopes are obtained using the Delta method. See the "Standard Errors" vignette on the package website for details (link above).

Functions

avg_slopes(): Average slopes

Model-Specific Arguments

Some model types allow model-specific arguments to modify the nature of marginal effects, predictions, marginal means, and contrasts.

Package	Class	Argument	Documentation

brms	brmsfit	ndraws	brms::posterior_predict

re_formula

lme4	merMod	include_random	insight::get_predicted

re.form	lme4::predict.merMod

allow.new.levels	lme4::predict.merMod

glmmTMB	glmmTMB	re.form	glmmTMB::predict.glmmTMB

allow.new.levels	glmmTMB::predict.glmmTMB

zitype	glmmTMB::predict.glmmTMB

mgcv	bam	exclude	mgcv::predict.bam

robustlmm	rlmerMod	re.form	robustlmm::predict.rlmerMod

allow.new.levels	robustlmm::predict.rlmerMod

MCMCglmm	MCMCglmm	ndraws

Bayesian posterior summaries

By default, credible intervals in bayesian models are built as equal-tailed intervals. This can be changed to a highest density interval by setting a global option:

options("marginaleffects_posterior_interval" = "eti")

options("marginaleffects_posterior_interval" = "hdi")

By default, the center of the posterior distribution in bayesian models is identified by the median. Users can use a different summary function by setting a global option:

options("marginaleffects_posterior_center" = "mean")

options("marginaleffects_posterior_center" = "median")

When estimates are averaged using the by argument, the tidy() function, or the summary() function, the posterior distribution is marginalized twice over. First, we take the average across units but within each iteration of the MCMC chain, according to what the user requested in by argument or tidy()/summary() functions. Then, we identify the center of the resulting posterior using the function supplied to the "marginaleffects_posterior_center" option (the median by default).

Examples

if (FALSE) { # interactive()
}

# Unit-level (conditional) Marginal Effects
mod <- glm(am ~ hp * wt, data = mtcars, family = binomial)
mfx <- slopes(mod)
head(mfx)
#> 
#>  Term Estimate Std. Error    z Pr(>|z|)    2.5 %  97.5 %
#>    hp  0.00698    0.00588 1.19     0.23 -0.00454 0.01851
#>    hp  0.01640    0.01339 1.23     0.22 -0.00984 0.04264
#>    hp  0.00283    0.00375 0.75     0.45 -0.00452 0.01018
#>    hp  0.00193    0.00245 0.79     0.43 -0.00287 0.00674
#>    hp  0.00299    0.00338 0.89     0.38 -0.00363 0.00961
#>    hp  0.00015    0.00035 0.43     0.67 -0.00053 0.00082
#> 
#> Prediction type:  response 
#> Columns: rowid, type, term, estimate, std.error, statistic, p.value, conf.low, conf.high, predicted, predicted_hi, predicted_lo, am, hp, wt, eps 
#> 

# Average Marginal Effect (AME)
avg_slopes(mod, by = TRUE)
#> 
#>  Term Estimate Std. Error    z Pr(>|z|)   2.5 %  97.5 %
#>    hp   0.0027     0.0019  1.4     0.17 -0.0011  0.0065
#>    wt  -0.4358     0.1021 -4.3   <0.001 -0.6358 -0.2357
#> 
#> Prediction type:  response 
#> Columns: type, term, estimate, std.error, statistic, p.value, conf.low, conf.high 
#> 

# Marginal Effect at the Mean (MEM)
slopes(mod, newdata = datagrid())
#> 
#>  Term Estimate Std. Error    z Pr(>|z|)   2.5 % 97.5 %  hp  wt
#>    hp   0.0085     0.0078  1.1     0.28 -0.0069  0.024 147 3.2
#>    wt  -1.7445     1.5863 -1.1     0.27 -4.8536  1.365 147 3.2
#> 
#> Prediction type:  response 
#> Columns: rowid, type, term, estimate, std.error, statistic, p.value, conf.low, conf.high, predicted, predicted_hi, predicted_lo, am, hp, wt, eps 
#> 

# Marginal Effect at User-Specified Values
# Variables not explicitly included in `datagrid()` are held at their means
slopes(mod, newdata = datagrid(hp = c(100, 110)))
#> 
#>  Term Estimate Std. Error     z Pr(>|z|)   2.5 % 97.5 %  wt  hp
#>    hp   0.0012     0.0018  0.67     0.51 -0.0023 0.0046 3.2 100
#>    hp   0.0019     0.0024  0.78     0.43 -0.0028 0.0066 3.2 110
#>    wt  -0.1947     0.3072 -0.63     0.53 -0.7968 0.4075 3.2 100
#>    wt  -0.3315     0.4361 -0.76     0.45 -1.1862 0.5232 3.2 110
#> 
#> Prediction type:  response 
#> Columns: rowid, type, term, estimate, std.error, statistic, p.value, conf.low, conf.high, predicted, predicted_hi, predicted_lo, am, wt, hp, eps 
#> 

# Group-Average Marginal Effects (G-AME)
# Calculate marginal effects for each observation, and then take the average
# marginal effect within each subset of observations with different observed
# values for the `cyl` variable:
mod2 <- lm(mpg ~ hp * cyl, data = mtcars)
avg_slopes(mod2, variables = "hp", by = "cyl")
#> 
#>  Term    Contrast cyl Estimate Std. Error     z Pr(>|z|)  2.5 % 97.5 %
#>    hp mean(dY/dX)   4   -0.092      0.035 -2.60   0.0094 -0.161 -0.022
#>    hp mean(dY/dX)   6   -0.052      0.020 -2.56   0.0104 -0.092 -0.012
#>    hp mean(dY/dX)   8   -0.013      0.014 -0.89   0.3728 -0.041  0.015
#> 
#> Prediction type:  response 
#> Columns: type, term, contrast, cyl, estimate, std.error, statistic, p.value, conf.low, conf.high, predicted, predicted_hi, predicted_lo 
#> 

# Marginal Effects at User-Specified Values (counterfactual)
# Variables not explicitly included in `datagrid()` are held at their
# original values, and the whole dataset is duplicated once for each
# combination of the values in `datagrid()`
mfx <- slopes(mod,
              newdata = datagrid(hp = c(100, 110),
              grid_type = "counterfactual"))
head(mfx)
#> 
#>  rowidcf Term Estimate Std. Error    z Pr(>|z|)    2.5 %  97.5 %  wt  hp
#>        1   hp  0.01203    0.00999 1.21     0.23 -0.00754 0.03161 2.6 100
#>        2   hp  0.01416    0.01081 1.31     0.19 -0.00702 0.03534 2.9 100
#>        3   hp  0.00156    0.00220 0.71     0.48 -0.00275 0.00588 2.3 100
#>        4   hp  0.00119    0.00178 0.67     0.50 -0.00230 0.00468 3.2 100
#>        5   hp  0.00015    0.00034 0.43     0.67 -0.00052 0.00081 3.4 100
#>        6   hp  0.00012    0.00029 0.41     0.68 -0.00045 0.00069 3.5 100
#> 
#> Prediction type:  response 
#> Columns: rowid, rowidcf, type, term, estimate, std.error, statistic, p.value, conf.low, conf.high, predicted, predicted_hi, predicted_lo, am, wt, hp, eps 
#> 

# Heteroskedasticity robust standard errors
mfx <- slopes(mod, vcov = sandwich::vcovHC(mod))
head(mfx)
#> 
#>  Term Estimate Std. Error    z Pr(>|z|)    2.5 %  97.5 %
#>    hp  0.00698    0.00919 0.76     0.45 -0.01102 0.02499
#>    hp  0.01640    0.01340 1.22     0.22 -0.00986 0.04267
#>    hp  0.00283    0.00491 0.58     0.56 -0.00680 0.01246
#>    hp  0.00193    0.00184 1.05     0.29 -0.00168 0.00555
#>    hp  0.00299    0.00278 1.08     0.28 -0.00246 0.00845
#>    hp  0.00015    0.00025 0.58     0.56 -0.00035 0.00065
#> 
#> Prediction type:  response 
#> Columns: rowid, type, term, estimate, std.error, statistic, p.value, conf.low, conf.high, predicted, predicted_hi, predicted_lo, am, hp, wt, eps 
#> 

# hypothesis test: is the `hp` marginal effect at the mean equal to the `drat` marginal effect
mod <- lm(mpg ~ wt + drat, data = mtcars)

slopes(
    mod,
    newdata = "mean",
    hypothesis = "wt = drat")
#> 
#>     Term Estimate Std. Error    z Pr(>|z|) 2.5 % 97.5 %
#>  wt=drat     -6.2        1.1 -5.9   <0.001  -8.3   -4.2
#> 
#> Prediction type:  response 
#> Columns: type, term, estimate, std.error, statistic, p.value, conf.low, conf.high 
#> 

# same hypothesis test using row indices
slopes(
    mod,
    newdata = "mean",
    hypothesis = "b1 - b2 = 0")
#> 
#>     Term Estimate Std. Error    z Pr(>|z|) 2.5 % 97.5 %
#>  b1-b2=0     -6.2        1.1 -5.9   <0.001  -8.3   -4.2
#> 
#> Prediction type:  response 
#> Columns: type, term, estimate, std.error, statistic, p.value, conf.low, conf.high 
#> 

# same hypothesis test using numeric vector of weights
slopes(
    mod,
    newdata = "mean",
    hypothesis = c(1, -1))
#> 
#>    Term Estimate Std. Error    z Pr(>|z|) 2.5 % 97.5 %
#>  custom     -6.2        1.1 -5.9   <0.001  -8.3   -4.2
#> 
#> Prediction type:  response 
#> Columns: type, term, estimate, std.error, statistic, p.value, conf.low, conf.high 
#> 

# two custom contrasts using a matrix of weights
lc <- matrix(c(
    1, -1,
    2, 3),
    ncol = 2)
colnames(lc) <- c("Contrast A", "Contrast B")
slopes(
    mod,
    newdata = "mean",
    hypothesis = lc)
#> 
#>        Term Estimate Std. Error     z Pr(>|z|) 2.5 % 97.5 %
#>  Contrast A     -6.2        1.1 -5.92   <0.001  -8.3   -4.2
#>  Contrast B     -5.2        5.6 -0.93     0.35 -16.3    5.8
#> 
#> Prediction type:  response 
#> Columns: type, term, estimate, std.error, statistic, p.value, conf.low, conf.high 
#>