Summarize a marginaleffects object

Summarize a marginaleffects object

# S3 method for marginaleffects
summary(object, conf.level = 0.95, by = NULL, ...)

Arguments

object

An object produced by the marginaleffects function

conf.level

The confidence level to use for the confidence interval if conf.int=TRUE. Must be strictly greater than 0 and less than 1. Defaults to 0.95, which corresponds to a 95 percent confidence interval.

by

Character vector of variable names over which to compute group-averaged marginal effects.

...

Additional arguments are passed to the predict() method used to compute adjusted predictions. These arguments are particularly useful for mixed-effects or bayesian models (see the online vignettes on the marginaleffects website). Available arguments can vary from model to model, depending on the range of supported arguments by each modeling package. See the "Model-Specific Arguments" section of the ?marginaleffects documentation for a non-exhaustive list of available arguments.

Value

Data frame of summary statistics for an object produced by the marginaleffects function

Examples

mod <- lm(mpg ~ hp * wt + factor(gear), data = mtcars)
mfx <- marginaleffects(mod)

# average marginal effects
summary(mfx)
#> Average marginal effects 
#>   Term Contrast   Effect Std. Error z value   Pr(>|z|)    2.5 %   97.5 %
#> 1   hp          -0.03533    0.01045 -3.3815 0.00072104 -0.05581 -0.01485
#> 2   wt          -3.52593    0.73202 -4.8167 1.4593e-06 -4.96065 -2.09120
#> 3 gear    4 - 3  0.82900    1.11804  0.7415 0.45840435 -1.36232  3.02032
#> 4 gear    5 - 3  1.81650    1.54761  1.1737 0.24049622 -1.21675  4.84975
#> 
#> Model type:  lm 
#> Prediction type:  response 

# average marginal effects by group
summary(mfx, by = "gear")
#> Average marginal effects 
#>    Term Contrast gear   Effect Std. Error z value   Pr(>|z|)    2.5 %    97.5 %
#> 1    hp             4 -0.05181    0.01142 -4.5352 5.7561e-06 -0.07420 -0.029418
#> 2    hp             3 -0.01680    0.01158 -1.4514 0.14667941 -0.03949  0.005888
#> 3    hp             5 -0.05137    0.01138 -4.5159 6.3062e-06 -0.07367 -0.029075
#> 4    wt             4 -5.09491    0.88830 -5.7356 9.7188e-09 -6.83594 -3.353869
#> 5    wt             3 -2.71806    0.74164 -3.6650 0.00024738 -4.17164 -1.264480
#> 6    wt             5 -2.18398    0.78317 -2.7886 0.00529297 -3.71896 -0.648994
#> 7  gear    4 - 3    4  0.82900    1.11804  0.7415 0.45840435 -1.36232  3.020322
#> 8  gear    4 - 3    3  0.82900    1.11804  0.7415 0.45840435 -1.36232  3.020322
#> 9  gear    4 - 3    5  0.82900    1.11804  0.7415 0.45840435 -1.36232  3.020322
#> 10 gear    5 - 3    4  1.81650    1.54761  1.1737 0.24049622 -1.21675  4.849748
#> 11 gear    5 - 3    3  1.81650    1.54761  1.1737 0.24049622 -1.21675  4.849748
#> 12 gear    5 - 3    5  1.81650    1.54761  1.1737 0.24049622 -1.21675  4.849748
#> 
#> Model type:  lm 
#> Prediction type:  response