Tidy a marginaleffects object

Tidy a marginaleffects object

# S3 method for marginaleffects
tidy(x, conf_level = 0.95, by = NULL, ...)

Arguments

x

An object produced by the marginaleffects function.

conf_level

numeric value between 0 and 1. Confidence level to use to build a confidence interval.

by

Character vector of variable names over which to compute group-averaged marginal effects.

...

Additional arguments are passed to the predict() method supplied by the modeling package.These arguments are particularly useful for mixed-effects or bayesian models (see the online vignettes on the marginaleffects website). Available arguments can vary from model to model, depending on the range of supported arguments by each modeling package. See the "Model-Specific Arguments" section of the ?marginaleffects documentation for a non-exhaustive list of available arguments.

Value

A "tidy" data.frame of summary statistics which conforms to the broom package specification.

Details

The tidy function calculates average marginal effects by taking the mean of all the unit-level marginal effects computed by the marginaleffects function.

The standard error of the average marginal effects is obtained by taking the mean of each column of the Jacobian. . Then, we use this "Jacobian at the mean" in the Delta method to obtained standard errors.

In Bayesian models (e.g., brms), we compute Average Marginal Effects by applying the mean function twice. First, we apply it to all marginal effects for each posterior draw, thereby estimating one Average (or Median) Marginal Effect per iteration of the MCMC chain. Second, we take the mean and quantile function to the results of Step 1 to obtain the Average (or Median) Marginal Effect and its associated interval.

Examples

mod <- lm(mpg ~ hp * wt + factor(gear), data = mtcars)
mfx <- marginaleffects(mod)

# average marginal effects
tidy(mfx)
#> Warning: column(s) not removed because not found: [marginaleffects_weights_internal]
#>       type term contrast    estimate  std.error  statistic      p.value
#> 1 response   hp    dY/dX -0.03533001 0.01044818 -3.3814514 7.210400e-04
#> 2 response   wt    dY/dX -3.52592595 0.73201579 -4.8167349 1.459264e-06
#> 3 response gear    4 - 3  0.82900152 1.11804097  0.7414769 4.584043e-01
#> 4 response gear    5 - 3  1.81649741 1.54760505  1.1737474 2.404962e-01
#>      conf.low   conf.high
#> 1 -0.05580807 -0.01485196
#> 2 -4.96065052 -2.09120137
#> 3 -1.36231851  3.02032155
#> 4 -1.21675276  4.84974757

# average marginal effects by group
tidy(mfx, by = "gear")
#> Warning: column(s) not removed because not found: [marginaleffects_weights_internal]
#>        type term contrast gear    estimate  std.error  statistic      p.value
#> 1  response   hp    dY/dX    4 -0.05180745 0.01142353 -4.5351547 5.756127e-06
#> 2  response   hp    dY/dX    3 -0.01680129 0.01157623 -1.4513608 1.466794e-01
#> 3  response   hp    dY/dX    5 -0.05137031 0.01137555 -4.5158527 6.306250e-06
#> 4  response   wt    dY/dX    4 -5.09490667 0.88830072 -5.7355652 9.718772e-09
#> 5  response   wt    dY/dX    3 -2.71805806 0.74163507 -3.6649535 2.473835e-04
#> 6  response   wt    dY/dX    5 -2.18397588 0.78316845 -2.7886413 5.292965e-03
#> 7  response gear    4 - 3    4  0.82900152 1.11804097  0.7414769 4.584043e-01
#> 8  response gear    4 - 3    3  0.82900152 1.11804097  0.7414769 4.584043e-01
#> 9  response gear    4 - 3    5  0.82900152 1.11804097  0.7414769 4.584043e-01
#> 10 response gear    5 - 3    4  1.81649741 1.54760505  1.1737474 2.404962e-01
#> 11 response gear    5 - 3    3  1.81649741 1.54760505  1.1737474 2.404962e-01
#> 12 response gear    5 - 3    5  1.81649741 1.54760505  1.1737474 2.404962e-01
#>       conf.low    conf.high
#> 1  -0.07419715 -0.029417756
#> 2  -0.03949030  0.005887709
#> 3  -0.07366598 -0.029074642
#> 4  -6.83594409 -3.353869244
#> 5  -4.17163610 -1.264480027
#> 6  -3.71895783 -0.648993930
#> 7  -1.36231851  3.020321553
#> 8  -1.36231851  3.020321553
#> 9  -1.36231851  3.020321553
#> 10 -1.21675276  4.849747574
#> 11 -1.21675276  4.849747574
#> 12 -1.21675276  4.849747574