Tidy a marginalmeans object

Tidy a marginalmeans object

Usage

# S3 method for marginalmeans
tidy(x, conf_level = 0.95, transform_avg = NULL, ...)

Arguments

x

An object produced by the marginalmeans function.

conf_level

numeric value between 0 and 1. Confidence level to use to build a confidence interval. The default NULL uses the conf_level value used in the original call to marginaleffects().

transform_avg

A function applied to the estimates and confidence intervals after the unit-level estimates have been averaged.

...

Additional arguments are passed to the predict() method supplied by the modeling package.These arguments are particularly useful for mixed-effects or bayesian models (see the online vignettes on the marginaleffects website). Available arguments can vary from model to model, depending on the range of supported arguments by each modeling package. See the "Model-Specific Arguments" section of the ?marginaleffects documentation for a non-exhaustive list of available arguments.

Value

A "tidy" data.frame of summary statistics which conforms to the broom package specification.

See also