Comparisons Between Predictions Made With Different Regressor Values

Predict the outcome variable at different regressor values (e.g., college graduates vs. others), and compare those predictions by computing a difference, ratio, or some other function. comparisons() can return many quantities of interest, such as contrasts, differences, risk ratios, changes in log odds, slopes, elasticities, etc.

comparisons(): unit-level (conditional) estimates.

avg_comparisons(): average (marginal) estimates.

The newdata argument and the datagrid() function can be used to control where statistics are evaluated in the predictor space: "at observed values", "at the mean", "at representative values", etc.

See the comparisons vignette and package website for worked examples and case studies:

https://vincentarelbundock.github.io/marginaleffects/articles/comparisons.html

https://vincentarelbundock.github.io/marginaleffects/

Usage

comparisons(
  model,
  newdata = NULL,
  variables = NULL,
  type = NULL,
  vcov = TRUE,
  by = FALSE,
  conf_level = 0.95,
  transform_pre = "difference",
  transform_post = NULL,
  cross = FALSE,
  wts = NULL,
  hypothesis = NULL,
  df = Inf,
  eps = NULL,
  ...
)

avg_comparisons(
  model,
  newdata = NULL,
  variables = NULL,
  type = NULL,
  vcov = TRUE,
  by = TRUE,
  conf_level = 0.95,
  transform_pre = "difference",
  transform_post = NULL,
  cross = FALSE,
  wts = NULL,
  hypothesis = NULL,
  df = Inf,
  eps = NULL,
  ...
)

Arguments

model

Model object

newdata

Grid of predictor values at which we evaluate the comparisons.

NULL (default): Unit-level contrasts for each observed value in the original dataset (empirical distribution).

data frame: Unit-level contrasts for each row of the newdata data frame.

string:

"mean": Contrasts at the Mean. Contrasts when each predictor is held at its mean or mode.

"median": Contrasts at the Median. Contrasts when each predictor is held at its median or mode.

"marginalmeans": Contrasts at Marginal Means.

"tukey": Contrasts at Tukey's 5 numbers.

"grid": Contrasts on a grid of representative numbers (Tukey's 5 numbers and unique values of categorical predictors).

datagrid() call to specify a custom grid of regressors. For example:

newdata = datagrid(cyl = c(4, 6)): cyl variable equal to 4 and 6 and other regressors fixed at their means or modes.

newdata = datagrid(mpg = fivenum): mpg variable held at Tukey's five numbers (using the fivenum function), and other regressors fixed at their means or modes.

See the Examples section and the datagrid documentation.

variables

Focal variables

NULL: compute comparisons for all the variables in the model object (can be slow).

Character vector: subset of variables (usually faster).

Named list: names identify the subset of variables of interest, and values define the type of contrast to compute. Acceptable values depend on the variable type:

Factor or character variables:

"reference": Each factor level is compared to the factor reference (base) level

"all": All combinations of observed levels

"sequential": Each factor level is compared to the previous factor level

"pairwise": Each factor level is compared to all other levels

"minmax": The highest and lowest levels of a factor.

Vector of length 2 with the two values to compare.

Logical variables:

NULL: contrast between TRUE and FALSE

Numeric variables:

Numeric of length 1: Contrast for a gap of x, computed at the observed value plus and minus x / 2. For example, estimating a +1 contrast compares adjusted predictions when the regressor is equal to its observed value minus 0.5 and its observed value plus 0.5.

Numeric vector of length 2: Contrast between the 2nd element and the 1st element of the x vector.

Function which accepts a numeric vector and returns a data frame with two columns of "low" and "high" values to compare. See examples below.

"iqr": Contrast across the interquartile range of the regressor.

"sd": Contrast across one standard deviation around the regressor mean.

"2sd": Contrast across two standard deviations around the regressor mean.

"minmax": Contrast between the maximum and the minimum values of the regressor.

Examples:

variables = list(gear = "pairwise", hp = 10)

variables = list(gear = "sequential", hp = c(100, 120))

See the Examples section below for more.

type

string indicates the type (scale) of the predictions used to compute contrasts or slopes. This can differ based on the model type, but will typically be a string such as: "response", "link", "probs", or "zero". When an unsupported string is entered, the model-specific list of acceptable values is returned in an error message. When type is NULL, the default value is used. This default is the first model-related row in the marginaleffects:::type_dictionary dataframe.

vcov

Type of uncertainty estimates to report (e.g., for robust standard errors). Acceptable values:

FALSE: Do not compute standard errors. This can speed up computation considerably.

TRUE: Unit-level standard errors using the default vcov(model) variance-covariance matrix.

String which indicates the kind of uncertainty estimates to return.

Heteroskedasticity-consistent: "HC", "HC0", "HC1", "HC2", "HC3", "HC4", "HC4m", "HC5". See ?sandwich::vcovHC

Heteroskedasticity and autocorrelation consistent: "HAC"

Mixed-Models degrees of freedom: "satterthwaite", "kenward-roger"

Other: "NeweyWest", "KernHAC", "OPG". See the sandwich package documentation.

One-sided formula which indicates the name of cluster variables (e.g., ~unit_id). This formula is passed to the cluster argument of the sandwich::vcovCL function.

Square covariance matrix

Function which returns a covariance matrix (e.g., stats::vcov(model))

by

Aggregate unit-level estimates (aka, marginalize, average over). Valid inputs:

FALSE: return the original unit-level estimates.

TRUE: aggregate estimates for each term.

Character vector of column names in newdata or in the data frame produced by calling the function without the by argument.

Data frame with a by column of group labels, and merging columns shared by newdata or the data frame produced by calling the same function without the by argument.

See examples below.

conf_level

numeric value between 0 and 1. Confidence level to use to build a confidence interval.

transform_pre

string or function. How should pairs of adjusted predictions be contrasted?

string: shortcuts to common contrast functions.

Supported shortcuts strings: difference, differenceavg, differenceavgwts, dydx, eyex, eydx, dyex, dydxavg, eyexavg, eydxavg, dyexavg, dydxavgwts, eyexavgwts, eydxavgwts, dyexavgwts, ratio, ratioavg, ratioavgwts, lnratio, lnratioavg, lnratioavgwts, lnor, lnoravg, lnoravgwts, expdydx, expdydxavg, expdydxavgwts

See the Transformations section below for definitions of each transformation.

function: accept two equal-length numeric vectors of adjusted predictions (hi and lo) and returns a vector of contrasts of the same length, or a unique numeric value.

See the Transformations section below for examples of valid functions.

transform_post

string or function. Transformation applied to unit-level estimates and confidence intervals just before the function returns results. Functions must accept a vector and return a vector of the same length. Support string shortcuts: "exp", "ln"

cross

TRUE or FALSE

FALSE: Contrasts represent the change in adjusted predictions when one predictor changes and all other variables are held constant.

TRUE: Contrasts represent the changes in adjusted predictions when all the predictors specified in the variables argument are manipulated simultaneously (a "cross-contrast").

wts

string or numeric: weights to use when computing average contrasts or slopes. These weights only affect the averaging in avg_*() or with the by argument, and not the unit-level estimates themselves.

string: column name of the weights variable in newdata. When supplying a column name to wts, it is recommended to supply the original data (including the weights variable) explicitly to newdata.

numeric: vector of length equal to the number of rows in the original data or in newdata (if supplied).

hypothesis

specify a hypothesis test or custom contrast using a numeric value, vector, or matrix, a string, or a string formula.

Numeric:

Single value: the null hypothesis used in the computation of Z and p (before applying transform_post).

Vector: Weights to compute a linear combination of (custom contrast between) estimates. Length equal to the number of rows generated by the same function call, but without the hypothesis argument.

Matrix: Each column is a vector of weights, as describe above, used to compute a distinct linear combination of (contrast between) estimates. The column names of the matrix are used as labels in the output.

String formula to specify linear or non-linear hypothesis tests. If the term column uniquely identifies rows, terms can be used in the formula. Otherwise, use b1, b2, etc. to identify the position of each parameter. Examples:

hp = drat

hp + drat = 12

b1 + b2 + b3 = 0

String:

"pairwise": pairwise differences between estimates in each row.

"reference": differences between the estimates in each row and the estimate in the first row.

"sequential": difference between an estimate and the estimate in the next row.

"revpairwise", "revreference", "revsequential": inverse of the corresponding hypotheses, as described above.

See the Examples section below and the vignette: https://vincentarelbundock.github.io/marginaleffects/articles/hypothesis.html

df

Degrees of freedom used to compute p values and confidence intervals. A single numeric value between 1 and Inf. When df is Inf, the normal distribution is used. When df is finite, the t distribution is used. See insight::get_df for a convenient function to extract degrees of freedom. Ex: slopes(model, df = insight::get_df(model))

eps

NULL or numeric value which determines the step size to use when calculating numerical derivatives: (f(x+eps)-f(x))/eps. When eps is NULL, the step size is 0.0001 multiplied by the difference between the maximum and minimum values of the variable with respect to which we are taking the derivative. Changing eps may be necessary to avoid numerical problems in certain models.

...

Additional arguments are passed to the predict() method supplied by the modeling package.These arguments are particularly useful for mixed-effects or bayesian models (see the online vignettes on the marginaleffects website). Available arguments can vary from model to model, depending on the range of supported arguments by each modeling package. See the "Model-Specific Arguments" section of the ?marginaleffects documentation for a non-exhaustive list of available arguments.

Functions

avg_comparisons(): Average comparisons

Model-Specific Arguments

Some model types allow model-specific arguments to modify the nature of marginal effects, predictions, marginal means, and contrasts.

Package	Class	Argument	Documentation

brms	brmsfit	ndraws	brms::posterior_predict

re_formula

lme4	merMod	include_random	insight::get_predicted

re.form	lme4::predict.merMod

allow.new.levels	lme4::predict.merMod

glmmTMB	glmmTMB	re.form	glmmTMB::predict.glmmTMB

allow.new.levels	glmmTMB::predict.glmmTMB

zitype	glmmTMB::predict.glmmTMB

mgcv	bam	exclude	mgcv::predict.bam

robustlmm	rlmerMod	re.form	robustlmm::predict.rlmerMod

allow.new.levels	robustlmm::predict.rlmerMod

MCMCglmm	MCMCglmm	ndraws

Transformations

The following transformations can be applied by supplying one of the shortcut strings to the transform_pre argument. hi is a vector of adjusted predictions for the "high" side of the contrast. lo is a vector of adjusted predictions for the "low" side of the contrast. y is a vector of adjusted predictions for the original data. x is the predictor in the original data. eps is the step size to use to compute derivatives and elasticities.

Shortcut	Function

difference	\(hi, lo) hi - lo

differenceavg	\(hi, lo) mean(hi) - mean(lo)

differenceavgwts	\(hi, lo, w) wmean(hi, w) - wmean(lo, w)

dydx	\(hi, lo, eps) (hi - lo)/eps

eyex	\(hi, lo, eps, y, x) (hi - lo)/eps * (x/y)

eydx	\(hi, lo, eps, y, x) ((hi - lo)/eps)/y

dyex	\(hi, lo, eps, x) ((hi - lo)/eps) * x

dydxavg	\(hi, lo, eps) mean((hi - lo)/eps)

eyexavg	\(hi, lo, eps, y, x) mean((hi - lo)/eps * (x/y))

eydxavg	\(hi, lo, eps, y, x) mean(((hi - lo)/eps)/y)

dyexavg	\(hi, lo, eps, x) mean(((hi - lo)/eps) * x)

dydxavgwts	\(hi, lo, eps, w) wmean((hi - lo)/eps, w)

eyexavgwts	\(hi, lo, eps, y, x, w) wmean((hi - lo)/eps * (x/y), w)

eydxavgwts	\(hi, lo, eps, y, x, w) wmean(((hi - lo)/eps)/y, w)

dyexavgwts	\(hi, lo, eps, x, w) wmean(((hi - lo)/eps) * x, w)

ratio	\(hi, lo) hi/lo

ratioavg	\(hi, lo) mean(hi)/mean(lo)

ratioavgwts	\(hi, lo) wmean(hi)/wmean(lo)

lnratio	\(hi, lo) log(hi/lo)

lnratioavg	\(hi, lo) log(mean(hi)/mean(lo))

lnratioavgwts	\(hi, lo) log(wmean(hi)/wmean(lo))

lnor	\(hi, lo) log((hi/(1 - hi))/(lo/(1 - lo)))

lnoravg	\(hi, lo) log((mean(hi)/(1 - mean(hi)))/(mean(lo)/(1 - mean(lo))))

lnoravgwts	\(hi, lo, w) log((wmean(hi, w)/(1 - wmean(hi, w)))/(wmean(lo, w)/(1 - wmean(lo, w))))

expdydx	\(hi, lo, eps) ((exp(hi) - exp(lo))/exp(eps))/eps

expdydxavg	\(hi, lo, eps) mean(((exp(hi) - exp(lo))/exp(eps))/eps)

expdydxavgwts	\(hi, lo, eps, w) wmean(((exp(hi) - exp(lo))/exp(eps))/eps, w)

Bayesian posterior summaries

By default, credible intervals in bayesian models are built as equal-tailed intervals. This can be changed to a highest density interval by setting a global option:

options("marginaleffects_posterior_interval" = "eti")

options("marginaleffects_posterior_interval" = "hdi")

By default, the center of the posterior distribution in bayesian models is identified by the median. Users can use a different summary function by setting a global option:

options("marginaleffects_posterior_center" = "mean")

options("marginaleffects_posterior_center" = "median")

When estimates are averaged using the by argument, the tidy() function, or the summary() function, the posterior distribution is marginalized twice over. First, we take the average across units but within each iteration of the MCMC chain, according to what the user requested in by argument or tidy()/summary() functions. Then, we identify the center of the resulting posterior using the function supplied to the "marginaleffects_posterior_center" option (the median by default).

Examples

library(marginaleffects)

# Linear model
tmp <- mtcars
tmp$am <- as.logical(tmp$am)
mod <- lm(mpg ~ am + factor(cyl), tmp)
comparisons(mod, variables = list(cyl = "reference")) |> tidy()
#> Warning: The `cyl` variable is treated as a categorical (factor) variable, but
#>   the original data is of class numeric. It is safer and faster to convert
#>   such variables to factor before fitting the model and calling `slopes`
#>   functions.
#>   
#>   This warning appears once per session.
#> # A tibble: 2 × 9
#>   type     term  contrast       estim…¹ std.e…² stati…³  p.value conf.…⁴ conf.…⁵
#>   <chr>    <chr> <chr>            <dbl>   <dbl>   <dbl>    <dbl>   <dbl>   <dbl>
#> 1 response cyl   mean(6) - mea…   -6.16    1.54   -4.01 6.11e- 5   -9.17   -3.15
#> 2 response cyl   mean(8) - mea…  -10.1     1.45   -6.93 4.11e-12  -12.9    -7.22
#> # … with abbreviated variable names ¹​estimate, ²​std.error, ³​statistic,
#> #   ⁴​conf.low, ⁵​conf.high
comparisons(mod, variables = list(cyl = "sequential")) |> tidy()
#> # A tibble: 2 × 9
#>   type     term  contrast        estim…¹ std.e…² stati…³ p.value conf.…⁴ conf.…⁵
#>   <chr>    <chr> <chr>             <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#> 1 response cyl   mean(6) - mean…   -6.16    1.54   -4.01 6.11e-5   -9.17   -3.15
#> 2 response cyl   mean(8) - mean…   -3.91    1.47   -2.66 7.81e-3   -6.79   -1.03
#> # … with abbreviated variable names ¹​estimate, ²​std.error, ³​statistic,
#> #   ⁴​conf.low, ⁵​conf.high
comparisons(mod, variables = list(cyl = "pairwise")) |> tidy()
#> # A tibble: 3 × 9
#>   type     term  contrast       estim…¹ std.e…² stati…³  p.value conf.…⁴ conf.…⁵
#>   <chr>    <chr> <chr>            <dbl>   <dbl>   <dbl>    <dbl>   <dbl>   <dbl>
#> 1 response cyl   mean(6) - mea…   -6.16    1.54   -4.01 6.11e- 5   -9.17   -3.15
#> 2 response cyl   mean(8) - mea…  -10.1     1.45   -6.93 4.11e-12  -12.9    -7.22
#> 3 response cyl   mean(8) - mea…   -3.91    1.47   -2.66 7.81e- 3   -6.79   -1.03
#> # … with abbreviated variable names ¹​estimate, ²​std.error, ³​statistic,
#> #   ⁴​conf.low, ⁵​conf.high

# GLM with different scale types
mod <- glm(am ~ factor(gear), data = mtcars)
comparisons(mod, type = "response") |> tidy()
#> # A tibble: 2 × 9
#>   type     term  contrast       estim…¹ std.e…² stati…³  p.value conf.…⁴ conf.…⁵
#>   <chr>    <chr> <chr>            <dbl>   <dbl>   <dbl>    <dbl>   <dbl>   <dbl>
#> 1 response gear  mean(4) - mea…   0.667   0.117    5.68 1.38e- 8   0.436   0.897
#> 2 response gear  mean(5) - mea…   1       0.157    6.39 1.70e-10   0.693   1.31 
#> # … with abbreviated variable names ¹​estimate, ²​std.error, ³​statistic,
#> #   ⁴​conf.low, ⁵​conf.high
comparisons(mod, type = "link") |> tidy()
#> # A tibble: 2 × 9
#>   type  term  contrast          estim…¹ std.e…² stati…³  p.value conf.…⁴ conf.…⁵
#>   <chr> <chr> <chr>               <dbl>   <dbl>   <dbl>    <dbl>   <dbl>   <dbl>
#> 1 link  gear  mean(4) - mean(3)   0.667   0.117    5.68 1.38e- 8   0.436   0.897
#> 2 link  gear  mean(5) - mean(3)   1       0.157    6.39 1.70e-10   0.693   1.31 
#> # … with abbreviated variable names ¹​estimate, ²​std.error, ³​statistic,
#> #   ⁴​conf.low, ⁵​conf.high

# Contrasts at the mean
comparisons(mod, newdata = "mean")
#> 
#>  Term Contrast Estimate Std. Error     z   Pr(>|z|)  2.5 % 97.5 %
#>  gear    4 - 3   0.6667     0.1174 5.676 1.3751e-08 0.4365 0.8969
#>  gear    5 - 3   1.0000     0.1566 6.386 1.7026e-10 0.6931 1.3069
#> 
#> Prediction type:  response 
#> Columns: rowid, type, term, contrast, estimate, std.error, statistic, p.value, conf.low, conf.high, predicted, predicted_hi, predicted_lo, am, gear 
#> 

# Contrasts between marginal means
comparisons(mod, newdata = "marginalmeans")
#> 
#>  Term Contrast Estimate Std. Error     z   Pr(>|z|)  2.5 % 97.5 %
#>  gear    4 - 3   0.6667     0.1174 5.676 1.3751e-08 0.4365 0.8969
#>  gear    5 - 3   1.0000     0.1566 6.386 1.7026e-10 0.6931 1.3069
#> 
#> Prediction type:  response 
#> Columns: type, term, contrast, estimate, std.error, statistic, p.value, conf.low, conf.high, predicted, predicted_hi, predicted_lo 
#> 

# Contrasts at user-specified values
comparisons(mod, newdata = datagrid(am = 0, gear = tmp$gear))
#> 
#>  Term Contrast Estimate Std. Error     z   Pr(>|z|)  2.5 % 97.5 %
#>  gear    4 - 3   0.6667     0.1174 5.676 1.3751e-08 0.4365 0.8969
#>  gear    4 - 3   0.6667     0.1174 5.676 1.3751e-08 0.4365 0.8969
#>  gear    4 - 3   0.6667     0.1174 5.676 1.3751e-08 0.4365 0.8969
#>  gear    5 - 3   1.0000     0.1566 6.386 1.7026e-10 0.6931 1.3069
#>  gear    5 - 3   1.0000     0.1566 6.386 1.7026e-10 0.6931 1.3069
#>  gear    5 - 3   1.0000     0.1566 6.386 1.7026e-10 0.6931 1.3069
#> 
#> Prediction type:  response 
#> Columns: rowid, type, term, contrast, estimate, std.error, statistic, p.value, conf.low, conf.high, predicted, predicted_hi, predicted_lo, am, gear 
#> 
comparisons(mod, newdata = datagrid(am = unique, gear = max))
#> 
#>  Term Contrast Estimate Std. Error     z   Pr(>|z|)  2.5 % 97.5 %
#>  gear    4 - 3   0.6667     0.1174 5.676 1.3751e-08 0.4365 0.8969
#>  gear    4 - 3   0.6667     0.1174 5.676 1.3751e-08 0.4365 0.8969
#>  gear    5 - 3   1.0000     0.1566 6.386 1.7026e-10 0.6931 1.3069
#>  gear    5 - 3   1.0000     0.1566 6.386 1.7026e-10 0.6931 1.3069
#> 
#> Prediction type:  response 
#> Columns: rowid, type, term, contrast, estimate, std.error, statistic, p.value, conf.low, conf.high, predicted, predicted_hi, predicted_lo, am, gear 
#> 

m <- lm(mpg ~ hp + drat + factor(cyl) + factor(am), data = mtcars)
comparisons(m, variables = "hp", newdata = datagrid(FUN_factor = unique, FUN_numeric = median))
#> 
#>  Term Contrast Estimate Std. Error      z  Pr(>|z|)    2.5 %   97.5 %  drat cyl
#>    hp       +1 -0.04523    0.01489 -3.037 0.0023856 -0.07441 -0.01604 3.695   6
#>    hp       +1 -0.04523    0.01489 -3.037 0.0023856 -0.07441 -0.01604 3.695   6
#>    hp       +1 -0.04523    0.01489 -3.037 0.0023856 -0.07441 -0.01604 3.695   4
#>    hp       +1 -0.04523    0.01489 -3.037 0.0023856 -0.07441 -0.01604 3.695   4
#>    hp       +1 -0.04523    0.01489 -3.037 0.0023856 -0.07441 -0.01604 3.695   8
#>    hp       +1 -0.04523    0.01489 -3.037 0.0023856 -0.07441 -0.01604 3.695   8
#>  am
#>   1
#>   0
#>   1
#>   0
#>   1
#>   0
#> 
#> Prediction type:  response 
#> Columns: rowid, type, term, contrast, estimate, std.error, statistic, p.value, conf.low, conf.high, predicted, predicted_hi, predicted_lo, mpg, hp, drat, cyl, am, eps 
#> 

# Numeric contrasts
mod <- lm(mpg ~ hp, data = mtcars)
comparisons(mod, variables = list(hp = 1)) |> tidy()
#> # A tibble: 1 × 9
#>   type     term  contrast estimate std.error statistic  p.value conf.low conf.…¹
#>   <chr>    <chr> <chr>       <dbl>     <dbl>     <dbl>    <dbl>    <dbl>   <dbl>
#> 1 response hp    mean(+1)  -0.0682    0.0101     -6.74 1.56e-11  -0.0881 -0.0484
#> # … with abbreviated variable name ¹​conf.high
comparisons(mod, variables = list(hp = 5)) |> tidy()
#> # A tibble: 1 × 9
#>   type     term  contrast estimate std.error statistic  p.value conf.low conf.…¹
#>   <chr>    <chr> <chr>       <dbl>     <dbl>     <dbl>    <dbl>    <dbl>   <dbl>
#> 1 response hp    mean(+5)   -0.341    0.0506     -6.74 1.56e-11   -0.440  -0.242
#> # … with abbreviated variable name ¹​conf.high
comparisons(mod, variables = list(hp = c(90, 100))) |> tidy()
#> # A tibble: 1 × 9
#>   type     term  contrast       estim…¹ std.e…² stati…³  p.value conf.…⁴ conf.…⁵
#>   <chr>    <chr> <chr>            <dbl>   <dbl>   <dbl>    <dbl>   <dbl>   <dbl>
#> 1 response hp    mean(100) - m…  -0.682   0.101   -6.74 1.56e-11  -0.881  -0.484
#> # … with abbreviated variable names ¹​estimate, ²​std.error, ³​statistic,
#> #   ⁴​conf.low, ⁵​conf.high
comparisons(mod, variables = list(hp = "iqr")) |> tidy()
#> # A tibble: 1 × 9
#>   type     term  contrast       estim…¹ std.e…² stati…³  p.value conf.…⁴ conf.…⁵
#>   <chr>    <chr> <chr>            <dbl>   <dbl>   <dbl>    <dbl>   <dbl>   <dbl>
#> 1 response hp    mean(Q3) - me…   -5.70   0.845   -6.74 1.56e-11   -7.35   -4.04
#> # … with abbreviated variable names ¹​estimate, ²​std.error, ³​statistic,
#> #   ⁴​conf.low, ⁵​conf.high
comparisons(mod, variables = list(hp = "sd")) |> tidy()
#> # A tibble: 1 × 9
#>   type     term  contrast       estim…¹ std.e…² stati…³  p.value conf.…⁴ conf.…⁵
#>   <chr>    <chr> <chr>            <dbl>   <dbl>   <dbl>    <dbl>   <dbl>   <dbl>
#> 1 response hp    mean(x + sd/2…   -4.68   0.694   -6.74 1.56e-11   -6.04   -3.32
#> # … with abbreviated variable names ¹​estimate, ²​std.error, ³​statistic,
#> #   ⁴​conf.low, ⁵​conf.high
comparisons(mod, variables = list(hp = "minmax")) |> tidy()
#> # A tibble: 1 × 9
#>   type     term  contrast       estim…¹ std.e…² stati…³  p.value conf.…⁴ conf.…⁵
#>   <chr>    <chr> <chr>            <dbl>   <dbl>   <dbl>    <dbl>   <dbl>   <dbl>
#> 1 response hp    mean(Max) - m…   -19.3    2.86   -6.74 1.56e-11   -24.9   -13.7
#> # … with abbreviated variable names ¹​estimate, ²​std.error, ³​statistic,
#> #   ⁴​conf.low, ⁵​conf.high

# using a function to specify a custom difference in one regressor
dat <- mtcars
dat$new_hp <- 49 * (dat$hp - min(dat$hp)) / (max(dat$hp) - min(dat$hp)) + 1
modlog <- lm(mpg ~ log(new_hp) + factor(cyl), data = dat)
fdiff <- \(x) data.frame(x, x + 10)
avg_comparisons(modlog, variables = list(new_hp = fdiff))
#> 
#>    Term Contrast Estimate Std. Error      z  Pr(>|z|)  2.5 %  97.5 %
#>  new_hp   custom   -1.974     0.7105 -2.778 0.0054696 -3.366 -0.5812
#> 
#> Prediction type:  response 
#> Columns: type, term, contrast, estimate, std.error, statistic, p.value, conf.low, conf.high 
#> 

# Adjusted Risk Ratio: see the contrasts vignette
mod <- glm(vs ~ mpg, data = mtcars, family = binomial)
avg_comparisons(mod, transform_pre = "lnratioavg", transform_post = exp)
#> 
#>  Term Contrast Estimate   Pr(>|z|) 2.5 % 97.5 %
#>   mpg mean(+1)    1.135 2.3808e-10 1.091   1.18
#> 
#> Prediction type:  response 
#> Columns: type, term, contrast, estimate, p.value, conf.low, conf.high, predicted, predicted_hi, predicted_lo 
#> 

# Adjusted Risk Ratio: Manual specification of the `transform_pre`
avg_comparisons(
     mod,
     transform_pre = function(hi, lo) log(mean(hi) / mean(lo)),
     transform_post = exp)
#> 
#>  Term Contrast Estimate   Pr(>|z|) 2.5 % 97.5 %
#>   mpg       +1    1.135 2.3808e-10 1.091   1.18
#> 
#> Prediction type:  response 
#> Columns: type, term, contrast, estimate, p.value, conf.low, conf.high, predicted, predicted_hi, predicted_lo 
#> 
# cross contrasts
mod <- lm(mpg ~ factor(cyl) * factor(gear) + hp, data = mtcars)
avg_comparisons(mod, variables = c("cyl", "gear"), cross = TRUE)
#> Warning: Model matrix is rank deficient. Some variance-covariance parameters are
#>   missing.
#> 
#>  C: cyl C: gear Estimate Std. Error       z Pr(>|z|)  2.5 % 97.5 %
#>   6 - 4   4 - 3  -0.6306      3.405 -0.1852  0.85307 -7.303  6.042
#>   6 - 4   5 - 3   2.6778      4.617  0.5800  0.56193 -6.372 11.727
#>   8 - 4   4 - 3   3.3475      6.427  0.5208  0.60249 -9.250 15.945
#>   8 - 4   5 - 3   5.5250      5.868  0.9416  0.34642 -5.976 17.026
#> 
#> Prediction type:  response 
#> Columns: type, term, contrast_cyl, contrast_gear, estimate, std.error, statistic, p.value, conf.low, conf.high 
#> 

# variable-specific contrasts
avg_comparisons(mod, variables = list(gear = "sequential", hp = 10))
#> Warning: Model matrix is rank deficient. Some variance-covariance parameters are
#>   missing.
#> 
#>  Term Contrast Estimate Std. Error       z Pr(>|z|)  2.5 %  97.5 %
#>  gear    4 - 3   3.4095      2.587  1.3177 0.187591 -1.662  8.4806
#>  gear    5 - 4   2.6277      2.747  0.9566 0.338744 -2.756  8.0113
#>    hp      +10  -0.5741      0.225 -2.5518 0.010718 -1.015 -0.1331
#> 
#> Prediction type:  response 
#> Columns: type, term, contrast, estimate, std.error, statistic, p.value, conf.low, conf.high 
#> 

# hypothesis test: is the `hp` marginal effect at the mean equal to the `drat` marginal effect
mod <- lm(mpg ~ wt + drat, data = mtcars)

comparisons(
    mod,
    newdata = "mean",
    hypothesis = "wt = drat")
#> 
#>     Term Estimate Std. Error      z   Pr(>|z|)  2.5 % 97.5 %
#>  wt=drat   -6.225      1.052 -5.919 3.2398e-09 -8.287 -4.164
#> 
#> Prediction type:  response 
#> Columns: type, term, estimate, std.error, statistic, p.value, conf.low, conf.high 
#> 

# same hypothesis test using row indices
comparisons(
    mod,
    newdata = "mean",
    hypothesis = "b1 - b2 = 0")
#> 
#>     Term Estimate Std. Error      z   Pr(>|z|)  2.5 % 97.5 %
#>  b1-b2=0   -6.225      1.052 -5.919 3.2398e-09 -8.287 -4.164
#> 
#> Prediction type:  response 
#> Columns: type, term, estimate, std.error, statistic, p.value, conf.low, conf.high 
#> 

# same hypothesis test using numeric vector of weights
comparisons(
    mod,
    newdata = "mean",
    hypothesis = c(1, -1))
#> 
#>    Term Estimate Std. Error      z   Pr(>|z|)  2.5 % 97.5 %
#>  custom   -6.225      1.052 -5.919 3.2398e-09 -8.287 -4.164
#> 
#> Prediction type:  response 
#> Columns: type, term, estimate, std.error, statistic, p.value, conf.low, conf.high 
#> 

# two custom contrasts using a matrix of weights
lc <- matrix(c(
    1, -1,
    2, 3),
    ncol = 2)
comparisons(
    mod,
    newdata = "mean",
    hypothesis = lc)
#> 
#>    Term Estimate Std. Error       z   Pr(>|z|)   2.5 % 97.5 %
#>  custom   -6.225      1.052 -5.9190 3.2398e-09  -8.287 -4.164
#>  custom   -5.238      5.624 -0.9315    0.35162 -16.261  5.784
#> 
#> Prediction type:  response 
#> Columns: type, term, estimate, std.error, statistic, p.value, conf.low, conf.high 
#> 

# `by` argument
mod <- lm(mpg ~ hp * am * vs, data = mtcars)
comparisons(mod, by = TRUE)
#> 
#>  Term Contrast Estimate Std. Error       z   Pr(>|z|)   2.5 %   97.5 %
#>    am    1 - 0  4.69801     1.0601  4.4316 9.3531e-06  2.6202  6.77580
#>    hp       +1 -0.06878     0.0182 -3.7797 0.00015703 -0.1044 -0.03311
#>    vs    1 - 0 -0.29433     2.3379 -0.1259 0.89981407 -4.8766  4.28789
#> 
#> Prediction type:  response 
#> Columns: type, term, contrast, estimate, std.error, statistic, p.value, conf.low, conf.high 
#> 

mod <- lm(mpg ~ hp * am * vs, data = mtcars)
avg_comparisons(mod, variables = "hp", by = c("vs", "am"))
#> 
#>  Term Contrast vs am Estimate Std. Error      z  Pr(>|z|)    2.5 %    97.5 %
#>    hp mean(+1)  0  0 -0.04215    0.02477 -1.702 0.0887904 -0.09070  0.006394
#>    hp mean(+1)  0  1 -0.03685    0.01240 -2.971 0.0029676 -0.06116 -0.012540
#>    hp mean(+1)  1  0 -0.09941    0.05345 -1.860 0.0628876 -0.20417  0.005344
#>    hp mean(+1)  1  1 -0.11115    0.04634 -2.399 0.0164473 -0.20197 -0.020336
#> 
#> Prediction type:  response 
#> Columns: type, term, contrast, vs, am, estimate, std.error, statistic, p.value, conf.low, conf.high, predicted, predicted_hi, predicted_lo 
#> 

library(nnet)
mod <- multinom(factor(gear) ~ mpg + am * vs, data = mtcars, trace = FALSE)
by <- data.frame(
    group = c("3", "4", "5"),
    by = c("3,4", "3,4", "5"))
comparisons(mod, type = "probs", by = by)
#> 
#>  Term  By   Estimate Std. Error        z  Pr(>|z|)    2.5 %   97.5 %
#>    am 3,4 -0.2227925   0.079558 -2.80038 0.0051042 -0.37872 -0.06686
#>    am   5  0.4455851   0.159116  2.80038 0.0051042  0.13372  0.75745
#>   mpg 3,4  0.0004635   0.005796  0.07997 0.9362637 -0.01090  0.01182
#>   mpg   5 -0.0009270   0.011592 -0.07997 0.9362637 -0.02365  0.02179
#>    vs 3,4  0.1021021   0.073236  1.39416 0.1632708 -0.04144  0.24564
#>    vs   5 -0.2042041   0.146472 -1.39416 0.1632708 -0.49128  0.08287
#> 
#> Prediction type:  probs 
#> Columns: type, term, by, estimate, std.error, statistic, p.value, conf.low, conf.high 
#>