Experimental function to compute contrasts between adjusted predictions

This function calculates contrasts (or comparisons) between adjusted predictions for each row of the dataset. The resulting object can processed by the tidy() or summary() functions, which compute Average Contrasts. The datagrid() function and the newdata argument can be used to calculate contrasts Contrasts at the Mean or Contrasts at User-Specified values (aka Contrasts at Representative values). Additional information can be found in the Details and Examples sections below, and in the vignette on the marginaleffects website.

comparisons(
  model,
  variables = NULL,
  newdata = insight::get_data(model),
  type = "response",
  vcov = TRUE,
  contrast_factor = "reference",
  contrast_numeric = 1,
  ...
)

Arguments

model

Model object

variables

Variables to consider (character vector). NULL calculates marginal effects for all terms in the model object.

newdata

A dataset over which to compute marginal effects. NULL uses the original data used to fit the model.

type

Type(s) of prediction as string or character vector. This can differ based on the model type, but will typically be a string such as: "response", "link", "probs", or "zero".

vcov

Matrix or boolean

FALSE: does not compute unit-level standard errors.

TRUE: computes unit-level standard errors using the default vcov(model) variance-covariance matrix.

Named square matrix: computes standard errors with a user-supplied variance-covariance matrix. This matrix must be square and have dimensions equal to the number of coefficients in get_coef(model).

contrast_factor

"reference" or "sequential"

"reference": Each factor level is compared to the factor reference (base) level

"sequential": Each factor level is compared to the previous factor level

"pairwise": Each factor level is compared to all other levels

contrast_numeric

string or numeric

Numeric of length 1: Contrast between the observed value and the observed value plus contrast_numeric

Numeric vector of length 2: Contrast between the 2nd element and the 1st element of the contrast_numeric vector.

"iqr": Contrast across the interquartile range of the regressor.

"sd": Contrast across one standard deviation around the regressor mean.

"2sd": Contrast across two standard deviations around the regressor mean.

"minmax": Contrast between the maximum and the minimum values of the regressor.

...

Additional arguments are passed to the predict() method used to compute adjusted predictions. These arguments are particularly useful for mixed-effects or bayesian models (see the online vignettes on the marginaleffects website). Available arguments can vary from model to model, depending on the range of supported arguments by each modeling package. See the "Model-Specific Arguments" section of the ?marginaleffects document for a non-exhaustive list of available arguments.

Details

A "contrast" is the difference between two adjusted predictions, calculated for meaningfully different regressor values (e.g., College graduates vs. Others). Uncertainty estimates are computed using the delta method.

Detailed vignettes on contrasts, marginal effects, predictions, and marginal means, as well as a list of supported models can be found on the package website:

https://vincentarelbundock.github.io/marginaleffects/

Examples

library(marginaleffects)
library(magrittr)

# Linear model
tmp <- mtcars
tmp$am <- as.logical(tmp$am)
mod <- lm(mpg ~ am + factor(cyl), tmp)
comparisons(mod, contrast_factor = "reference") %>% tidy()
#>   term     contrast   estimate std.error statistic      p.value     conf.low
#> 1   am TRUE - FALSE   2.559954  1.297579  1.972869 4.851044e-02   0.01674586
#> 2  cyl        6 - 4  -6.156118  1.535723 -4.008612 6.107658e-05  -9.16607927
#> 3  cyl        8 - 4 -10.067560  1.452082 -6.933187 4.114709e-12 -12.91358877
#>   conf.high
#> 1  5.103162
#> 2 -3.146156
#> 3 -7.221530
comparisons(mod, contrast_factor = "sequential") %>% tidy()
#>   term     contrast  estimate std.error statistic      p.value    conf.low
#> 1   am TRUE - FALSE  2.559954  1.297579  1.972869 4.851044e-02  0.01674586
#> 2  cyl        6 - 4 -6.156118  1.535723 -4.008612 6.107658e-05 -9.16607927
#> 3  cyl        8 - 6 -3.911442  1.470254 -2.660385 7.805144e-03 -6.79308707
#>   conf.high
#> 1  5.103162
#> 2 -3.146156
#> 3 -1.029797
comparisons(mod, contrast_factor = "pairwise") %>% tidy()
#>   term     contrast   estimate std.error statistic      p.value     conf.low
#> 1   am TRUE - FALSE   2.559954  1.297579  1.972869 4.851044e-02   0.01674586
#> 2  cyl        6 - 4  -6.156118  1.535723 -4.008612 6.107658e-05  -9.16607927
#> 3  cyl        8 - 4 -10.067560  1.452082 -6.933187 4.114709e-12 -12.91358877
#> 4  cyl        8 - 6  -3.911442  1.470254 -2.660385 7.805144e-03  -6.79308707
#>   conf.high
#> 1  5.103162
#> 2 -3.146156
#> 3 -7.221530
#> 4 -1.029797

# GLM with different scale types
mod <- glm(am ~ factor(gear), data = mtcars)
comparisons(mod) %>% tidy()
#>   term contrast  estimate std.error statistic      p.value  conf.low conf.high
#> 1 gear    4 - 3 0.6666667 0.1174440  5.676462 1.375090e-08 0.4364806 0.8968528
#> 2 gear    5 - 3 1.0000000 0.1565921  6.386020 1.702589e-10 0.6930852 1.3069148
comparisons(mod, type = "link") %>% tidy()
#>   term contrast  estimate std.error statistic      p.value  conf.low conf.high
#> 1 gear    4 - 3 0.6666667 0.1174440  5.676462 1.375090e-08 0.4364806 0.8968528
#> 2 gear    5 - 3 1.0000000 0.1565921  6.386020 1.702589e-10 0.6930852 1.3069148

# Numeric contrasts
mod <- lm(mpg ~ hp, data = mtcars)
comparisons(mod, contrast_numeric = 1) %>% tidy()
#>   term contrast    estimate std.error statistic      p.value    conf.low
#> 1   hp       +1 -0.06822828 0.0101193 -6.742389 1.558043e-11 -0.08806175
#>     conf.high
#> 1 -0.04839481
comparisons(mod, contrast_numeric = 5) %>% tidy()
#>   term contrast   estimate  std.error statistic      p.value   conf.low
#> 1   hp       +5 -0.3411414 0.05059652 -6.742389 1.558043e-11 -0.4403087
#>   conf.high
#> 1 -0.241974
comparisons(mod, contrast_numeric = c(90, 100)) %>% tidy()
#>   term contrast   estimate std.error statistic      p.value   conf.low
#> 1   hp 100 - 90 -0.6822828  0.101193 -6.742389 1.558043e-11 -0.8806175
#>    conf.high
#> 1 -0.4839481
comparisons(mod, contrast_numeric = "iqr") %>% tidy()
#>   term contrast  estimate std.error statistic      p.value  conf.low conf.high
#> 1   hp      IQR -5.697061 0.8449619 -6.742389 1.558043e-11 -7.353156 -4.040966
comparisons(mod, contrast_numeric = "sd") %>% tidy()
#>   term contrast  estimate std.error statistic      p.value  conf.low conf.high
#> 1   hp       sd -4.677926 0.6938085 -6.742389 1.558043e-11 -6.037766 -3.318087
comparisons(mod, contrast_numeric = "minmax") %>% tidy()
#>   term  contrast estimate std.error statistic      p.value  conf.low conf.high
#> 1   hp Max - Min -19.3086  2.863763 -6.742389 1.558043e-11 -24.92147 -13.69573