Generate "counterfactual" datasets for use in marginaleffects's newdata argument

Generate "counterfactual" datasets for use in marginaleffects's newdata argument

counterfactual(..., model = NULL, data = NULL)

Arguments

...
named arguments with vectors of values for the variables to construct (see Examples below.)

model
Model object

data
data.frame (one and only one of the model and data arguments must be true).

Value

A data.frame where each row of the original data is repeated multiple times for each of the values of the variables in the at list. See example below.

If users supply a model, the data used to fit that model is retrieved using the insight::get_data function, and then replicated with different values of the variables in the at list.

Examples

# All rows are repeated twice, with different values of `hp`
cd <- counterfactual(data = mtcars, hp = c(100, 110))
cd[cd$rowid %in% 1:3,]

#>    rowid  mpg cyl disp drat    wt  qsec vs am gear carb  hp
#> 1      1 21.0   6  160 3.90 2.620 16.46  0  1    4    4 100
#> 2      2 21.0   6  160 3.90 2.875 17.02  0  1    4    4 100
#> 3      3 22.8   4  108 3.85 2.320 18.61  1  1    4    1 100
#> 33     1 21.0   6  160 3.90 2.620 16.46  0  1    4    4 110
#> 34     2 21.0   6  160 3.90 2.875 17.02  0  1    4    4 110
#> 35     3 22.8   4  108 3.85 2.320 18.61  1  1    4    1 110

# We get the same result by feeding a model instead of a data.frame
mod <- lm(mpg ~ hp + wt, mtcars)
cd <- counterfactual(model = mod, hp = c(100, 110))
cd[cd$rowid %in% 1:3,]

#>    rowid    wt  hp
#> 1      1 2.620 100
#> 2      2 2.875 100
#> 3      3 2.320 100
#> 33     1 2.620 110
#> 34     2 2.875 110
#> 35     3 2.320 110

# Use in `marginaleffects` to compute "Counterfactual Average Marginal Effects"
marginaleffects(mod, newdata = counterfactual(hp = c(100, 110)))

#>     rowid     type term        dydx   std.error predicted_response    wt  hp
#> 1       1 response   hp -0.03177295 0.009029704           23.89006 2.620 100
#> 2       1 response   hp -0.03177295 0.009029704           23.57233 2.620 110
#> 3       1 response   wt -3.87783074 0.632733547           23.89006 2.620 100
#> 4       1 response   wt -3.87783074 0.632733547           23.57233 2.620 110
#> 5       2 response   hp -0.03177295 0.009029704           22.90121 2.875 100
#> 6       2 response   hp -0.03177295 0.009029704           22.58348 2.875 110
#> 7       2 response   wt -3.87783074 0.632733322           22.90121 2.875 100
#> 8       2 response   wt -3.87783074 0.632733322           22.58348 2.875 110
#> 9       3 response   wt -3.87783074 0.632733097           24.73568 2.320 110
#> 10      3 response   wt -3.87783074 0.632733097           25.05341 2.320 100
#> 11      3 response   hp -0.03177295 0.009029704           24.73568 2.320 110
#> 12      3 response   hp -0.03177295 0.009029704           25.05341 2.320 100
#> 13      4 response   wt -3.87783074 0.632733322           21.26502 3.215 110
#> 14      4 response   wt -3.87783074 0.632733322           21.58275 3.215 100
#> 15      4 response   hp -0.03177295 0.009029704           21.26502 3.215 110
#> 16      4 response   hp -0.03177295 0.009029704           21.58275 3.215 100
#> 17      5 response   hp -0.03177295 0.009029704           20.71024 3.440 100
#> 18      5 response   hp -0.03177295 0.009029704           20.39251 3.440 110
#> 19      5 response   wt -3.87783074 0.632733322           20.71024 3.440 100
#> 20      5 response   wt -3.87783074 0.632733322           20.39251 3.440 110
#> 21      6 response   wt -3.87783074 0.632733547           20.63268 3.460 100
#> 22      6 response   wt -3.87783074 0.632733547           20.31495 3.460 110
#> 23      6 response   hp -0.03177295 0.009029704           20.63268 3.460 100
#> 24      6 response   hp -0.03177295 0.009029704           20.31495 3.460 110
#> 25      7 response   wt -3.87783074 0.632733322           19.88839 3.570 110
#> 26      7 response   wt -3.87783074 0.632733322           20.20612 3.570 100
#> 27      7 response   hp -0.03177295 0.009029704           19.88839 3.570 110
#> 28      7 response   hp -0.03177295 0.009029704           20.20612 3.570 100
#> 29      8 response   wt -3.87783074 0.632733322           21.36197 3.190 110
#> 30      8 response   wt -3.87783074 0.632733322           21.67970 3.190 100
#> 31      8 response   hp -0.03177295 0.009029704           21.36197 3.190 110
#> 32      8 response   hp -0.03177295 0.009029704           21.67970 3.190 100
#> 33      9 response   hp -0.03177295 0.009029704           21.51708 3.150 110
#> 34      9 response   hp -0.03177295 0.009029704           21.83481 3.150 100
#> 35      9 response   wt -3.87783074 0.632733547           21.51708 3.150 110
#> 36      9 response   wt -3.87783074 0.632733547           21.83481 3.150 100
#> 37     10 response   wt -3.87783074 0.632733322           20.71024 3.440 100
#> 38     10 response   wt -3.87783074 0.632733322           20.39251 3.440 110
#> 39     10 response   hp -0.03177295 0.009029704           20.71024 3.440 100
#> 40     10 response   hp -0.03177295 0.009029704           20.39251 3.440 110
#> 41     11 response   hp -0.03177295 0.009029704           20.39251 3.440 110
#> 42     11 response   hp -0.03177295 0.009029704           20.71024 3.440 100
#> 43     11 response   wt -3.87783074 0.632733322           20.39251 3.440 110
#> 44     11 response   wt -3.87783074 0.632733322           20.71024 3.440 100
#> 45     12 response   hp -0.03177295 0.009029704           17.94947 4.070 110
#> 46     12 response   hp -0.03177295 0.009029704           18.26720 4.070 100
#> 47     12 response   wt -3.87783074 0.632733547           17.94947 4.070 110
#> 48     12 response   wt -3.87783074 0.632733547           18.26720 4.070 100
#> 49     13 response   hp -0.03177295 0.009029704           19.26794 3.730 110
#> 50     13 response   hp -0.03177295 0.009029704           19.58567 3.730 100
#> 51     13 response   wt -3.87783074 0.632733547           19.26794 3.730 110
#> 52     13 response   wt -3.87783074 0.632733547           19.58567 3.730 100
#> 53     14 response   hp -0.03177295 0.009029704           19.39178 3.780 100
#> 54     14 response   hp -0.03177295 0.009029704           19.07405 3.780 110
#> 55     14 response   wt -3.87783074 0.632733322           19.39178 3.780 100
#> 56     14 response   wt -3.87783074 0.632733322           19.07405 3.780 110
#> 57     15 response   hp -0.03177295 0.009029704           13.69136 5.250 100
#> 58     15 response   hp -0.03177295 0.009029704           13.37363 5.250 110
#> 59     15 response   wt -3.87783074 0.632733547           13.69136 5.250 100
#> 60     15 response   wt -3.87783074 0.632733547           13.37363 5.250 110
#> 61     16 response   hp -0.03177295 0.009029704           12.69889 5.424 110
#> 62     16 response   hp -0.03177295 0.009029704           13.01662 5.424 100
#> 63     16 response   wt -3.87783074 0.632733322           12.69889 5.424 110
#> 64     16 response   wt -3.87783074 0.632733322           13.01662 5.424 100
#> 65     17 response   wt -3.87783074 0.632733547           13.00524 5.345 110
#> 66     17 response   wt -3.87783074 0.632733547           13.32297 5.345 100
#> 67     17 response   hp -0.03177295 0.009029704           13.00524 5.345 110
#> 68     17 response   hp -0.03177295 0.009029704           13.32297 5.345 100
#> 69     18 response   hp -0.03177295 0.009029704           25.51875 2.200 100
#> 70     18 response   hp -0.03177295 0.009029704           25.20102 2.200 110
#> 71     18 response   wt -3.87783074 0.632733322           25.51875 2.200 100
#> 72     18 response   wt -3.87783074 0.632733322           25.20102 2.200 110
#> 73     19 response   hp -0.03177295 0.009029704           27.78728 1.615 100
#> 74     19 response   hp -0.03177295 0.009029704           27.46955 1.615 110
#> 75     19 response   wt -3.87783074 0.632733772           27.78728 1.615 100
#> 76     19 response   wt -3.87783074 0.632733772           27.46955 1.615 110
#> 77     20 response   wt -3.87783074 0.632733322           26.61643 1.835 110
#> 78     20 response   wt -3.87783074 0.632733322           26.93416 1.835 100
#> 79     20 response   hp -0.03177295 0.009029704           26.61643 1.835 110
#> 80     20 response   hp -0.03177295 0.009029704           26.93416 1.835 100
#> 81     21 response   wt -3.87783074 0.632733097           24.17339 2.465 110
#> 82     21 response   wt -3.87783074 0.632733097           24.49112 2.465 100
#> 83     21 response   hp -0.03177295 0.009029704           24.17339 2.465 110
#> 84     21 response   hp -0.03177295 0.009029704           24.49112 2.465 100
#> 85     22 response   hp -0.03177295 0.009029704           20.08228 3.520 110
#> 86     22 response   hp -0.03177295 0.009029704           20.40001 3.520 100
#> 87     22 response   wt -3.87783074 0.632733322           20.08228 3.520 110
#> 88     22 response   wt -3.87783074 0.632733322           20.40001 3.520 100
#> 89     23 response   hp -0.03177295 0.009029704           20.72963 3.435 100
#> 90     23 response   hp -0.03177295 0.009029704           20.41190 3.435 110
#> 91     23 response   wt -3.87783074 0.632733547           20.72963 3.435 100
#> 92     23 response   wt -3.87783074 0.632733547           20.41190 3.435 110
#> 93     24 response   hp -0.03177295 0.009029704           18.84138 3.840 110
#> 94     24 response   hp -0.03177295 0.009029704           19.15911 3.840 100
#> 95     24 response   wt -3.87783074 0.632733996           18.84138 3.840 110
#> 96     24 response   wt -3.87783074 0.632733996           19.15911 3.840 100
#> 97     25 response   hp -0.03177295 0.009029704           18.82199 3.845 110
#> 98     25 response   hp -0.03177295 0.009029704           19.13972 3.845 100
#> 99     25 response   wt -3.87783074 0.632733322           18.82199 3.845 110
#> 100    25 response   wt -3.87783074 0.632733322           19.13972 3.845 100
#> 101    26 response   wt -3.87783074 0.632733322           26.22864 1.935 110
#> 102    26 response   wt -3.87783074 0.632733322           26.54637 1.935 100
#> 103    26 response   hp -0.03177295 0.009029704           26.22864 1.935 110
#> 104    26 response   hp -0.03177295 0.009029704           26.54637 1.935 100
#> 105    27 response   wt -3.87783074 0.632733547           25.75142 2.140 100
#> 106    27 response   wt -3.87783074 0.632733547           25.43369 2.140 110
#> 107    27 response   hp -0.03177295 0.009029704           25.75142 2.140 100
#> 108    27 response   hp -0.03177295 0.009029704           25.43369 2.140 110
#> 109    28 response   hp -0.03177295 0.009029704           28.18282 1.513 100
#> 110    28 response   hp -0.03177295 0.009029704           27.86509 1.513 110
#> 111    28 response   wt -3.87783074 0.632733322           28.18282 1.513 100
#> 112    28 response   wt -3.87783074 0.632733322           27.86509 1.513 110
#> 113    29 response   hp -0.03177295 0.009029704           21.43952 3.170 110
#> 114    29 response   hp -0.03177295 0.009029704           21.75725 3.170 100
#> 115    29 response   wt -3.87783074 0.632733547           21.43952 3.170 110
#> 116    29 response   wt -3.87783074 0.632733547           21.75725 3.170 100
#> 117    30 response   wt -3.87783074 0.632733547           22.99065 2.770 110
#> 118    30 response   wt -3.87783074 0.632733547           23.30838 2.770 100
#> 119    30 response   hp -0.03177295 0.009029704           22.99065 2.770 110
#> 120    30 response   hp -0.03177295 0.009029704           23.30838 2.770 100
#> 121    31 response   wt -3.87783074 0.632733322           20.20612 3.570 100
#> 122    31 response   wt -3.87783074 0.632733322           19.88839 3.570 110
#> 123    31 response   hp -0.03177295 0.009029704           20.20612 3.570 100
#> 124    31 response   hp -0.03177295 0.009029704           19.88839 3.570 110
#> 125    32 response   hp -0.03177295 0.009029704           23.26961 2.780 100
#> 126    32 response   hp -0.03177295 0.009029704           22.95188 2.780 110
#> 127    32 response   wt -3.87783074 0.632733322           23.26961 2.780 100
#> 128    32 response   wt -3.87783074 0.632733322           22.95188 2.780 110