Generate "counterfactual" datasets for use in marginaleffects's newdata argument

Generate "counterfactual" datasets for use in marginaleffects's newdata argument

counterfactual(..., model = NULL, data = NULL)

Arguments

...
named arguments with vectors of values for the variables to construct (see Examples below.)

model
Model object

data
data.frame (one and only one of the model and data arguments must be true).

Value

A data.frame where each row of the original data is repeated multiple times for each of the values of the variables in the at list. See example below.

If users supply a model, the data used to fit that model is retrieved using the insight::get_data function, and then replicated with different values of the variables in the at list.

Examples

# All rows are repeated twice, with different values of `hp`
cd <- counterfactual(data = mtcars, hp = c(100, 110))
cd[cd$rowid %in% 1:3,]

#>    rowid  mpg cyl disp drat    wt  qsec vs am gear carb  hp
#> 1      1 21.0   6  160 3.90 2.620 16.46  0  1    4    4 100
#> 2      2 21.0   6  160 3.90 2.875 17.02  0  1    4    4 100
#> 3      3 22.8   4  108 3.85 2.320 18.61  1  1    4    1 100
#> 33     1 21.0   6  160 3.90 2.620 16.46  0  1    4    4 110
#> 34     2 21.0   6  160 3.90 2.875 17.02  0  1    4    4 110
#> 35     3 22.8   4  108 3.85 2.320 18.61  1  1    4    1 110

# We get the same result by feeding a model instead of a data.frame
mod <- lm(mpg ~ hp + wt, mtcars)
cd <- counterfactual(model = mod, hp = c(100, 110))
cd[cd$rowid %in% 1:3,]

#>    rowid    wt  hp
#> 1      1 2.620 100
#> 2      2 2.875 100
#> 3      3 2.320 100
#> 33     1 2.620 110
#> 34     2 2.875 110
#> 35     3 2.320 110

# Use in `marginaleffects` to compute "Counterfactual Average Marginal Effects"
marginaleffects(mod, newdata = counterfactual(hp = c(100, 110)))

#>     rowid     type term        dydx   std.error predicted_response    wt  hp
#> 1       1 response   hp -0.03177295 0.009029704           23.89006 2.620 100
#> 2       1 response   wt -3.87783074 0.632733547           23.89006 2.620 100
#> 3       2 response   hp -0.03177295 0.009029704           22.90121 2.875 100
#> 4       2 response   wt -3.87783074 0.632733322           22.90121 2.875 100
#> 5       3 response   hp -0.03177295 0.009029704           25.05341 2.320 100
#> 6       3 response   wt -3.87783074 0.632733097           25.05341 2.320 100
#> 7       4 response   wt -3.87783074 0.632733322           21.58275 3.215 100
#> 8       4 response   hp -0.03177295 0.009029704           21.58275 3.215 100
#> 9       5 response   hp -0.03177295 0.009029704           20.71024 3.440 100
#> 10      5 response   wt -3.87783074 0.632733322           20.71024 3.440 100
#> 11      6 response   hp -0.03177295 0.009029704           20.63268 3.460 100
#> 12      6 response   wt -3.87783074 0.632733547           20.63268 3.460 100
#> 13      7 response   hp -0.03177295 0.009029704           20.20612 3.570 100
#> 14      7 response   wt -3.87783074 0.632733322           20.20612 3.570 100
#> 15      8 response   wt -3.87783074 0.632733322           21.67970 3.190 100
#> 16      8 response   hp -0.03177295 0.009029704           21.67970 3.190 100
#> 17      9 response   hp -0.03177295 0.009029704           21.83481 3.150 100
#> 18      9 response   wt -3.87783074 0.632733547           21.83481 3.150 100
#> 19     10 response   hp -0.03177295 0.009029704           20.71024 3.440 100
#> 20     10 response   wt -3.87783074 0.632733322           20.71024 3.440 100
#> 21     11 response   hp -0.03177295 0.009029704           20.71024 3.440 100
#> 22     11 response   wt -3.87783074 0.632733322           20.71024 3.440 100
#> 23     12 response   wt -3.87783074 0.632733547           18.26720 4.070 100
#> 24     12 response   hp -0.03177295 0.009029704           18.26720 4.070 100
#> 25     13 response   wt -3.87783074 0.632733547           19.58567 3.730 100
#> 26     13 response   hp -0.03177295 0.009029704           19.58567 3.730 100
#> 27     14 response   hp -0.03177295 0.009029704           19.39178 3.780 100
#> 28     14 response   wt -3.87783074 0.632733322           19.39178 3.780 100
#> 29     15 response   hp -0.03177295 0.009029704           13.69136 5.250 100
#> 30     15 response   wt -3.87783074 0.632733547           13.69136 5.250 100
#> 31     16 response   hp -0.03177295 0.009029704           13.01662 5.424 100
#> 32     16 response   wt -3.87783074 0.632733322           13.01662 5.424 100
#> 33     17 response   wt -3.87783074 0.632733547           13.32297 5.345 100
#> 34     17 response   hp -0.03177295 0.009029704           13.32297 5.345 100
#> 35     18 response   hp -0.03177295 0.009029704           25.51875 2.200 100
#> 36     18 response   wt -3.87783074 0.632733322           25.51875 2.200 100
#> 37     19 response   hp -0.03177295 0.009029704           27.78728 1.615 100
#> 38     19 response   wt -3.87783074 0.632733772           27.78728 1.615 100
#> 39     20 response   hp -0.03177295 0.009029704           26.93416 1.835 100
#> 40     20 response   wt -3.87783074 0.632733322           26.93416 1.835 100
#> 41     21 response   wt -3.87783074 0.632733097           24.49112 2.465 100
#> 42     21 response   hp -0.03177295 0.009029704           24.49112 2.465 100
#> 43     22 response   hp -0.03177295 0.009029704           20.40001 3.520 100
#> 44     22 response   wt -3.87783074 0.632733322           20.40001 3.520 100
#> 45     23 response   hp -0.03177295 0.009029704           20.72963 3.435 100
#> 46     23 response   wt -3.87783074 0.632733547           20.72963 3.435 100
#> 47     24 response   hp -0.03177295 0.009029704           19.15911 3.840 100
#> 48     24 response   wt -3.87783074 0.632733996           19.15911 3.840 100
#> 49     25 response   wt -3.87783074 0.632733322           19.13972 3.845 100
#> 50     25 response   hp -0.03177295 0.009029704           19.13972 3.845 100
#> 51     26 response   wt -3.87783074 0.632733322           26.54637 1.935 100
#> 52     26 response   hp -0.03177295 0.009029704           26.54637 1.935 100
#> 53     27 response   wt -3.87783074 0.632733547           25.75142 2.140 100
#> 54     27 response   hp -0.03177295 0.009029704           25.75142 2.140 100
#> 55     28 response   hp -0.03177295 0.009029704           28.18282 1.513 100
#> 56     28 response   wt -3.87783074 0.632733322           28.18282 1.513 100
#> 57     29 response   hp -0.03177295 0.009029704           21.75725 3.170 100
#> 58     29 response   wt -3.87783074 0.632733547           21.75725 3.170 100
#> 59     30 response   wt -3.87783074 0.632733547           23.30838 2.770 100
#> 60     30 response   hp -0.03177295 0.009029704           23.30838 2.770 100
#> 61     31 response   wt -3.87783074 0.632733322           20.20612 3.570 100
#> 62     31 response   hp -0.03177295 0.009029704           20.20612 3.570 100
#> 63     32 response   hp -0.03177295 0.009029704           23.26961 2.780 100
#> 64     32 response   wt -3.87783074 0.632733322           23.26961 2.780 100
#> 65     33 response   hp -0.03177295 0.009029710           23.57233 2.620 110
#> 66     33 response   wt -3.87783074 0.632733547           23.57233 2.620 110
#> 67     34 response   wt -3.87783074 0.632733322           22.58348 2.875 110
#> 68     34 response   hp -0.03177295 0.009029710           22.58348 2.875 110
#> 69     35 response   wt -3.87783074 0.632733097           24.73568 2.320 110
#> 70     35 response   hp -0.03177295 0.009029710           24.73568 2.320 110
#> 71     36 response   hp -0.03177295 0.009029710           21.26502 3.215 110
#> 72     36 response   wt -3.87783074 0.632733322           21.26502 3.215 110
#> 73     37 response   hp -0.03177295 0.009029710           20.39251 3.440 110
#> 74     37 response   wt -3.87783074 0.632733322           20.39251 3.440 110
#> 75     38 response   wt -3.87783074 0.632733547           20.31495 3.460 110
#> 76     38 response   hp -0.03177295 0.009029710           20.31495 3.460 110
#> 77     39 response   wt -3.87783074 0.632733322           19.88839 3.570 110
#> 78     39 response   hp -0.03177295 0.009029710           19.88839 3.570 110
#> 79     40 response   wt -3.87783074 0.632733322           21.36197 3.190 110
#> 80     40 response   hp -0.03177295 0.009029710           21.36197 3.190 110
#> 81     41 response   hp -0.03177295 0.009029710           21.51708 3.150 110
#> 82     41 response   wt -3.87783074 0.632733547           21.51708 3.150 110
#> 83     42 response   hp -0.03177295 0.009029710           20.39251 3.440 110
#> 84     42 response   wt -3.87783074 0.632733322           20.39251 3.440 110
#> 85     43 response   hp -0.03177295 0.009029710           20.39251 3.440 110
#> 86     43 response   wt -3.87783074 0.632733322           20.39251 3.440 110
#> 87     44 response   wt -3.87783074 0.632733547           17.94947 4.070 110
#> 88     44 response   hp -0.03177295 0.009029710           17.94947 4.070 110
#> 89     45 response   hp -0.03177295 0.009029710           19.26794 3.730 110
#> 90     45 response   wt -3.87783074 0.632733547           19.26794 3.730 110
#> 91     46 response   hp -0.03177295 0.009029710           19.07405 3.780 110
#> 92     46 response   wt -3.87783074 0.632733322           19.07405 3.780 110
#> 93     47 response   hp -0.03177295 0.009029710           13.37363 5.250 110
#> 94     47 response   wt -3.87783074 0.632733547           13.37363 5.250 110
#> 95     48 response   wt -3.87783074 0.632733322           12.69889 5.424 110
#> 96     48 response   hp -0.03177295 0.009029710           12.69889 5.424 110
#> 97     49 response   hp -0.03177295 0.009029710           13.00524 5.345 110
#> 98     49 response   wt -3.87783074 0.632733547           13.00524 5.345 110
#> 99     50 response   hp -0.03177295 0.009029710           25.20102 2.200 110
#> 100    50 response   wt -3.87783074 0.632733322           25.20102 2.200 110
#> 101    51 response   hp -0.03177295 0.009029710           27.46955 1.615 110
#> 102    51 response   wt -3.87783074 0.632733772           27.46955 1.615 110
#> 103    52 response   wt -3.87783074 0.632733322           26.61643 1.835 110
#> 104    52 response   hp -0.03177295 0.009029710           26.61643 1.835 110
#> 105    53 response   wt -3.87783074 0.632733097           24.17339 2.465 110
#> 106    53 response   hp -0.03177295 0.009029710           24.17339 2.465 110
#> 107    54 response   hp -0.03177295 0.009029710           20.08228 3.520 110
#> 108    54 response   wt -3.87783074 0.632733322           20.08228 3.520 110
#> 109    55 response   hp -0.03177295 0.009029710           20.41190 3.435 110
#> 110    55 response   wt -3.87783074 0.632733547           20.41190 3.435 110
#> 111    56 response   hp -0.03177295 0.009029710           18.84138 3.840 110
#> 112    56 response   wt -3.87783074 0.632733996           18.84138 3.840 110
#> 113    57 response   wt -3.87783074 0.632733322           18.82199 3.845 110
#> 114    57 response   hp -0.03177295 0.009029710           18.82199 3.845 110
#> 115    58 response   hp -0.03177295 0.009029710           26.22864 1.935 110
#> 116    58 response   wt -3.87783074 0.632733322           26.22864 1.935 110
#> 117    59 response   hp -0.03177295 0.009029710           25.43369 2.140 110
#> 118    59 response   wt -3.87783074 0.632733547           25.43369 2.140 110
#> 119    60 response   hp -0.03177295 0.009029710           27.86509 1.513 110
#> 120    60 response   wt -3.87783074 0.632733322           27.86509 1.513 110
#> 121    61 response   wt -3.87783074 0.632733547           21.43952 3.170 110
#> 122    61 response   hp -0.03177295 0.009029710           21.43952 3.170 110
#> 123    62 response   hp -0.03177295 0.009029710           22.99065 2.770 110
#> 124    62 response   wt -3.87783074 0.632733547           22.99065 2.770 110
#> 125    63 response   hp -0.03177295 0.009029710           19.88839 3.570 110
#> 126    63 response   wt -3.87783074 0.632733322           19.88839 3.570 110
#> 127    64 response   hp -0.03177295 0.009029710           22.95188 2.780 110
#> 128    64 response   wt -3.87783074 0.632733322           22.95188 2.780 110