Generate "counterfactual" datasets for use in marginaleffects's newdata argument

Generate "counterfactual" datasets for use in marginaleffects's newdata argument

counterfactual(..., model = NULL, newdata = NULL)

Arguments

...

named arguments with vectors of values for the variables to construct (see Examples below.)

model

Model object

newdata

data.frame (one and only one of the model and newdata arguments must be true).

Value

A data.frame where each row of the original data is repeated multiple times for each of the values of the variables in the at list. See example below.

Details

If counterfactual is used in a marginaleffects or predictions call as the newdata argument, users do not need to specify the model or newdata argument. The data is extracted automatically from the model.

If users supply a model, the data used to fit that model is retrieved using the insight::get_data function.

If users supply a model, the data used to fit that model is retrieved using the insight::get_data function, and then replicated with different values of the variables in the at list.

Examples

# All rows are repeated twice, with different values of `hp`
cd <- counterfactual(newdata = mtcars, hp = c(100, 110))
cd[cd$rowid %in% 1:3,]
#>    rowid_original  mpg cyl disp drat    wt  qsec vs am gear carb  hp
#> 1               1 21.0   6  160 3.90 2.620 16.46  0  1    4    4 100
#> 2               2 21.0   6  160 3.90 2.875 17.02  0  1    4    4 100
#> 3               3 22.8   4  108 3.85 2.320 18.61  1  1    4    1 100
#> 33              1 21.0   6  160 3.90 2.620 16.46  0  1    4    4 110
#> 34              2 21.0   6  160 3.90 2.875 17.02  0  1    4    4 110
#> 35              3 22.8   4  108 3.85 2.320 18.61  1  1    4    1 110

# We get the same result by feeding a model instead of a data.frame
mod <- lm(mpg ~ hp + wt, mtcars)
cd <- counterfactual(model = mod, hp = c(100, 110))
cd[cd$rowid %in% 1:3,]
#>    rowid_original    wt  hp
#> 1               1 2.620 100
#> 2               2 2.875 100
#> 3               3 2.320 100
#> 33              1 2.620 110
#> 34              2 2.875 110
#> 35              3 2.320 110

# Use in `marginaleffects` to compute "Counterfactual Average Marginal Effects"
marginaleffects(mod, newdata = counterfactual(hp = c(100, 110)))
#>     rowid     type term        dydx   std.error rowid_original    wt  hp
#> 1       1 response   hp -0.03177295 0.009029704              1 2.620 100
#> 2       2 response   hp -0.03177295 0.009029704              2 2.875 100
#> 3       3 response   hp -0.03177295 0.009029704              3 2.320 100
#> 4       4 response   hp -0.03177295 0.009029704              4 3.215 100
#> 5       5 response   hp -0.03177295 0.009029704              5 3.440 100
#> 6       6 response   hp -0.03177295 0.009029704              6 3.460 100
#> 7       7 response   hp -0.03177295 0.009029704              7 3.570 100
#> 8       8 response   hp -0.03177295 0.009029704              8 3.190 100
#> 9       9 response   hp -0.03177295 0.009029704              9 3.150 100
#> 10     10 response   hp -0.03177295 0.009029704             10 3.440 100
#> 11     11 response   hp -0.03177295 0.009029704             11 3.440 100
#> 12     12 response   hp -0.03177295 0.009029704             12 4.070 100
#> 13     13 response   hp -0.03177295 0.009029704             13 3.730 100
#> 14     14 response   hp -0.03177295 0.009029704             14 3.780 100
#> 15     15 response   hp -0.03177295 0.009029704             15 5.250 100
#> 16     16 response   hp -0.03177295 0.009029704             16 5.424 100
#> 17     17 response   hp -0.03177295 0.009029704             17 5.345 100
#> 18     18 response   hp -0.03177295 0.009029704             18 2.200 100
#> 19     19 response   hp -0.03177295 0.009029704             19 1.615 100
#> 20     20 response   hp -0.03177295 0.009029704             20 1.835 100
#> 21     21 response   hp -0.03177295 0.009029704             21 2.465 100
#> 22     22 response   hp -0.03177295 0.009029704             22 3.520 100
#> 23     23 response   hp -0.03177295 0.009029704             23 3.435 100
#> 24     24 response   hp -0.03177295 0.009029704             24 3.840 100
#> 25     25 response   hp -0.03177295 0.009029704             25 3.845 100
#> 26     26 response   hp -0.03177295 0.009029704             26 1.935 100
#> 27     27 response   hp -0.03177295 0.009029704             27 2.140 100
#> 28     28 response   hp -0.03177295 0.009029704             28 1.513 100
#> 29     29 response   hp -0.03177295 0.009029704             29 3.170 100
#> 30     30 response   hp -0.03177295 0.009029704             30 2.770 100
#> 31     31 response   hp -0.03177295 0.009029704             31 3.570 100
#> 32     32 response   hp -0.03177295 0.009029704             32 2.780 100
#> 33     33 response   hp -0.03177295 0.009029710              1 2.620 110
#> 34     34 response   hp -0.03177295 0.009029710              2 2.875 110
#> 35     35 response   hp -0.03177295 0.009029710              3 2.320 110
#> 36     36 response   hp -0.03177295 0.009029710              4 3.215 110
#> 37     37 response   hp -0.03177295 0.009029710              5 3.440 110
#> 38     38 response   hp -0.03177295 0.009029710              6 3.460 110
#> 39     39 response   hp -0.03177295 0.009029710              7 3.570 110
#> 40     40 response   hp -0.03177295 0.009029710              8 3.190 110
#> 41     41 response   hp -0.03177295 0.009029710              9 3.150 110
#> 42     42 response   hp -0.03177295 0.009029710             10 3.440 110
#> 43     43 response   hp -0.03177295 0.009029710             11 3.440 110
#> 44     44 response   hp -0.03177295 0.009029710             12 4.070 110
#> 45     45 response   hp -0.03177295 0.009029710             13 3.730 110
#> 46     46 response   hp -0.03177295 0.009029710             14 3.780 110
#> 47     47 response   hp -0.03177295 0.009029710             15 5.250 110
#> 48     48 response   hp -0.03177295 0.009029710             16 5.424 110
#> 49     49 response   hp -0.03177295 0.009029710             17 5.345 110
#> 50     50 response   hp -0.03177295 0.009029710             18 2.200 110
#> 51     51 response   hp -0.03177295 0.009029710             19 1.615 110
#> 52     52 response   hp -0.03177295 0.009029710             20 1.835 110
#> 53     53 response   hp -0.03177295 0.009029710             21 2.465 110
#> 54     54 response   hp -0.03177295 0.009029710             22 3.520 110
#> 55     55 response   hp -0.03177295 0.009029710             23 3.435 110
#> 56     56 response   hp -0.03177295 0.009029710             24 3.840 110
#> 57     57 response   hp -0.03177295 0.009029710             25 3.845 110
#> 58     58 response   hp -0.03177295 0.009029710             26 1.935 110
#> 59     59 response   hp -0.03177295 0.009029710             27 2.140 110
#> 60     60 response   hp -0.03177295 0.009029710             28 1.513 110
#> 61     61 response   hp -0.03177295 0.009029710             29 3.170 110
#> 62     62 response   hp -0.03177295 0.009029710             30 2.770 110
#> 63     63 response   hp -0.03177295 0.009029710             31 3.570 110
#> 64     64 response   hp -0.03177295 0.009029710             32 2.780 110
#> 65      1 response   wt -3.87783074 0.632733547              1 2.620 100
#> 66      2 response   wt -3.87783074 0.632733322              2 2.875 100
#> 67      3 response   wt -3.87783074 0.632733097              3 2.320 100
#> 68      4 response   wt -3.87783074 0.632733322              4 3.215 100
#> 69      5 response   wt -3.87783074 0.632733322              5 3.440 100
#> 70      6 response   wt -3.87783074 0.632733547              6 3.460 100
#> 71      7 response   wt -3.87783074 0.632733322              7 3.570 100
#> 72      8 response   wt -3.87783074 0.632733322              8 3.190 100
#> 73      9 response   wt -3.87783074 0.632733547              9 3.150 100
#> 74     10 response   wt -3.87783074 0.632733322             10 3.440 100
#> 75     11 response   wt -3.87783074 0.632733322             11 3.440 100
#> 76     12 response   wt -3.87783074 0.632733547             12 4.070 100
#> 77     13 response   wt -3.87783074 0.632733547             13 3.730 100
#> 78     14 response   wt -3.87783074 0.632733322             14 3.780 100
#> 79     15 response   wt -3.87783074 0.632733547             15 5.250 100
#> 80     16 response   wt -3.87783074 0.632733322             16 5.424 100
#> 81     17 response   wt -3.87783074 0.632733547             17 5.345 100
#> 82     18 response   wt -3.87783074 0.632733322             18 2.200 100
#> 83     19 response   wt -3.87783074 0.632733772             19 1.615 100
#> 84     20 response   wt -3.87783074 0.632733322             20 1.835 100
#> 85     21 response   wt -3.87783074 0.632733097             21 2.465 100
#> 86     22 response   wt -3.87783074 0.632733322             22 3.520 100
#> 87     23 response   wt -3.87783074 0.632733547             23 3.435 100
#> 88     24 response   wt -3.87783074 0.632733996             24 3.840 100
#> 89     25 response   wt -3.87783074 0.632733322             25 3.845 100
#> 90     26 response   wt -3.87783074 0.632733322             26 1.935 100
#> 91     27 response   wt -3.87783074 0.632733547             27 2.140 100
#> 92     28 response   wt -3.87783074 0.632733322             28 1.513 100
#> 93     29 response   wt -3.87783074 0.632733547             29 3.170 100
#> 94     30 response   wt -3.87783074 0.632733547             30 2.770 100
#> 95     31 response   wt -3.87783074 0.632733322             31 3.570 100
#> 96     32 response   wt -3.87783074 0.632733322             32 2.780 100
#> 97     33 response   wt -3.87783074 0.632733547              1 2.620 110
#> 98     34 response   wt -3.87783074 0.632733322              2 2.875 110
#> 99     35 response   wt -3.87783074 0.632733097              3 2.320 110
#> 100    36 response   wt -3.87783074 0.632733322              4 3.215 110
#> 101    37 response   wt -3.87783074 0.632733322              5 3.440 110
#> 102    38 response   wt -3.87783074 0.632733547              6 3.460 110
#> 103    39 response   wt -3.87783074 0.632733322              7 3.570 110
#> 104    40 response   wt -3.87783074 0.632733322              8 3.190 110
#> 105    41 response   wt -3.87783074 0.632733547              9 3.150 110
#> 106    42 response   wt -3.87783074 0.632733322             10 3.440 110
#> 107    43 response   wt -3.87783074 0.632733322             11 3.440 110
#> 108    44 response   wt -3.87783074 0.632733547             12 4.070 110
#> 109    45 response   wt -3.87783074 0.632733547             13 3.730 110
#> 110    46 response   wt -3.87783074 0.632733322             14 3.780 110
#> 111    47 response   wt -3.87783074 0.632733547             15 5.250 110
#> 112    48 response   wt -3.87783074 0.632733322             16 5.424 110
#> 113    49 response   wt -3.87783074 0.632733547             17 5.345 110
#> 114    50 response   wt -3.87783074 0.632733322             18 2.200 110
#> 115    51 response   wt -3.87783074 0.632733772             19 1.615 110
#> 116    52 response   wt -3.87783074 0.632733322             20 1.835 110
#> 117    53 response   wt -3.87783074 0.632733097             21 2.465 110
#> 118    54 response   wt -3.87783074 0.632733322             22 3.520 110
#> 119    55 response   wt -3.87783074 0.632733547             23 3.435 110
#> 120    56 response   wt -3.87783074 0.632733996             24 3.840 110
#> 121    57 response   wt -3.87783074 0.632733322             25 3.845 110
#> 122    58 response   wt -3.87783074 0.632733322             26 1.935 110
#> 123    59 response   wt -3.87783074 0.632733547             27 2.140 110
#> 124    60 response   wt -3.87783074 0.632733322             28 1.513 110
#> 125    61 response   wt -3.87783074 0.632733547             29 3.170 110
#> 126    62 response   wt -3.87783074 0.632733547             30 2.770 110
#> 127    63 response   wt -3.87783074 0.632733322             31 3.570 110
#> 128    64 response   wt -3.87783074 0.632733322             32 2.780 110