Generate "counterfactual" datasets for use in marginaleffects's newdata argument

Generate "counterfactual" datasets for use in marginaleffects's newdata argument

counterfactual(..., model = NULL, newdata = NULL)

Arguments

...
named arguments with vectors of values for the variables to construct (see Examples below.)

model
Model object

newdata
data.frame (one and only one of the model and newdata arguments must be true).

Value

A data.frame where each row of the original data is repeated multiple times for each of the values of the variables in the at list. See example below.

Details

If counterfactual is used in a marginaleffects or predictions call as the newdata argument, users do not need to specify the model or newdata argument. The data is extracted automatically from the model.

If users supply a model, the data used to fit that model is retrieved using the insight::get_data function.

If users supply a model, the data used to fit that model is retrieved using the insight::get_data function, and then replicated with different values of the variables in the at list.

Examples

# All rows are repeated twice, with different values of `hp`
cd <- counterfactual(newdata = mtcars, hp = c(100, 110))
cd[cd$rowid %in% 1:3,]
#>    rowid  mpg cyl disp drat    wt  qsec vs am gear carb  hp
#> 1      1 21.0   6  160 3.90 2.620 16.46  0  1    4    4 100
#> 2      2 21.0   6  160 3.90 2.875 17.02  0  1    4    4 100
#> 3      3 22.8   4  108 3.85 2.320 18.61  1  1    4    1 100
#> 33     1 21.0   6  160 3.90 2.620 16.46  0  1    4    4 110
#> 34     2 21.0   6  160 3.90 2.875 17.02  0  1    4    4 110
#> 35     3 22.8   4  108 3.85 2.320 18.61  1  1    4    1 110

# We get the same result by feeding a model instead of a data.frame
mod <- lm(mpg ~ hp + wt, mtcars)
cd <- counterfactual(model = mod, hp = c(100, 110))
cd[cd$rowid %in% 1:3,]
#>    rowid    wt  hp
#> 1      1 2.620 100
#> 2      2 2.875 100
#> 3      3 2.320 100
#> 33     1 2.620 110
#> 34     2 2.875 110
#> 35     3 2.320 110

# Use in `marginaleffects` to compute "Counterfactual Average Marginal Effects"
marginaleffects(mod, newdata = counterfactual(hp = c(100, 110)))
#>     rowid     type term        dydx   std.error    wt  hp predicted
#> 1       1 response   hp -0.03177295 0.009029704 2.620 100  23.89006
#> 2       1 response   hp -0.03177295 0.009029704 2.620 110  23.57233
#> 3       1 response   wt -3.87783074 0.632733547 2.620 100  23.89006
#> 4       1 response   wt -3.87783074 0.632733547 2.620 110  23.57233
#> 5       2 response   hp -0.03177295 0.009029704 2.875 100  22.90121
#> 6       2 response   hp -0.03177295 0.009029704 2.875 110  22.58348
#> 7       2 response   wt -3.87783074 0.632733322 2.875 100  22.90121
#> 8       2 response   wt -3.87783074 0.632733322 2.875 110  22.58348
#> 9       3 response   wt -3.87783074 0.632733097 2.320 110  24.73568
#> 10      3 response   wt -3.87783074 0.632733097 2.320 100  25.05341
#> 11      3 response   hp -0.03177295 0.009029704 2.320 110  24.73568
#> 12      3 response   hp -0.03177295 0.009029704 2.320 100  25.05341
#> 13      4 response   wt -3.87783074 0.632733322 3.215 110  21.26502
#> 14      4 response   wt -3.87783074 0.632733322 3.215 100  21.58275
#> 15      4 response   hp -0.03177295 0.009029704 3.215 110  21.26502
#> 16      4 response   hp -0.03177295 0.009029704 3.215 100  21.58275
#> 17      5 response   hp -0.03177295 0.009029704 3.440 100  20.71024
#> 18      5 response   hp -0.03177295 0.009029704 3.440 110  20.39251
#> 19      5 response   wt -3.87783074 0.632733322 3.440 100  20.71024
#> 20      5 response   wt -3.87783074 0.632733322 3.440 110  20.39251
#> 21      6 response   wt -3.87783074 0.632733547 3.460 100  20.63268
#> 22      6 response   wt -3.87783074 0.632733547 3.460 110  20.31495
#> 23      6 response   hp -0.03177295 0.009029704 3.460 100  20.63268
#> 24      6 response   hp -0.03177295 0.009029704 3.460 110  20.31495
#> 25      7 response   wt -3.87783074 0.632733322 3.570 110  19.88839
#> 26      7 response   wt -3.87783074 0.632733322 3.570 100  20.20612
#> 27      7 response   hp -0.03177295 0.009029704 3.570 110  19.88839
#> 28      7 response   hp -0.03177295 0.009029704 3.570 100  20.20612
#> 29      8 response   wt -3.87783074 0.632733322 3.190 110  21.36197
#> 30      8 response   wt -3.87783074 0.632733322 3.190 100  21.67970
#> 31      8 response   hp -0.03177295 0.009029704 3.190 110  21.36197
#> 32      8 response   hp -0.03177295 0.009029704 3.190 100  21.67970
#> 33      9 response   hp -0.03177295 0.009029704 3.150 110  21.51708
#> 34      9 response   hp -0.03177295 0.009029704 3.150 100  21.83481
#> 35      9 response   wt -3.87783074 0.632733547 3.150 110  21.51708
#> 36      9 response   wt -3.87783074 0.632733547 3.150 100  21.83481
#> 37     10 response   wt -3.87783074 0.632733322 3.440 100  20.71024
#> 38     10 response   wt -3.87783074 0.632733322 3.440 110  20.39251
#> 39     10 response   hp -0.03177295 0.009029704 3.440 100  20.71024
#> 40     10 response   hp -0.03177295 0.009029704 3.440 110  20.39251
#> 41     11 response   hp -0.03177295 0.009029704 3.440 110  20.39251
#> 42     11 response   hp -0.03177295 0.009029704 3.440 100  20.71024
#> 43     11 response   wt -3.87783074 0.632733322 3.440 110  20.39251
#> 44     11 response   wt -3.87783074 0.632733322 3.440 100  20.71024
#> 45     12 response   hp -0.03177295 0.009029704 4.070 110  17.94947
#> 46     12 response   hp -0.03177295 0.009029704 4.070 100  18.26720
#> 47     12 response   wt -3.87783074 0.632733547 4.070 110  17.94947
#> 48     12 response   wt -3.87783074 0.632733547 4.070 100  18.26720
#> 49     13 response   hp -0.03177295 0.009029704 3.730 110  19.26794
#> 50     13 response   hp -0.03177295 0.009029704 3.730 100  19.58567
#> 51     13 response   wt -3.87783074 0.632733547 3.730 110  19.26794
#> 52     13 response   wt -3.87783074 0.632733547 3.730 100  19.58567
#> 53     14 response   hp -0.03177295 0.009029704 3.780 100  19.39178
#> 54     14 response   hp -0.03177295 0.009029704 3.780 110  19.07405
#> 55     14 response   wt -3.87783074 0.632733322 3.780 100  19.39178
#> 56     14 response   wt -3.87783074 0.632733322 3.780 110  19.07405
#> 57     15 response   hp -0.03177295 0.009029704 5.250 100  13.69136
#> 58     15 response   hp -0.03177295 0.009029704 5.250 110  13.37363
#> 59     15 response   wt -3.87783074 0.632733547 5.250 100  13.69136
#> 60     15 response   wt -3.87783074 0.632733547 5.250 110  13.37363
#> 61     16 response   hp -0.03177295 0.009029704 5.424 110  12.69889
#> 62     16 response   hp -0.03177295 0.009029704 5.424 100  13.01662
#> 63     16 response   wt -3.87783074 0.632733322 5.424 110  12.69889
#> 64     16 response   wt -3.87783074 0.632733322 5.424 100  13.01662
#> 65     17 response   wt -3.87783074 0.632733547 5.345 110  13.00524
#> 66     17 response   wt -3.87783074 0.632733547 5.345 100  13.32297
#> 67     17 response   hp -0.03177295 0.009029704 5.345 110  13.00524
#> 68     17 response   hp -0.03177295 0.009029704 5.345 100  13.32297
#> 69     18 response   hp -0.03177295 0.009029704 2.200 100  25.51875
#> 70     18 response   hp -0.03177295 0.009029704 2.200 110  25.20102
#> 71     18 response   wt -3.87783074 0.632733322 2.200 100  25.51875
#> 72     18 response   wt -3.87783074 0.632733322 2.200 110  25.20102
#> 73     19 response   hp -0.03177295 0.009029704 1.615 100  27.78728
#> 74     19 response   hp -0.03177295 0.009029704 1.615 110  27.46955
#> 75     19 response   wt -3.87783074 0.632733772 1.615 100  27.78728
#> 76     19 response   wt -3.87783074 0.632733772 1.615 110  27.46955
#> 77     20 response   wt -3.87783074 0.632733322 1.835 110  26.61643
#> 78     20 response   wt -3.87783074 0.632733322 1.835 100  26.93416
#> 79     20 response   hp -0.03177295 0.009029704 1.835 110  26.61643
#> 80     20 response   hp -0.03177295 0.009029704 1.835 100  26.93416
#> 81     21 response   wt -3.87783074 0.632733097 2.465 110  24.17339
#> 82     21 response   wt -3.87783074 0.632733097 2.465 100  24.49112
#> 83     21 response   hp -0.03177295 0.009029704 2.465 110  24.17339
#> 84     21 response   hp -0.03177295 0.009029704 2.465 100  24.49112
#> 85     22 response   hp -0.03177295 0.009029704 3.520 110  20.08228
#> 86     22 response   hp -0.03177295 0.009029704 3.520 100  20.40001
#> 87     22 response   wt -3.87783074 0.632733322 3.520 110  20.08228
#> 88     22 response   wt -3.87783074 0.632733322 3.520 100  20.40001
#> 89     23 response   hp -0.03177295 0.009029704 3.435 100  20.72963
#> 90     23 response   hp -0.03177295 0.009029704 3.435 110  20.41190
#> 91     23 response   wt -3.87783074 0.632733547 3.435 100  20.72963
#> 92     23 response   wt -3.87783074 0.632733547 3.435 110  20.41190
#> 93     24 response   hp -0.03177295 0.009029704 3.840 110  18.84138
#> 94     24 response   hp -0.03177295 0.009029704 3.840 100  19.15911
#> 95     24 response   wt -3.87783074 0.632733996 3.840 110  18.84138
#> 96     24 response   wt -3.87783074 0.632733996 3.840 100  19.15911
#> 97     25 response   hp -0.03177295 0.009029704 3.845 110  18.82199
#> 98     25 response   hp -0.03177295 0.009029704 3.845 100  19.13972
#> 99     25 response   wt -3.87783074 0.632733322 3.845 110  18.82199
#> 100    25 response   wt -3.87783074 0.632733322 3.845 100  19.13972
#> 101    26 response   wt -3.87783074 0.632733322 1.935 110  26.22864
#> 102    26 response   wt -3.87783074 0.632733322 1.935 100  26.54637
#> 103    26 response   hp -0.03177295 0.009029704 1.935 110  26.22864
#> 104    26 response   hp -0.03177295 0.009029704 1.935 100  26.54637
#> 105    27 response   wt -3.87783074 0.632733547 2.140 100  25.75142
#> 106    27 response   wt -3.87783074 0.632733547 2.140 110  25.43369
#> 107    27 response   hp -0.03177295 0.009029704 2.140 100  25.75142
#> 108    27 response   hp -0.03177295 0.009029704 2.140 110  25.43369
#> 109    28 response   hp -0.03177295 0.009029704 1.513 100  28.18282
#> 110    28 response   hp -0.03177295 0.009029704 1.513 110  27.86509
#> 111    28 response   wt -3.87783074 0.632733322 1.513 100  28.18282
#> 112    28 response   wt -3.87783074 0.632733322 1.513 110  27.86509
#> 113    29 response   hp -0.03177295 0.009029704 3.170 110  21.43952
#> 114    29 response   hp -0.03177295 0.009029704 3.170 100  21.75725
#> 115    29 response   wt -3.87783074 0.632733547 3.170 110  21.43952
#> 116    29 response   wt -3.87783074 0.632733547 3.170 100  21.75725
#> 117    30 response   wt -3.87783074 0.632733547 2.770 110  22.99065
#> 118    30 response   wt -3.87783074 0.632733547 2.770 100  23.30838
#> 119    30 response   hp -0.03177295 0.009029704 2.770 110  22.99065
#> 120    30 response   hp -0.03177295 0.009029704 2.770 100  23.30838
#> 121    31 response   wt -3.87783074 0.632733322 3.570 100  20.20612
#> 122    31 response   wt -3.87783074 0.632733322 3.570 110  19.88839
#> 123    31 response   hp -0.03177295 0.009029704 3.570 100  20.20612
#> 124    31 response   hp -0.03177295 0.009029704 3.570 110  19.88839
#> 125    32 response   hp -0.03177295 0.009029704 2.780 100  23.26961
#> 126    32 response   hp -0.03177295 0.009029704 2.780 110  22.95188
#> 127    32 response   wt -3.87783074 0.632733322 2.780 100  23.26961
#> 128    32 response   wt -3.87783074 0.632733322 2.780 110  22.95188