Generate "counterfactual" datasets for use in marginaleffects's newdata argument

Generate "counterfactual" datasets for use in marginaleffects's newdata argument

counterfactual(..., model = NULL, data = NULL)

Arguments

...
named arguments with vectors of values for the variables to construct (see Examples below.)

model
Model object

data
data.frame (one and only one of the model and data arguments must be true).

Value

A data.frame where each row of the original data is repeated multiple times for each of the values of the variables in the at list. See example below.

Details

If counterfactual is used in a marginaleffects or predictions call as the newdata argument, users do not need to specify the model or data argument. The data is extracted automatically from the model.

If users supply a model, the data used to fit that model is retrieved using the insight::get_data function.

If users supply a model, the data used to fit that model is retrieved using the insight::get_data function, and then replicated with different values of the variables in the at list.

Examples

# All rows are repeated twice, with different values of `hp`
cd <- counterfactual(data = mtcars, hp = c(100, 110))
cd[cd$rowid %in% 1:3,]

#>    rowid  mpg cyl disp drat    wt  qsec vs am gear carb  hp
#> 1      1 21.0   6  160 3.90 2.620 16.46  0  1    4    4 100
#> 2      2 21.0   6  160 3.90 2.875 17.02  0  1    4    4 100
#> 3      3 22.8   4  108 3.85 2.320 18.61  1  1    4    1 100
#> 33     1 21.0   6  160 3.90 2.620 16.46  0  1    4    4 110
#> 34     2 21.0   6  160 3.90 2.875 17.02  0  1    4    4 110
#> 35     3 22.8   4  108 3.85 2.320 18.61  1  1    4    1 110

# We get the same result by feeding a model instead of a data.frame
mod <- lm(mpg ~ hp + wt, mtcars)
cd <- counterfactual(model = mod, hp = c(100, 110))
cd[cd$rowid %in% 1:3,]

#>    rowid    wt  hp
#> 1      1 2.620 100
#> 2      2 2.875 100
#> 3      3 2.320 100
#> 33     1 2.620 110
#> 34     2 2.875 110
#> 35     3 2.320 110

# Use in `marginaleffects` to compute "Counterfactual Average Marginal Effects"
marginaleffects(mod, newdata = counterfactual(hp = c(100, 110)))

#>     rowid     type term        dydx   std.error    wt  hp predicted
#> 1       1 response   hp -0.03177295 0.009029704 2.620 100  23.89006
#> 2       1 response   hp -0.03177295 0.009029704 2.620 110  23.57233
#> 3       2 response   hp -0.03177295 0.009029704 2.875 100  22.90121
#> 4       2 response   hp -0.03177295 0.009029704 2.875 110  22.58348
#> 5       3 response   hp -0.03177295 0.009029704 2.320 110  24.73568
#> 6       3 response   hp -0.03177295 0.009029704 2.320 100  25.05341
#> 7       4 response   hp -0.03177295 0.009029704 3.215 110  21.26502
#> 8       4 response   hp -0.03177295 0.009029704 3.215 100  21.58275
#> 9       5 response   hp -0.03177295 0.009029704 3.440 100  20.71024
#> 10      5 response   hp -0.03177295 0.009029704 3.440 110  20.39251
#> 11      6 response   hp -0.03177295 0.009029704 3.460 100  20.63268
#> 12      6 response   hp -0.03177295 0.009029704 3.460 110  20.31495
#> 13      7 response   hp -0.03177295 0.009029704 3.570 110  19.88839
#> 14      7 response   hp -0.03177295 0.009029704 3.570 100  20.20612
#> 15      8 response   hp -0.03177295 0.009029704 3.190 110  21.36197
#> 16      8 response   hp -0.03177295 0.009029704 3.190 100  21.67970
#> 17      9 response   hp -0.03177295 0.009029704 3.150 110  21.51708
#> 18      9 response   hp -0.03177295 0.009029704 3.150 100  21.83481
#> 19     10 response   hp -0.03177295 0.009029704 3.440 100  20.71024
#> 20     10 response   hp -0.03177295 0.009029704 3.440 110  20.39251
#> 21     11 response   hp -0.03177295 0.009029704 3.440 110  20.39251
#> 22     11 response   hp -0.03177295 0.009029704 3.440 100  20.71024
#> 23     12 response   hp -0.03177295 0.009029704 4.070 110  17.94947
#> 24     12 response   hp -0.03177295 0.009029704 4.070 100  18.26720
#> 25     13 response   hp -0.03177295 0.009029704 3.730 110  19.26794
#> 26     13 response   hp -0.03177295 0.009029704 3.730 100  19.58567
#> 27     14 response   hp -0.03177295 0.009029704 3.780 100  19.39178
#> 28     14 response   hp -0.03177295 0.009029704 3.780 110  19.07405
#> 29     15 response   hp -0.03177295 0.009029704 5.250 100  13.69136
#> 30     15 response   hp -0.03177295 0.009029704 5.250 110  13.37363
#> 31     16 response   hp -0.03177295 0.009029704 5.424 110  12.69889
#> 32     16 response   hp -0.03177295 0.009029704 5.424 100  13.01662
#> 33     17 response   hp -0.03177295 0.009029704 5.345 110  13.00524
#> 34     17 response   hp -0.03177295 0.009029704 5.345 100  13.32297
#> 35     18 response   hp -0.03177295 0.009029704 2.200 100  25.51875
#> 36     18 response   hp -0.03177295 0.009029704 2.200 110  25.20102
#> 37     19 response   hp -0.03177295 0.009029704 1.615 100  27.78728
#> 38     19 response   hp -0.03177295 0.009029704 1.615 110  27.46955
#> 39     20 response   hp -0.03177295 0.009029704 1.835 110  26.61643
#> 40     20 response   hp -0.03177295 0.009029704 1.835 100  26.93416
#> 41     21 response   hp -0.03177295 0.009029704 2.465 110  24.17339
#> 42     21 response   hp -0.03177295 0.009029704 2.465 100  24.49112
#> 43     22 response   hp -0.03177295 0.009029704 3.520 110  20.08228
#> 44     22 response   hp -0.03177295 0.009029704 3.520 100  20.40001
#> 45     23 response   hp -0.03177295 0.009029704 3.435 100  20.72963
#> 46     23 response   hp -0.03177295 0.009029704 3.435 110  20.41190
#> 47     24 response   hp -0.03177295 0.009029704 3.840 110  18.84138
#> 48     24 response   hp -0.03177295 0.009029704 3.840 100  19.15911
#> 49     25 response   hp -0.03177295 0.009029704 3.845 110  18.82199
#> 50     25 response   hp -0.03177295 0.009029704 3.845 100  19.13972
#> 51     26 response   hp -0.03177295 0.009029704 1.935 110  26.22864
#> 52     26 response   hp -0.03177295 0.009029704 1.935 100  26.54637
#> 53     27 response   hp -0.03177295 0.009029704 2.140 100  25.75142
#> 54     27 response   hp -0.03177295 0.009029704 2.140 110  25.43369
#> 55     28 response   hp -0.03177295 0.009029704 1.513 100  28.18282
#> 56     28 response   hp -0.03177295 0.009029704 1.513 110  27.86509
#> 57     29 response   hp -0.03177295 0.009029704 3.170 110  21.43952
#> 58     29 response   hp -0.03177295 0.009029704 3.170 100  21.75725
#> 59     30 response   hp -0.03177295 0.009029704 2.770 110  22.99065
#> 60     30 response   hp -0.03177295 0.009029704 2.770 100  23.30838
#> 61     31 response   hp -0.03177295 0.009029704 3.570 100  20.20612
#> 62     31 response   hp -0.03177295 0.009029704 3.570 110  19.88839
#> 63     32 response   hp -0.03177295 0.009029704 2.780 100  23.26961
#> 64     32 response   hp -0.03177295 0.009029704 2.780 110  22.95188
#> 65      1 response   wt -3.87783074 0.632733547 2.620 100  23.89006
#> 66      1 response   wt -3.87783074 0.632733547 2.620 110  23.57233
#> 67      2 response   wt -3.87783074 0.632733322 2.875 100  22.90121
#> 68      2 response   wt -3.87783074 0.632733322 2.875 110  22.58348
#> 69      3 response   wt -3.87783074 0.632733097 2.320 110  24.73568
#> 70      3 response   wt -3.87783074 0.632733097 2.320 100  25.05341
#> 71      4 response   wt -3.87783074 0.632733322 3.215 110  21.26502
#> 72      4 response   wt -3.87783074 0.632733322 3.215 100  21.58275
#> 73      5 response   wt -3.87783074 0.632733322 3.440 100  20.71024
#> 74      5 response   wt -3.87783074 0.632733322 3.440 110  20.39251
#> 75      6 response   wt -3.87783074 0.632733547 3.460 100  20.63268
#> 76      6 response   wt -3.87783074 0.632733547 3.460 110  20.31495
#> 77      7 response   wt -3.87783074 0.632733322 3.570 110  19.88839
#> 78      7 response   wt -3.87783074 0.632733322 3.570 100  20.20612
#> 79      8 response   wt -3.87783074 0.632733322 3.190 110  21.36197
#> 80      8 response   wt -3.87783074 0.632733322 3.190 100  21.67970
#> 81      9 response   wt -3.87783074 0.632733547 3.150 110  21.51708
#> 82      9 response   wt -3.87783074 0.632733547 3.150 100  21.83481
#> 83     10 response   wt -3.87783074 0.632733322 3.440 100  20.71024
#> 84     10 response   wt -3.87783074 0.632733322 3.440 110  20.39251
#> 85     11 response   wt -3.87783074 0.632733322 3.440 110  20.39251
#> 86     11 response   wt -3.87783074 0.632733322 3.440 100  20.71024
#> 87     12 response   wt -3.87783074 0.632733547 4.070 110  17.94947
#> 88     12 response   wt -3.87783074 0.632733547 4.070 100  18.26720
#> 89     13 response   wt -3.87783074 0.632733547 3.730 110  19.26794
#> 90     13 response   wt -3.87783074 0.632733547 3.730 100  19.58567
#> 91     14 response   wt -3.87783074 0.632733322 3.780 100  19.39178
#> 92     14 response   wt -3.87783074 0.632733322 3.780 110  19.07405
#> 93     15 response   wt -3.87783074 0.632733547 5.250 100  13.69136
#> 94     15 response   wt -3.87783074 0.632733547 5.250 110  13.37363
#> 95     16 response   wt -3.87783074 0.632733322 5.424 110  12.69889
#> 96     16 response   wt -3.87783074 0.632733322 5.424 100  13.01662
#> 97     17 response   wt -3.87783074 0.632733547 5.345 110  13.00524
#> 98     17 response   wt -3.87783074 0.632733547 5.345 100  13.32297
#> 99     18 response   wt -3.87783074 0.632733322 2.200 100  25.51875
#> 100    18 response   wt -3.87783074 0.632733322 2.200 110  25.20102
#> 101    19 response   wt -3.87783074 0.632733772 1.615 100  27.78728
#> 102    19 response   wt -3.87783074 0.632733772 1.615 110  27.46955
#> 103    20 response   wt -3.87783074 0.632733322 1.835 110  26.61643
#> 104    20 response   wt -3.87783074 0.632733322 1.835 100  26.93416
#> 105    21 response   wt -3.87783074 0.632733097 2.465 110  24.17339
#> 106    21 response   wt -3.87783074 0.632733097 2.465 100  24.49112
#> 107    22 response   wt -3.87783074 0.632733322 3.520 110  20.08228
#> 108    22 response   wt -3.87783074 0.632733322 3.520 100  20.40001
#> 109    23 response   wt -3.87783074 0.632733547 3.435 100  20.72963
#> 110    23 response   wt -3.87783074 0.632733547 3.435 110  20.41190
#> 111    24 response   wt -3.87783074 0.632733996 3.840 110  18.84138
#> 112    24 response   wt -3.87783074 0.632733996 3.840 100  19.15911
#> 113    25 response   wt -3.87783074 0.632733322 3.845 110  18.82199
#> 114    25 response   wt -3.87783074 0.632733322 3.845 100  19.13972
#> 115    26 response   wt -3.87783074 0.632733322 1.935 110  26.22864
#> 116    26 response   wt -3.87783074 0.632733322 1.935 100  26.54637
#> 117    27 response   wt -3.87783074 0.632733547 2.140 100  25.75142
#> 118    27 response   wt -3.87783074 0.632733547 2.140 110  25.43369
#> 119    28 response   wt -3.87783074 0.632733322 1.513 100  28.18282
#> 120    28 response   wt -3.87783074 0.632733322 1.513 110  27.86509
#> 121    29 response   wt -3.87783074 0.632733547 3.170 110  21.43952
#> 122    29 response   wt -3.87783074 0.632733547 3.170 100  21.75725
#> 123    30 response   wt -3.87783074 0.632733547 2.770 110  22.99065
#> 124    30 response   wt -3.87783074 0.632733547 2.770 100  23.30838
#> 125    31 response   wt -3.87783074 0.632733322 3.570 100  20.20612
#> 126    31 response   wt -3.87783074 0.632733322 3.570 110  19.88839
#> 127    32 response   wt -3.87783074 0.632733322 2.780 100  23.26961
#> 128    32 response   wt -3.87783074 0.632733322 2.780 110  22.95188