EXPERIMENTAL Uncertainty Estimates for marginaleffects Objects

Warning: This function is experimental. It may be renamed, the user interface may change, or the functionality may migrate to arguments in other marginaleffects functions.

Apply this function to a model object to change the inferential strategy used to compute uncertainty estimates: delta method or simulation-based inference.

Usage

inferences(model, method = "simulation", iter = 1000, ...)

Arguments

model

Model object

method

String: "simulation" or "delta". See below for references and details.

iter

Number of simulations.

...

Other arguments are ignored.

Value

A "decorated" model with additional information on how to conduct inference and compute uncertainty estimates. This decorated model can then fed to any of the marginaleffects functions, as one would for any other supported models.

Details

When method="simulation", we conduct simulation-based inference following the method discussed in Krinsky & Robb (1986):

Draw iter sets of simulated coefficients from a multivariate normal distribution with mean equal to the original model's estimated coefficients and variance equal to the model's variance-covariance matrix (classical, "HC3", or other).

Use the iter sets of coefficients to compute iter sets of estimands: predictions, comparisons, or slopes.

Take quantiles of the resulting distribution of estimands to obtain a confidence interval and the standard deviation of simulated estimates to estimate the standard error.

References

Krinsky, I., and A. L. Robb. 1986. “On Approximating the Statistical Properties of Elasticities.” Review of Economics and Statistics 68 (4): 715–9.

King, Gary, Michael Tomz, and Jason Wittenberg. "Making the most of statistical analyses: Improving interpretation and presentation." American journal of political science (2000): 347-361

Dowd, Bryan E., William H. Greene, and Edward C. Norton. "Computation of standard errors." Health services research 49.2 (2014): 731-750.

Examples

library(marginaleffects)
library(magrittr)
mod <- glm(vs ~ hp * wt + factor(gear), data = mtcars, family = binomial)

mod %>% inferences() %>%
        avg_predictions(by = "gear")
#> 
#>  gear Estimate Std. Error   2.5 % 97.5 %
#>     3   0.2022     0.1538 0.07417 0.7634
#>     4   0.7931     0.1166 0.48274 0.9445
#>     5   0.2135     0.1309 0.03458 0.5602
#> 
#> Prediction type:  response 
#> Columns: type, gear, estimate, std.error, conf.low, conf.high 
#> 

mod %>% inferences() %>%
        slopes() %>%
        head()
#> 
#>  Term Contrast   Estimate Std. Error    2.5 %    97.5 %
#>    hp    dY/dX -0.0091195   0.010031 -0.03491 0.0012889
#>    hp    dY/dX -0.0106273   0.009523 -0.03358 0.0012339
#>    hp    dY/dX -0.0037059   0.008115 -0.02804 0.0025639
#>    hp    dY/dX -0.0067763   0.007519 -0.02682 0.0008921
#>    hp    dY/dX -0.0008534   0.003502 -0.01171 0.0011195
#>    hp    dY/dX -0.0049370   0.006791 -0.02317 0.0010450
#> 
#> Prediction type:  response 
#> Columns: rowid, type, term, contrast, estimate, std.error, conf.low, conf.high, predicted, predicted_hi, predicted_lo, vs, hp, wt, gear, eps 
#> 

mod %>% inferences() %>%
        avg_slopes() %>%
        posterior_draws("rvar")
#> 
#>  Term Contrast  Estimate Std. Error    2.5 %   97.5 %
#>  gear    4 - 3 -0.035126     0.1442 -0.24786 0.326803
#>  gear    5 - 3 -0.176318     0.2307 -0.54161 0.353115
#>    hp    dY/dX -0.004462     0.0030 -0.01065 0.001938
#>    wt    dY/dX  0.018325     0.1789 -0.32781 0.398963
#> 
#> Prediction type:  response 
#> Columns: type, term, contrast, estimate, std.error, conf.low, conf.high, rvar 
#>