EXPERIMENTAL Uncertainty Estimates for marginaleffects Objects

Warning: This function is experimental. It may be renamed, the user interface may change, or the functionality may migrate to arguments in other marginaleffects functions.

Apply this function to a marginaleffects object to change the inferential method used to compute uncertainty estimates.

Usage

inferences(x, method = "simulation", R = 1000, conf_type = "perc", ...)

Arguments

x

Object produced by one of the core marginaleffects functions.

method

String

"delta": delta method standard errors

"boot" package

"rsample" package

"simulation" from a multivariate normal distribution (Krinsky & Robb, 1986)

R

Number of resamples or simulations.

conf_type

String: type of bootstrap interval to construct.

boot: "perc", "norm", "basic", or "bca"

rsample: "perc" or "bca"

simulation: argument ignored.

...

If method="boot", additional arguments are passed to boot::boot().

If method="rsample", additional arguments are passed to rsample::bootstraps().

If method="simulation", additional arguments are ignored.

Value

A marginaleffects object with simulation or bootstrap resamples and objects attached.

Details

When method="simulation", we conduct simulation-based inference following the method discussed in Krinsky & Robb (1986):

Draw R sets of simulated coefficients from a multivariate normal distribution with mean equal to the original model's estimated coefficients and variance equal to the model's variance-covariance matrix (classical, "HC3", or other).

Use the R sets of coefficients to compute R sets of estimands: predictions, comparisons, or slopes.

Take quantiles of the resulting distribution of estimands to obtain a confidence interval and the standard deviation of simulated estimates to estimate the standard error.

References

Krinsky, I., and A. L. Robb. 1986. “On Approximating the Statistical Properties of Elasticities.” Review of Economics and Statistics 68 (4): 715–9.

King, Gary, Michael Tomz, and Jason Wittenberg. "Making the most of statistical analyses: Improving interpretation and presentation." American journal of political science (2000): 347-361

Dowd, Bryan E., William H. Greene, and Edward C. Norton. "Computation of standard errors." Health services research 49.2 (2014): 731-750.

Examples

library(marginaleffects)
library(magrittr)
mod <- glm(vs ~ hp * wt + factor(gear), data = mtcars, family = binomial)

avg_predictions(mod, by = "gear") %>%
  inferences()
#> 
#>  gear Estimate Std. Error   2.5 % 97.5 %
#>     3   0.2055     0.1425 0.08148 0.7329
#>     4   0.7918     0.1134 0.50665 0.9405
#>     5   0.2125     0.1310 0.03715 0.5513
#> 
#> Prediction type:  response 
#> Columns: type, gear, estimate, std.error, conf.low, conf.high 
#> 

slopes(mod) %>%
  inferences() %>%
  head()
#> 
#>  Term Contrast   Estimate Std. Error    2.5 %   97.5 %
#>    hp    dY/dX -0.0089718   0.009993 -0.03396 0.001524
#>    hp    dY/dX -0.0101572   0.009798 -0.03465 0.001190
#>    hp    dY/dX -0.0034492   0.007977 -0.02803 0.002963
#>    hp    dY/dX -0.0066931   0.007915 -0.02833 0.001253
#>    hp    dY/dX -0.0007352   0.003551 -0.01251 0.001528
#>    hp    dY/dX -0.0051158   0.007015 -0.02459 0.001775
#> 
#> Prediction type:  response 
#> Columns: rowid, type, term, contrast, estimate, std.error, conf.low, conf.high, predicted, predicted_hi, predicted_lo, vs, hp, wt, gear, eps 
#> 

avg_slopes(mod) %>%
  inferences() %>%
  posterior_draws("rvar")
#> 
#>  Term Contrast  Estimate Std. Error    2.5 %   97.5 %
#>  gear    4 - 3 -0.032990   0.144602 -0.23676 0.371167
#>  gear    5 - 3 -0.164495   0.233371 -0.55117 0.363858
#>    hp    dY/dX -0.004517   0.002999 -0.01008 0.001752
#>    wt    dY/dX  0.014645   0.173003 -0.32321 0.379605
#> 
#> Prediction type:  response 
#> Columns: type, term, contrast, estimate, std.error, conf.low, conf.high, rvar 
#>