marginal_means() is an alias to marginal_means()

This alias is kept for backward compatibility and because some users may prefer that name.

Usage

marginalmeans(
  model,
  variables = NULL,
  variables_grid = NULL,
  vcov = TRUE,
  conf_level = 0.95,
  type = NULL,
  transform_post = NULL,
  cross = FALSE,
  hypothesis = NULL,
  df = Inf,
  wts = "equal",
  by = NULL,
  ...
)

Arguments

model

Model object

variables

character vector Categorical predictors over which to compute marginal means. NULL calculates marginal means for all logical, character, or factor variables in the dataset used to fit model. Set cross=TRUE to compute marginal means at combinations of the predictors specified in the variables argument.

variables_grid

character vector Categorical predictors used to construct the prediction grid over which adjusted predictions are averaged (character vector). NULL creates a grid with all combinations of all categorical predictors. This grid can be very large when there are many variables and many response levels, so it is advisable to select a limited number of variables in the variables and variables_grid arguments.

vcov

Type of uncertainty estimates to report (e.g., for robust standard errors). Acceptable values:

FALSE: Do not compute standard errors. This can speed up computation considerably.

TRUE: Unit-level standard errors using the default vcov(model) variance-covariance matrix.

String which indicates the kind of uncertainty estimates to return.

Heteroskedasticity-consistent: "HC", "HC0", "HC1", "HC2", "HC3", "HC4", "HC4m", "HC5". See ?sandwich::vcovHC

Heteroskedasticity and autocorrelation consistent: "HAC"

Mixed-Models degrees of freedom: "satterthwaite", "kenward-roger"

Other: "NeweyWest", "KernHAC", "OPG". See the sandwich package documentation.

One-sided formula which indicates the name of cluster variables (e.g., ~unit_id). This formula is passed to the cluster argument of the sandwich::vcovCL function.

Square covariance matrix

Function which returns a covariance matrix (e.g., stats::vcov(model))

conf_level

numeric value between 0 and 1. Confidence level to use to build a confidence interval.

type

string indicates the type (scale) of the predictions used to compute marginal effects or contrasts. This can differ based on the model type, but will typically be a string such as: "response", "link", "probs", or "zero". When an unsupported string is entered, the model-specific list of acceptable values is returned in an error message. When type is NULL, the default value is used. This default is the first model-related row in the marginaleffects:::type_dictionary dataframe. If type is NULL and the default value is "response", the function tries to compute marginal means on the link scale before backtransforming them using the inverse link function.

transform_post

(experimental) A function applied to unit-level adjusted predictions and confidence intervals just before the function returns results. For bayesian models, this function is applied to individual draws from the posterior distribution, before computing summaries.

cross

TRUE or FALSE

FALSE (default): Marginal means are computed for each predictor individually.

TRUE: Marginal means are computed for each combination of predictors specified in the variables argument.

hypothesis

specify a hypothesis test or custom contrast using a numeric value, vector, or matrix, a string, or a string formula.

Numeric:

Single value: the null hypothesis used in the computation of Z and p (before applying transform_post).

Vector: Weights to compute a linear combination of (custom contrast between) estimates. Length equal to the number of rows generated by the same function call, but without the hypothesis argument.

Matrix: Each column is a vector of weights, as describe above, used to compute a distinct linear combination of (contrast between) estimates. The column names of the matrix are used as labels in the output.

String formula to specify linear or non-linear hypothesis tests. If the term column uniquely identifies rows, terms can be used in the formula. Otherwise, use b1, b2, etc. to identify the position of each parameter. Examples:

hp = drat

hp + drat = 12

b1 + b2 + b3 = 0

String:

"pairwise": pairwise differences between estimates in each row.

"reference": differences between the estimates in each row and the estimate in the first row.

"sequential": difference between an estimate and the estimate in the next row.

"revpairwise", "revreference", "revsequential": inverse of the corresponding hypotheses, as described above.

See the Examples section below and the vignette: https://vincentarelbundock.github.io/marginaleffects/articles/hypothesis.html

df

Degrees of freedom used to compute p values and confidence intervals. A single numeric value between 1 and Inf. When df is Inf, the normal distribution is used. When df is finite, the t distribution is used. See insight::get_df for a convenient function to extract degrees of freedom. Ex: slopes(model, df = insight::get_df(model))

wts

character value. Weights to use in the averaging.

"equal": each combination of variables in variables_grid gets an equal weight.

"cells": each combination of values for the variables in the variables_grid gets a weight proportional to its frequency in the original data.

"proportional": each combination of values for the variables in the variables_grid -- except for those in the variables argument -- gets a weight proportional to its frequency in the original data.

by

Collapse marginal means into categories. Data frame with a by column of group labels, and merging columns shared by newdata or the data frame produced by calling the same function without the by argument.

...

Additional arguments are passed to the predict() method supplied by the modeling package.These arguments are particularly useful for mixed-effects or bayesian models (see the online vignettes on the marginaleffects website). Available arguments can vary from model to model, depending on the range of supported arguments by each modeling package. See the "Model-Specific Arguments" section of the ?marginaleffects documentation for a non-exhaustive list of available arguments.

Value

Data frame of marginal means with one row per variable-value combination.

Details

This function begins by calling the predictions function to obtain a grid of predictors, and adjusted predictions for each cell. The grid includes all combinations of the categorical variables listed in the variables and variables_grid arguments, or all combinations of the categorical variables used to fit the model if variables_grid is NULL. In the prediction grid, numeric variables are held at their means.

After constructing the grid and filling the grid with adjusted predictions, marginal_means computes marginal means for the variables listed in the variables argument, by average across all categories in the grid.

marginal_means can only compute standard errors for linear models, or for predictions on the link scale, that is, with the type argument set to "link".

The marginaleffects website compares the output of this function to the popular emmeans package, which provides similar but more advanced functionality: https://vincentarelbundock.github.io/marginaleffects/

Vignettes and documentation

Vignettes:

Predictions

Comparisons

Slopes

Marginal Means

Hypothesis Tests and Custom Contrasts

Case studies:

Bayesian Analyses with brms

Causal Inference with the g-Formula

Elasticity

Experiments

Generalized Additive Models

Mixed effects models

Multinomial Logit and Discrete Choice Models

Multiple Imputation

Plots: interactions, predictions, contrasts, and slopes

Python NumPyro models in marginaleffects

Unit-level contrasts in logistic regressions

Tips and technical notes:

73 Supported Classes of Models

Index of Functions and Documentation

Extending marginaleffects: add new models or modify existing ones

Standard Errors and Confidence Intervals

Tables and Plots

Performance

Alternative Software

Frequently Asked Questions

Model-Specific Arguments

Some model types allow model-specific arguments to modify the nature of marginal effects, predictions, marginal means, and contrasts.

Package	Class	Argument	Documentation

brms	brmsfit	ndraws	brms::posterior_predict

re_formula

lme4	merMod	include_random	insight::get_predicted

re.form	lme4::predict.merMod

allow.new.levels	lme4::predict.merMod

glmmTMB	glmmTMB	re.form	glmmTMB::predict.glmmTMB

allow.new.levels	glmmTMB::predict.glmmTMB

zitype	glmmTMB::predict.glmmTMB

mgcv	bam	exclude	mgcv::predict.bam

robustlmm	rlmerMod	re.form	robustlmm::predict.rlmerMod

allow.new.levels	robustlmm::predict.rlmerMod

Bayesian posterior summaries

By default, credible intervals in bayesian models are built as equal-tailed intervals. This can be changed to a highest density interval by setting a global option:

options("marginaleffects_posterior_interval" = "eti")

options("marginaleffects_posterior_interval" = "hdi")

By default, the center of the posterior distribution in bayesian models is identified by the median. Users can use a different summary function by setting a global option:

options("marginaleffects_posterior_center" = "mean")

options("marginaleffects_posterior_center" = "median")

When estimates are averaged using the by argument, the tidy() function, or the summary() function, the posterior distribution is marginalized twice over. First, we take the average across units but within each iteration of the MCMC chain, according to what the user requested in by argument or tidy()/summary() functions. Then, we identify the center of the resulting posterior using the function supplied to the "marginaleffects_posterior_center" option (the median by default).

Examples

library(marginaleffects)

# simple marginal means for each level of `cyl`
dat <- mtcars
dat$carb <- factor(dat$carb)
dat$cyl <- factor(dat$cyl)
dat$am <- as.logical(dat$am)
mod <- lm(mpg ~ carb + cyl + am, dat)

marginal_means(
  mod,
  variables = "cyl")
#>       type term value cyl estimate std.error statistic      p.value conf.low
#> 1 response  cyl     4   4 23.11334  1.658224  13.93861 3.690579e-44 19.86328
#> 2 response  cyl     6   6 20.38477  1.337106  15.24544 1.765453e-52 17.76409
#> 3 response  cyl     8   8 16.21082  1.072551  15.11427 1.304150e-51 14.10866
#>   conf.high
#> 1  26.36340
#> 2  23.00545
#> 3  18.31299

# collapse levels of cyl by averaging
by <- data.frame(
  cyl = c(4, 6, 8),
  by = c("4 & 6", "4 & 6", "8"))
marginal_means(mod,
  variables = "cyl",
  by = by)
#>      by estimate std.error statistic      p.value conf.low conf.high
#> 1 4 & 6 21.74905  1.132084  19.21152 2.964293e-82 19.53021  23.96790
#> 2     8 16.21082  1.072551  15.11427 1.304150e-51 14.10866  18.31299

# pairwise differences between collapsed levels
marginal_means(mod,
  variables = "cyl",
  by = by,
  hypothesis = "pairwise")
#>        term estimate std.error statistic      p.value conf.low conf.high
#> 1 4 & 6 - 8 5.538228  1.513052  3.660302 0.0002519178   2.5727  8.503755

# cross
marginal_means(mod,
  variables = c("cyl", "carb"),
  cross = TRUE)
#>        type cyl carb estimate std.error statistic       p.value  conf.low
#> 1  response   4    1 25.81720  1.263909 20.426468  9.729029e-93 23.339980
#> 2  response   4    2 25.58836  1.166714 21.931995 1.286605e-106 23.301643
#> 3  response   4    3 25.33929  2.365152 10.713598  8.787848e-27 20.703673
#> 4  response   4    4 21.87765  1.901048 11.508200  1.199563e-30 18.151659
#> 5  response   4    6 20.29180  3.766424  5.387551  7.142425e-08 12.909744
#> 6  response   4    8 19.76574  3.814283  5.182034  2.194796e-07 12.289884
#> 7  response   6    1 23.08862  1.765271 13.079362  4.320526e-39 19.628756
#> 8  response   6    2 22.85979  1.867465 12.241080  1.875206e-34 19.199624
#> 9  response   6    3 22.61071  2.365152  9.559942  1.178225e-21 17.975102
#> 10 response   6    4 19.14907  1.338478 14.306600  1.990073e-46 16.525705
#> 11 response   6    6 17.56323  3.003716  5.847167  5.000145e-09 11.676053
#> 12 response   6    8 17.03717  3.481207  4.894041  9.878645e-07 10.214129
#> 13 response   8    1 18.91468  1.942030  9.739643  2.042705e-22 15.108373
#> 14 response   8    2 18.68585  1.570513 11.897929  1.213213e-32 15.607698
#> 15 response   8    3 18.43677  1.830964 10.069436  7.540931e-24 14.848149
#> 16 response   8    4 14.97513  1.195291 12.528445  5.217536e-36 12.632406
#> 17 response   8    6 13.38929  3.356162  3.989464  6.622284e-05  6.811330
#> 18 response   8    8 12.86323  3.003716  4.282439  1.848561e-05  6.976053
#>    conf.high
#> 1   28.29441
#> 2   27.87508
#> 3   29.97490
#> 4   25.60363
#> 5   27.67385
#> 6   27.24160
#> 7   26.54849
#> 8   26.51995
#> 9   27.24633
#> 10  21.77244
#> 11  23.45040
#> 12  23.86021
#> 13  22.72099
#> 14  21.76399
#> 15  22.02540
#> 16  17.31786
#> 17  19.96724
#> 18  18.75040

# collapsed cross
by <- expand.grid(
  cyl = unique(mtcars$cyl),
  carb = unique(mtcars$carb))
by$by <- ifelse(
  by$cyl == 4,
  paste("Control:", by$carb),
  paste("Treatment:", by$carb))

# Convert numeric variables to categorical before fitting the model
dat <- mtcars
dat$am <- as.logical(dat$am)
dat$carb <- as.factor(dat$carb)
mod <- lm(mpg ~ hp + am + carb, data = dat)

# Compute and summarize marginal means
marginal_means(mod)
#>       type term value estimate std.error statistic       p.value conf.low
#> 1 response carb     1 21.99084 1.3449858 16.350238  4.332854e-60 19.35471
#> 2 response carb     2 21.47658 1.0249035 20.954731  1.698845e-97 19.46780
#> 3 response carb     3 20.55258 1.7795791 11.549126  7.457511e-31 17.06467
#> 4 response carb     4 18.81535 1.0415402 18.064933  6.019468e-73 16.77397
#> 5 response carb     6 18.47146 3.0191868  6.118024  9.474270e-10 12.55396
#> 6 response carb     8 21.62292 4.0550250  5.332377  9.693562e-08 13.67522
#> 7 response   am FALSE 17.87040 1.2439568 14.365776  8.485108e-47 15.43229
#> 8 response   am  TRUE 23.10617 0.9739524 23.724129 2.032675e-124 21.19726
#>   conf.high
#> 1  24.62696
#> 2  23.48535
#> 3  24.04049
#> 4  20.85674
#> 5  24.38895
#> 6  29.57062
#> 7  20.30852
#> 8  25.01508

# Contrast between marginal means (carb2 - carb1), or "is the 1st marginal means equal to the 2nd?"
# see the vignette on "Hypothesis Tests and Custom Contrasts" on the `marginaleffects` website.
lc <- c(-1, 1, 0, 0, 0, 0)
marginal_means(mod, variables = "carb", hypothesis = "b2 = b1")
#>    term   estimate std.error  statistic   p.value  conf.low conf.high
#> 1 b2=b1 -0.5142619  1.477885 -0.3479715 0.7278616 -3.410863   2.38234

marginal_means(mod, variables = "carb", hypothesis = lc)
#>     term   estimate std.error  statistic   p.value  conf.low conf.high
#> 1 custom -0.5142619  1.477885 -0.3479715 0.7278616 -3.410863   2.38234

# Multiple custom contrasts
lc <- matrix(c(
    -2, 1, 1, 0, -1, 1,
    -1, 1, 0, 0, 0, 0
    ),
  ncol = 2,
  dimnames = list(NULL, c("A", "B")))
marginal_means(mod, variables = "carb", hypothesis = lc)
#>   term   estimate std.error  statistic   p.value   conf.low conf.high
#> 1    A  1.1989451  6.149643  0.1949618 0.8454229 -10.854133  13.25202
#> 2    B -0.5142619  1.477885 -0.3479715 0.7278616  -3.410863   2.38234