Tidy a marginaleffects object

Tidy a marginaleffects object

# S3 method for marginaleffects
tidy(x, conf.int = TRUE, conf.level = 0.95, ...)

Arguments

x
An object produced by the marginaleffects function.

conf.int
Logical indicating whether or not to include a confidence interval.

conf.level
The confidence level to use for the confidence interval if conf.int=TRUE. Must be strictly greater than 0 and less than 1. Defaults to 0.95, which corresponds to a 95 percent confidence interval.

...
Additional arguments are pushed forward to predict().

Details

The tidy function calculates average marginal effects by taking the mean of all the unit-level marginal effects computed by the marginaleffects function.

Examples

mod <- lm(mpg ~ hp * wt + factor(gear), data = mtcars)
mfx <- marginaleffects(mod)
tidy(mfx)

#> # A tibble: 5 × 9
#> # Groups:   type [1]
#>   type     term  contrast estimate std.error statistic p.value conf.low conf.high
#>   <chr>    <chr> <chr>       <dbl>     <dbl>     <dbl>   <dbl>    <dbl>     <dbl>
#> 1 response hp    NA        -0.0353    0.0124    -2.84  4.50e-3  -0.0597   -0.0110
#> 2 response wt    NA        -3.53      0.877     -4.02  5.80e-5  -5.24     -1.81  
#> 3 response gear  4 - 3      0.829     1.12       0.741 7.41e-1  -1.36      3.02  
#> 4 response gear  5 - 3      1.82      1.55       1.17  4.79e-1  -1.22      4.85  
#> 5 response gear  5 - 4      0.987     1.60       0.617 8.12e-1  -2.15      4.12